### 2020-07-25 #2516 @whedon
[PRE REVIEW]: PyAutoFit`: A Classy Probabilistic Programming Language for Model Composition and Fitting
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** <a href="https://github.com/rhayes777/PyAutoFit" target ="_blank">https://github.com/rhayes777/PyAutoFit</a>
**Version:** v0.64.3
