### 2020-07-25 #2516 @Kevin-Mattheus-Moerman
@Jammy2211 After investigating if this submission is suitable for JOSS I'll work on finding an editor to handle this submission. In the mean time perhaps you can work on those missing and invalid DOI's :point_up:. You can call `@whedon generate pdf` to update the paper and you can call `@whedon check references` to check those references.
