### 2020-07-30 b52a2c2
faster execution with different likelihood function
