### 2020-07-30 2cad16c
retain function shape to vastly speed up execution - is this dangerous?
