### 2020-07-30 a5b78ac
another slight speed up
