### 2020-07-30 61e98a0
introduced implicit variables although not working yet
