### 2020-07-30 de1b181
use getattr to perform variable matching
