### 2020-07-30 b225918
variables are a kind of model object
