### 2020-07-30 8f27b26
merged factors with factor nodes
