+++ autofit/message_passing/factor_graphs/__init__.py
@@ -1,6 +1,6 @@
 from autofit.message_passing.factor_graphs.abstract import AbstractNode
 from autofit.message_passing.factor_graphs.abstract import AbstractNode
-from autofit.message_passing.factor_graphs.factor import FactorNode, FactorValue, Factor
+from autofit.message_passing.factor_graphs.factor import FactorValue, Factor
 from autofit.message_passing.factor_graphs.graph import DeterministicFactorNode, FactorGraph
 from autofit.message_passing.factor_graphs.numerical import numerical_jacobian
 from autofit.message_passing.factor_graphs.variable import Plate, Variable
