### 2020-07-30 ccb455d
use variables as key in the message dict
