+++ test_autofit/unit/message_passing/gaussian/test_syntax.py
@@ -40,6 +40,11 @@ def test_gaussian():
         _gaussian,
         x=x_
     ) == z
+
+    centre = gaussian.centre
+    intensity = gaussian.intensity
+    sigma = gaussian.sigma
+
     likelihood = mp.Factor(
         _likelihood,
         z=z,
@@ -49,35 +54,37 @@ def test_gaussian():
     # TODO: Can priors look like autofit priors? Could mp objects derive promise functionality from autofit?
     prior_centre = mp.Factor(
         prior,
-        x=gaussian.centre
+        x=centre
     )
     prior_intensity = mp.Factor(
         prior,
-        x=gaussian.intensity
+        x=intensity
     )
     prior_sigma = mp.Factor(
         prior,
-        x=gaussian.sigma
+        x=sigma
     )
 
     model = likelihood * gaussian * prior_centre * prior_sigma * prior_intensity
 
     model_approx = mp.MeanFieldApproximation.from_kws(
         model,
-        centre=mp.NormalMessage.from_prior(
-            prior
-        ),
-        intensity=mp.NormalMessage.from_prior(
-            prior
-        ),
-        sigma=mp.NormalMessage.from_prior(
-            prior
-        ),
-        x=mp.FixedMessage(x),
-        y=mp.FixedMessage(y),
-        z=mp.NormalMessage.from_mode(
-            np.zeros(n_observations), 100
-        ),
+        {
+            centre: mp.NormalMessage.from_prior(
+                prior
+            ),
+            intensity: mp.NormalMessage.from_prior(
+                prior
+            ),
+            sigma: mp.NormalMessage.from_prior(
+                prior
+            ),
+            x_: mp.FixedMessage(x),
+            y_: mp.FixedMessage(y),
+            z: mp.NormalMessage.from_mode(
+                np.zeros(n_observations), 100
+            ),
+        }
     )
 
     opt = mp.optimise.LaplaceOptimiser(
@@ -86,5 +93,5 @@ def test_gaussian():
     )
     opt.run()
 
-    for string in ("centre", "intensity", "sigma"):
-        print(f"{string} = {opt.model_approx[string].mu}")
+    for variable in (centre, intensity, sigma):
+        print(f"{variable.name} = {opt.model_approx[variable].mu}")

[…]
