### 2020-08-05 408bc74
retrieving correct variable count for the model
