### 2020-08-05 39154b5
fixed det vars updating
