+++ autofit/message_passing/factor_graphs/graph.py
@@ -281,7 +281,9 @@ class FactorGraph(AbstractNode):
                 ret_value = self.broadcast_plates(factor.plates, ret.log_value)
                 log_value = add_arrays(log_value, ret_value)
                 det_values.update(ret.deterministic_values)
-                variables.update(ret.deterministic_values)
+                variables.update({
+                    variable.name: array for variable, array in ret.deterministic_values.items()
+                })
 
         return FactorValue(log_value, det_values)
 
