+++ test_autofit/unit/message_passing/test_scratch.py
@@ -10,10 +10,11 @@ def func(x):
 def test_():
     x = mp.Variable("x")
     factor = mp.Factor(
-        func
-    )(x)
+        func,
+        x=x
+    )
 
     model = factor * factor
 
-    value = model(np.array([1.0]))
+    value = model(x=np.array([1.0]))
     assert value.log_value == np.array([4.0])
