+++ test_autofit/unit/message_passing/regression/test_logistic_regression.py
@@ -34,7 +34,11 @@ def make_model(
 )
 def make_model_approx(
         model,
-
+        a_,
+        b_,
+        z_,
+        x_,
+        y_,
 ):
     a = np.array([[-1.3], [0.7]])
     b = np.array([-0.5])
@@ -51,27 +55,31 @@ def make_model_approx(
 
     return mp.MeanFieldApproximation.from_kws(
         model,
-        a=mp.NormalMessage.from_mode(
-            np.zeros((n_features, n_dims)), 10),
-        b=mp.NormalMessage.from_mode(
-            np.zeros(n_dims), 10),
-        z=mp.NormalMessage.from_mode(
-            np.zeros((n_obs, n_dims)), 10),
-        x=mp.FixedMessage(x),
-        y=mp.FixedMessage(y)
+        {
+            a_: mp.NormalMessage.from_mode(
+                np.zeros((n_features, n_dims)), 10),
+            b_: mp.NormalMessage.from_mode(
+                np.zeros(n_dims), 10),
+            z_: mp.NormalMessage.from_mode(
+                np.zeros((n_obs, n_dims)), 10),
+            x_: mp.FixedMessage(x),
+            y_: mp.FixedMessage(y)
+        }
     )
 
 
 def test_laplace(
-        model_approx
+        model_approx,
+        a_,
+        b_
 ):
     opt = mp.optimise.LaplaceOptimiser(
         model_approx
     )
     opt.run()
 
-    q_a = opt.model_approx['a']
-    q_b = opt.model_approx['b']
+    q_a = opt.model_approx[a_]
+    q_b = opt.model_approx[b_]
 
     assert q_a.mu[0] == pytest.approx(-1.2, rel=1)
     assert q_a.sigma[0][0] == pytest.approx(0.09, rel=1)
@@ -82,7 +90,9 @@ def test_laplace(
 
 def test_importance_sampling(
         model,
-        model_approx
+        model_approx,
+        a_,
+        b_,
 ):
     sampler = mp.ImportanceSampler(n_samples=500)
     history = {}
@@ -101,8 +111,8 @@ def test_importance_sampling(
             # save and print current approximation
             history[i, factor] = model_approx
 
-    q_a = model_approx['a']
-    q_b = model_approx['b']
+    q_a = model_approx[a_]
+    q_b = model_approx[b_]
 
     assert q_a.mu[0] == pytest.approx(-1.2, rel=1)
     assert q_a.sigma[0][0] == pytest.approx(0.08, rel=1)
