### 2020-08-05 12beb4f
decorator to convert variables to kwargs
