### 2020-08-05 80eb656
finished cleaning up variable -> kwargs dict comprehensions
