+++ autofit/message_passing/factor_graphs/abstract.py
@@ -11,11 +11,14 @@ def accept_variable_dict(func):
     @wraps(func)
     def wrapper(self, variable_dict=None, **kwargs):
         if variable_dict is not None:
-            kwargs = {
-                variable.name: array
+            kwargs.update({
+                variable.name if isinstance(
+                    variable,
+                    Variable
+                ) else variable: array
                 for variable, array
                 in variable_dict.items()
-            }
+            })
         return func(
             self,
             **kwargs
