### 2020-08-05 870a6d5
packaged prior functionality into class
