+++ autofit/message_passing/__init__.py
@@ -1,4 +1,5 @@
 from .factor_graphs import Plate, Variable, Factor, FactorGraph
+from .fascade import MeanFieldPriorModel
 from .mean_field import FactorApproximation, MeanFieldApproximation
 from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, lstsq_laplace_factor_approx
