+++ test_autofit/unit/message_passing/gaussian/test_syntax.py
@@ -41,56 +41,33 @@ def test_gaussian():
         x=x_
     ) == z
 
-    centre = gaussian.centre
-    intensity = gaussian.intensity
-    sigma = gaussian.sigma
-
     likelihood = mp.Factor(
         _likelihood,
         z=z,
         y=y_
     )
 
-    prior_centre = mp.Factor(
-        prior,
-        x=centre
-    )
-    prior_intensity = mp.Factor(
-        prior,
-        x=intensity
-    )
-    prior_sigma = mp.Factor(
-        prior,
-        x=sigma
+    prior_model = mp.MeanFieldPriorModel(
+        likelihood * gaussian,
+        centre=prior,
+        intensity=prior,
+        sigma=prior
     )
 
-    model = likelihood * gaussian * prior_centre * prior_sigma * prior_intensity
-
-    model_approx = mp.MeanFieldApproximation.from_kws(
-        model,
-        {
-            centre: mp.NormalMessage.from_prior(
-                prior
-            ),
-            intensity: mp.NormalMessage.from_prior(
-                prior
-            ),
-            sigma: mp.NormalMessage.from_prior(
-                prior
-            ),
-            x_: mp.FixedMessage(x),
-            y_: mp.FixedMessage(y),
-            z: mp.NormalMessage.from_mode(
-                np.zeros(n_observations), 100
-            ),
-        }
-    )
+    model_approx = prior_model.mean_field_approximation({
+        x_: mp.FixedMessage(x),
+        y_: mp.FixedMessage(y),
+        z: mp.NormalMessage.from_mode(
+            np.zeros(n_observations), 100
+        ),
+    })
 
     opt = mp.optimise.LaplaceOptimiser(
         model_approx,
         n_iter=3
     )
+
     opt.run()
 
-    for variable in (centre, intensity, sigma):
+    for variable in prior_model.prior_variables:
         print(f"{variable.name} = {opt.model_approx[variable].mu}")
