### 2020-08-05 7de0631
compute messages and models at point of use
