### 2020-08-05 ab9227d
handle variable groups being passed in as arguments
