### 2020-08-05 #2550 @dfm
@Jammy2211: Please address Kevin's comments from the pre-review (copied here):

> Fixing those missing and invalid DOI's ☝️. You can call @whedon generate pdf to update the paper and you can call @whedon check references to check those references.
>
> Making your affiliations complete for the paper, i.e. include city and country (do not use acronyms for states and countries).
