+++ paper/paper.bib
@@ -2,6 +2,7 @@
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
   journal={Computing in Science   Engineering},
   title={The NumPy Array: A Structure for Efficient Numerical Computation},
+  doi={ 10.1109/MCSE.2011.37},
   year={2011},
   volume={13},
   number={2},
@@ -97,7 +98,7 @@ year = {2013}
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 @article{Haussler2013,
 abstract = {In this paper, we demonstrate a new method for fitting galaxy profiles which makes use of the full multiwavelength data provided by modern large optical-near-infrared imaging surveys. We present a new version of GALAPAGOS, which utilizes a recently developed multiwavelength version of GALFIT, and enables the automated measurement of wavelength-dependent S{\'{e}}rsic profile parameters for very large samples of galaxies. Our newtechnique is extensively tested to assess the reliability of both pieces of software, GALFIT and GALAPAGOS on both real ugrizY JHK imaging data from the Galaxy And Mass Assembly survey and simulated data made to the same specifications. We find that fitting galaxy light profiles with multiwavelength data increases the stability and accuracy of the measured parameters, and hence produces more complete and meaningful multiwavelength photometry than has been available previously. The improvement is particularly significant for magnitudes in low-S/N bands and for structural parameters like half-light radius re and S{\'{e}}rsic index n for which a prior is used by constraining these parameters to a polynomial as a function of wavelength. This allows the fitting routines to push the magnitude of galaxies for which sensible values can be derived to fainter limits. The technique utilizes a smooth transition of galaxy parameters with wavelength, creating more physically meaningful transitions than single-band fitting and allows accurate interpolation between passbands, perfect for derivation of rest-frame values. {\textcopyright} 2013 The Authors. Published by Oxford University Press on behalf of the Royal Astronomical Society.},
