+++ citations.bib
@@ -78,7 +78,8 @@ year = {2009}
 }
 @ARTICLE{numpy,
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
-  journal={Computing in Science   Engineering},
+  doi={ 10.1109/MCSE.2011.37},
+  journal={Computing in Science Engineering},
   title={The NumPy Array: A Structure for Efficient Numerical Computation},
   year={2011},
   volume={13},
@@ -147,5 +148,5 @@ year = {2014}
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
