+++ paper/paper.bib
@@ -2,7 +2,7 @@
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
   journal={Computing in Science   Engineering},
   title={The NumPy Array: A Structure for Efficient Numerical Computation},
-  doi={ 10.1109/MCSE.2011.37},
+  doi={10.1109/MCSE.2011.37},
   year={2011},
   volume={13},
   number={2},
