+++ citations.bib
@@ -78,7 +78,7 @@ year = {2009}
 }
 @ARTICLE{numpy,
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
-  doi={ 10.1109/MCSE.2011.37},
+  doi={10.1109/MCSE.2011.37},
   journal={Computing in Science Engineering},
   title={The NumPy Array: A Structure for Efficient Numerical Computation},
   year={2011},
