### 2020-08-19 rhayes777/PyAutoFit#114 @arm61
DOC: Suggestion to have `conda env` method mentioned above `pip`
I know I am [not alone](https://xkcd.com/1987/) in having a python environment that is a mess. So having the docs mention to build in a `conda env` before a simple `pip`. I found that I couldn't get `autofit_workspace/examples/simple/fit.py` to work, and I suspect that my *awful* base python environment is to blame.

This is very much just a suggestion and not a problem that would limit the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2550) (no reason you should be put out by my chaotic lifestyle choices).
