### 2020-08-19 rhayes777/PyAutoFit#114 @Jammy2211
We actually had the conda instructions first for a long time, but this caused a lot of issues with compatibility / upgrades on other software projects. I think those are fixed now, but I'm of the view that having to do a full conda build to install the software is a bit cumbersome and would recommend people use a virtual environment to do the installation via pip.
