### 2020-08-19 rhayes777/PyAutoFit#114 @arm61
Sounds fair. conda is rather slow 🐢
