### 2020-08-20 f60c793
Updates to Search output, making it sampler agnostic
