+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/src/__init__.py
@@ -3,4 +3,4 @@ from .model import profiles
 from .fit.fit import FitDataset
 from . import plot
 from .phase.phase import Phase
-from .phase.settings import PhaseSettings
+from .phase.settings import SettingsPhase
