+++ autofit/non_linear/samples.py
@@ -1,4 +1,5 @@
 import csv
+import json
 import math
 from typing import List
 
@@ -11,15 +12,38 @@ from autofit.mapper.model_mapper import ModelMapper
 from autofit.text import model_text
 
 
+def load_from_table(filename, model):
+
+    parameters = []
+    log_likelihoods = []
+    log_priors = []
+    log_posteriors = []
+    weights = []
+
+    with open(filename, "r+") as f:
+        reader = csv.reader(f)
+        for index, row in enumerate(reader):
+            if index > 0:
+                parameters.append(
+                    [float(param) for param in row[0: model.prior_count]]
+                )
+                log_likelihoods.append(float(row[model.prior_count]))
+                log_priors.append(float(row[model.prior_count + 1]))
+                log_posteriors.append(float(row[model.prior_count + 2]))
+                weights.append(float(row[model.prior_count + 3]))
+
+    return parameters, log_likelihoods, log_priors, log_posteriors, weights
+
+
 class OptimizerSamples:
     def __init__(
-            self,
-            model: ModelMapper,
-            parameters: List[List[float]],
-            log_likelihoods: List[float],
-            log_priors: List[float],
-            weights: List[float],
-            time: float = None
+        self,
+        model: ModelMapper,
+        parameters: List[List[float]],
+        log_likelihoods: List[float],
+        log_priors: List[float],
+        weights: List[float],
+        time: float = None,
     ):
         """The *Samples* of a non-linear search, specifically the samples of an search which only provides
         information on the global maximum likelihood solutions, but does not map-out the posterior and thus does
@@ -47,14 +71,22 @@ class OptimizerSamples:
 
     @property
     def parameters_extract(self):
-        return [[params[i] for params in self.parameters] for i in range(self.model.prior_count)]
+        return [
+            [params[i] for params in self.parameters]
+            for i in range(self.model.prior_count)
+        ]
 
     @property
     def _headers(self) -> List[str]:
         """
         Headers for the samples table
         """
-        return self.parameter_names + ["log_likelihood", "log_prior", "log_posterior", "weights"]
+        return self.parameter_names + [
+            "log_likelihood",
+            "log_prior",
+            "log_posterior",
+            "weights",
+        ]
 
     @property
     def _rows(self) -> List[List[float]]:
@@ -66,7 +98,7 @@ class OptimizerSamples:
                 self.log_likelihoods[index],
                 self.log_priors[index],
                 self.log_posteriors[index],
-                self.weights[index]
+                self.weights[index],
             ]
 
     def write_table(self, filename: str):
@@ -83,6 +115,13 @@ class OptimizerSamples:
             writer.writerow(self._headers)
             for row in self._rows:
                 writer.writerow(row)
+
+    def info_to_json(self, filename):
+
+        info = {}
+
+        with open(filename, 'w') as outfile:
+            json.dump(info, outfile)
 
     @property
     def parameter_labels(self):
@@ -147,14 +186,14 @@ class OptimizerSamples:
 
 class PDFSamples(OptimizerSamples):
     def __init__(
-            self,
-            model: ModelMapper,
-            parameters: List[List[float]],
-            log_likelihoods: List[float],
-            log_priors: List[float],
-            weights: List[float],
-            unconverged_sample_size: int = 100,
-            time: float = None,
+        self,
+        model: ModelMapper,
+        parameters: List[List[float]],
+        log_likelihoods: List[float],
+        log_priors: List[float],
+        weights: List[float],
+        unconverged_sample_size: int = 100,
+        time: float = None,
     ):
[…]
