### 2020-08-26 d33a744
multi gaussian model working with mean field prior model
