### 2020-08-26 54108f0
compute variables on the fly
