### 2020-08-26 44ef7dd
assertion and simplification
