+++ autofit/message_passing/fascade.py
@@ -1,13 +1,27 @@
 from abc import ABC, abstractmethod
-from typing import Dict
+from typing import Dict, Callable, List
 
-from .factor_graphs import Variable, Factor
+from .factor_graphs import Variable, Factor, FactorGraph, AbstractNode
 from .mean_field import MeanFieldApproximation
 from .messages import NormalMessage, AbstractMessage
 
 
 class PriorVariable:
-    def __init__(self, prior, variable):
+    def __init__(
+            self,
+            prior: Callable,
+            variable: Variable
+    ):
+        """
+        Comprises a prior and the variable to which it corresponds
+
+        Parameters
+        ----------
+        prior
+            A prior on the variable
+        variable
+            A placeholder for a variable
+        """
         self.variable = variable
         self.factor = Factor(
             prior,
@@ -22,7 +36,20 @@ class AbstractMeanFieldPriorModel(ABC):
     def mean_field_approximation(
             self,
             variable_message_dict: Dict[Variable, AbstractMessage]
-    ):
+    ) -> MeanFieldApproximation:
+        """
+        Create a mean field approximation from this model.
+
+        Parameters
+        ----------
+        variable_message_dict
+            A dictionary mapping variables to messages. This is so messages corresponding
+            to data can be passed in.
+
+        Returns
+        -------
+        A mean-field approximation of this model
+        """
         return MeanFieldApproximation.from_kws(
             self.model,
             {
@@ -31,7 +58,22 @@ class AbstractMeanFieldPriorModel(ABC):
             }
         )
 
-    def __mul__(self, other):
+    def __mul__(
+            self,
+            other: "AbstractMeanFieldPriorModel"
+    ) -> "CompoundMeanFieldPriorModel":
+        """
+        Combine this model with another model
+
+        Parameters
+        ----------
+        other
+            Another model
+
+        Returns
+        -------
+        A combined model
+        """
         return CompoundMeanFieldPriorModel(
             self, other
         )
@@ -53,15 +95,25 @@ class AbstractMeanFieldPriorModel(ABC):
 
 
 class CompoundMeanFieldPriorModel(AbstractMeanFieldPriorModel):
+    """
+    A combination of two or more models
+    """
+
     @property
-    def model(self):
+    def model(self) -> FactorGraph:
+        """
+        Combine the factor graphs of the underlying models
+        """
         model = self.mean_field_prior_models[0].model
         for mean_field_prior_model in self.mean_field_prior_models[1:]:
             model *= mean_field_prior_model.model
         return model
 
     @property
-    def message_dict(self):
+    def message_dict(self) -> Dict[Variable, AbstractMessage]:
+        """
+        Combine the message dictionaries of the underlying models
+        """
         message_dict = dict()
         for mean_field_prior_model in self.mean_field_prior_models:
             message_dict.update(
@@ -73,7 +125,10 @@ class CompoundMeanFieldPriorModel(AbstractMeanFieldPriorModel):
         self.mean_field_prior_models = mean_field_prior_models
 
     @property
-    def prior_variables(self):
+    def prior_variables(self) -> List[PriorVariable]:
+        """
+        Combine the prior-variables of the underlying models
+        """
         return [
             prior_variable
             for model
@@ -86,47 +141,50 @@ class CompoundMeanFieldPriorModel(AbstractMeanFieldPriorModel):
 class MeanFieldPriorModel(AbstractMeanFieldPriorModel):
     def __init__(
             self,
-            model,
-            **kwargs
+            model: AbstractNode,
+            **kwargs: Callable
     ):
+        """
+        Wraps a factor or graph
+
+        Parameters
+        ----------
+        model
+            A factor or graph
+        kwargs
+            Additional model components, such as priors
+        """
[…]
