+++ test_autofit/unit/expectation_propagation/regression/test_linear_regression.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from autofit import message_passing as mp
+from autofit import expectation_propagation as mp
 
 
 @pytest.fixture(
