### 2020-08-26 f9efe81
EmceeSamples oobject created separate from MCMCSamples
