+++ howtofit/chapter_1_introduction/tutorial_4_source_code/src/plot/fit_plots.py
@@ -15,7 +15,7 @@ def data(fit):
     Parameters
     -----------
     Fit : fit.Fit
-        The observed Fit dataset whose data is plotted.
+        The observed Fit _Dataset_ whose data is plotted.
     """
     line_plots.line(xvalues=fit.xvalues, line=fit.data, ylabel="Data Values")
 

[…]
