### 2020-08-26 6f2e358
building towards declarative template
