### 2020-08-26 08d8418
starting to convert a collection of prior models into a graph
