+++ autofit/expectation_propagation/__init__.py
@@ -1,6 +1,6 @@
 from .factor_graphs import Plate, Variable, Factor, FactorGraph
 from .fascade import MeanFieldPriorModel, PriorVariable
-from .declarative import ModelFactor, MessagePassingPriorModel
+from .declarative import ModelFactor
 from .mean_field import FactorApproximation, MeanFieldApproximation
 from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, lstsq_laplace_factor_approx
