+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/tutorial_7_phase_customization.py
@@ -1,0 +1,186 @@
+# %%
+"""
+Tutorial 7: Phase Customization
+===============================
+
+In this tutorial, we're going to add input parameters to a Phase object that customizes the analysis. We'll use the
+specific example of two input parameters that trim our data-set from the left and right before fitting it. This
+example is somewhat trivial (we could achieve almost the same effect with masking), but it will serve to illustrate
+phase customization.
+
+When we customize a phase, we'll use these 'phase settings' to perform phase tagging. Here, we 'tag' the output
+path of the phase's results, such that every time a phase is run with a different customization a new set of
+unique results are stored for those settings. For a given data-set we are thus able to fit it multiple times using
+different settings to compare the results.
+
+These new features have led to an additional module in the 'phase' package called 'settings.py', as well as extensions
+to the 'dataset.py' module. Before looking at these modules, lets first perform a series of Emcee fits to see how
+they change the behaviour of PyAutoFit.
+"""
+
+# %%
+#%matplotlib inline
+
+from autoconf import conf
+import autofit as af
+import numpy as np
+
+from howtofit.chapter_1_introduction.tutorial_7_phase_customization import src as htf
+
+from pyprojroot import here
+
+workspace_path = str(here())
+print("Workspace Path: ", workspace_path)
+
+# %%
+"""
+Setup the configs as we did in the previous tutorial, as well as the output folder for our non-linear search.
+"""
+
+# %%
+conf.instance = conf.Config(
+    config_path=f"{workspace_path}/howtofit/config",
+    output_path=f"{workspace_path}/howtofit/output/chapter_1",
+)
+
+# %%
+"""
+We're now going to perform multiple fits, where each fit trims the data-set that is fitted.
+
+To do this, we'll set up phases with a new class called _SettingsMaskedDataset_, which contains the settings that 
+customize how a _MaskedDataset_ is created. This has two inputs, 'data_trim_left' and 'data_trim_right':
+
+- data_trim_left:
+
+  The dataset's image and noise-map are trimmed and removed from the left (e.g. 1d index values from 0).
+  
+  For example, if the _Dataset_ has shape (100,) and we set data_trim_left=10, the _MaskedDataset_ that is fitted will 
+  have shape (90,). The mask is trimmed in the same way.
+
+- data_trim_right:
+
+  This behaves the same as data_trim_left, however data is removed from the right (e.g. 1D index values from the
+  shape[0] value of the 1D data).
+
+For our first phase, we will omit both of these settings (by setting them to None) and perform the fit from tutorial
+4 where we fit a single _Gaussian_ profile to data composed of a single _Gaussian_ (unlike tutorial 4, we'll use a
+CollectionPriorModel to do this).
+"""
+
+# %%
+settings_masked_dataset = htf.SettingsMaskedDataset(
+    data_trim_left=None, data_trim_right=None
+)
+
+settings = htf.SettingsPhase(settings_masked_dataset=settings_masked_dataset)
+
+phase = htf.Phase(
+    phase_name="phase_t7",
+    profiles=af.CollectionPriorModel(gaussian=htf.profiles.Gaussian),
+    settings=settings,
+    search=af.Emcee(),
+)
+
+# %%
+"""
+Import the simulator module, set up the _Dataset_ and mask and set up the _Dataset_.
+"""
+
+# %%
+from howtofit.simulators.chapter_1 import gaussian_x1
+
+dataset = htf.Dataset(data=gaussian_x1.data, noise_map=gaussian_x1.noise_map)
+mask = np.full(fill_value=False, shape=dataset.data.shape)
+
+print(
[…]
