+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/src/model/profiles.py
@@ -1,0 +1,102 @@
+import numpy as np
+
+"""The 'profiles.py' module is unchanged from the previous tutorial."""
+
+
+class Profile:
+    def __init__(self, centre=0.0, intensity=0.01):
+        """Represents an Abstract 1D profile.
+
+        Parameters
+        ----------
+        centre : float
+            The x coordinate of the profile centre.
+        intensity : float
+            Overall intensity normalisation of the profile.
+        """
+
+        self.centre = centre
+        self.intensity = intensity
+
+
+class Gaussian(Profile):
+    def __init__(
+        self,
+        centre=0.0,  # <- PyAutoFit recognises these constructor arguments
+        intensity=0.1,  # <- are the Gaussian's model parameters.
+        sigma=0.01,
+    ):
+        """Represents a 1D _Gaussian_ profile, which may be treated as a model-component of PyAutoFit the
+        parameters of which are fitted for by a non-linear search.
+
+        Parameters
+        ----------
+        centre : float
+            The x coordinate of the profile centre.
+        intensity : float
+            Overall intensity normalisation of the _Gaussian_ profile.
+        sigma : float
+            The sigma value controlling the size of the Gaussian.
+        """
+
+        super(Gaussian, self).__init__(centre=centre, intensity=intensity)
+
+        self.sigma = sigma
+
+    def profile_from_xvalues(self, xvalues):
+        """
+        Calculate the intensity of the profile on a line of Cartesian x coordinates.
+
+        The input xvalues are translated to a coordinate system centred on the Gaussian, using its centre.
+
+        Parameters
+        ----------
+        values : ndarray
+            The x coordinates in the original reference frame of the grid.
+        """
+        transformed_xvalues = np.subtract(xvalues, self.centre)
+        return np.multiply(
+            np.divide(self.intensity, self.sigma * np.sqrt(2.0 * np.pi)),
+            np.exp(-0.5 * np.square(np.divide(transformed_xvalues, self.sigma))),
+        )
+
+
+class Exponential(Profile):
+    def __init__(
+        self,
+        centre=0.0,  # <- PyAutoFit recognises these constructor arguments are the model
+        intensity=0.1,  # <- parameters of the Gaussian.
+        rate=0.01,
+    ):
+        """Represents a 1D Exponential profile, which may be treated as a model-component of PyAutoFit the
+        parameters of which are fitted for by a non-linear search.
+
+        Parameters
+        ----------
+        centre : float
+            The x coordinate of the profile centre.
+        intensity : float
+            Overall intensity normalisation of the _Gaussian_ profile.
+        ratw : float
+            The decay rate controlling has fast the Exponential declines.
+        """
+
+        super(Exponential, self).__init__(centre=centre, intensity=intensity)
+
+        self.rate = rate
+
+    def profile_from_xvalues(self, xvalues):
+        """
+        Calculate the intensity of the profile on a line of Cartesian x coordinates.
+
+        The input xvalues are translated to a coordinate system centred on the Exponential, using its centre.
+
+        Parameters
+        ----------
+        values : ndarray
+            The x coordinates in the original reference frame of the grid.
+        """
+        transformed_xvalues = np.subtract(xvalues, self.centre)
+        return self.intensity * np.multiply(
+            self.rate, np.exp(-1.0 * self.rate * abs(transformed_xvalues))
+        )
