+++ howtofit/chapter_1_introduction/tutorial_5_visualization_masking/src/model/gaussian.py
@@ -1,0 +1,42 @@
+import numpy as np
+
+
+class Gaussian:
+    def __init__(
+        self,
+        centre=0.0,  # <- PyAutoFit recognises these constructor arguments
+        intensity=0.1,  # <- are the Gaussian's model parameters.
+        sigma=0.01,
+    ):
+        """Represents a 1D _Gaussian_ profile, which may be treated as a model-component of PyAutoFit the
+        parameters of which are fitted for by a non-linear search.
+
+        Parameters
+        ----------
+        centre : float
+            The x coordinate of the profile centre.
+        intensity : float
+            Overall intensity normalisation of the _Gaussian_ profile.
+        sigma : float
+            The sigma value controlling the size of the Gaussian.
+        """
+        self.centre = centre
+        self.intensity = intensity
+        self.sigma = sigma
+
+    def profile_from_xvalues(self, xvalues):
+        """
+        Calculate the intensity of the light profile on a line of Cartesian x coordinates.
+
+        The input xvalues are translated to a coordinate system centred on the Gaussian, using its centre.
+
+        Parameters
+        ----------
+        values : ndarray
+            The x coordinates in the original reference frame of the grid.
+        """
+        transformed_xvalues = np.subtract(xvalues, self.centre)
+        return np.multiply(
+            np.divide(self.intensity, self.sigma * np.sqrt(2.0 * np.pi)),
+            np.exp(-0.5 * np.square(np.divide(transformed_xvalues, self.sigma))),
+        )

[…]
