### 2020-08-27 c9530d6
refactoring and notes
