+++ test_autofit/unit/expectation_propagation/gaussian/test_declarative.py
@@ -1,5 +1,7 @@
 import numpy as np
 
+import pytest
+
 import autofit as af
 import autofit.expectation_propagation as ep
 from .model import Gaussian, make_data
@@ -12,20 +14,17 @@ class FactorModel:
             image_function,
             likelihood_function
     ):
-        self.prior_models = prior_models
-        self.image_function = image_function
-        self.likelihood_function = likelihood_function
-
-    @property
-    def graph(self):
+        self._prior_models = prior_models
+        self._image_function = image_function
+        self._likelihood_function = likelihood_function
         unique_priors = {
             prior: path
             for model
-            in self.prior_models
+            in self._prior_models
             for path, prior
             in model.path_priors_tuples
         }
-        prior_variables = [
+        self._prior_variables = [
             ep.declarative.PriorVariable(
                 "_".join(
                     path
@@ -35,32 +34,66 @@ class FactorModel:
             for prior, path in unique_priors.items()
         ]
 
+    def _node_for_prior_model(
+            self,
+            prior_model
+    ):
+        return ep.ModelFactor(
+            prior_model,
+            self._image_function,
+            self._prior_variables
+        )
+
+    def _graph_for_prior_model(
+            self,
+            prior_model
+    ):
+        z = ep.Variable("z")
+        likelihood_factor = ep.Factor(
+            self._likelihood_function,
+            z=z
+        )
+        model_factor = self._node_for_prior_model(
+            prior_model
+        ) == z
+        return model_factor * likelihood_factor
+
+    @property
+    def graph(self):
         """
-        - Probably need to make an intermediate class for retaining variable state
-        - Each prior model factor should be combined with a unique fitness factor
-        - That means creating y and z variables, with fitness being a function of y and z and prior model factor == z
+        - Test graph with associated fitness function
+        - Test running an actual fit for a Gaussian
+        - Test creating a graph with multiple gaussians and shared priors
         - Also need to generate a dictionary mapping each of the prior variables to an initial message
+        - Can multiple instances of this class be combined? That would allow customisation of image and likelihood
+        functions
         """
 
-        graph = ep.ModelFactor(
-            self.prior_models[0],
-            self.image_function,
-            prior_variables
+        graph = self._graph_for_prior_model(
+            self._prior_models[0],
         )
-        for prior_model in self.prior_models[1:]:
-            graph *= ep.ModelFactor(
-                prior_model,
-                self.image_function,
-                prior_variables
+        for prior_model in self._prior_models[1:]:
+            graph *= self._graph_for_prior_model(
+                prior_model
             )
         return graph
 
 
-def test_factor_model():
-    pass
+@pytest.fixture(
+    name="prior_model"
+)
+def make_prior_model():
+    return af.PriorModel(
+        Gaussian
+    )
 
 
-def test_model_factor():
+@pytest.fixture(
+    name="factor_model"
+)
+def make_factor_model(
+        prior_model
+):
     def image_function(
             instance
     ):
@@ -69,19 +102,30 @@ def test_model_factor():
             x=np.zeros(100)
         )
 
-    model_factor = FactorModel(
-        af.PriorModel(
-            Gaussian
-        ),
-        image_function=image_function,
-        likelihood_function=None
-    )
-    graph = model_factor.graph
+    def likelihood_function(
+            z
+    ):
+        return 1
 
-    result = graph({
-        graph.centre: 1.0,
-        graph.intensity: 0.5,
-        graph.sigma: 0.5
+    return FactorModel(
+        prior_model,
+        image_function=image_function,
+        likelihood_function=likelihood_function
+    )
+
+
+def test_prior_model_node(
+        prior_model,
[…]
