### 2020-08-27 279839f
factor graph includes factors for priors
