### 2020-08-27 6d95e3c
mean field approximation
