### 2020-09-02 5df5459
attempting custom factor and likelihood model
