+++ test_autofit/unit/expectation_propagation/gaussian/test_declarative.py
@@ -1,3 +1,5 @@
+from typing import List
+
 import numpy as np
 import pytest
 
@@ -9,19 +11,15 @@ from test_autofit.unit.expectation_propagation.gaussian.model import Gaussian, m
 class FactorModel:
     def __init__(
             self,
-            *prior_models: af.PriorModel,
-            image_function,
-            likelihood_function
+            likelihood_models: List["LikelihoodModel"]
     ):
-        self._prior_models = prior_models
-        self._image_function = image_function
-        self._likelihood_function = likelihood_function
+        self.likelihood_models = likelihood_models
         self._unique_priors = {
             prior: path
-            for model
-            in self._prior_models
+            for prior_model
+            in self.prior_models
             for path, prior
-            in model.path_priors_tuples
+            in prior_model.path_priors_tuples
         }
         self._prior_variables = [
             ep.declarative.PriorVariable(
@@ -32,17 +30,18 @@ class FactorModel:
             )
             for prior, path in self._unique_priors.items()
         ]
-
-        self._deterministic_variables = {
-            prior_model: ep.Variable(
-                "z",
-                ep.Plate(
-                    "observations"
-                )
-            )
-            for prior_model
-            in prior_models
+        self._prior_variable_map = {
+            prior_variable.prior: prior_variable
+            for prior_variable in self._prior_variables
         }
+
+    @property
+    def prior_models(self):
+        return [
+            model.prior_model
+            for model
+            in self.likelihood_models
+        ]
 
     @property
     def prior_factors(self):
@@ -56,56 +55,31 @@ class FactorModel:
         ]
 
     @property
-    def shape(self):
-        return self._image_function(
-            self._prior_models[0].instance_from_prior_medians()
-        ).shape
-
-    @property
     def message_dict(self):
         return {
-            **{
-                variable: ep.NormalMessage.from_mode(
-                    np.zeros(self.shape),
-                    100
-                )
-                for variable
-                in self._deterministic_variables.values()
-            },
-            **{
-                variable: ep.NormalMessage.from_prior(
-                    variable.prior
-                )
-                for variable
-                in self._prior_variables
-            }
+            variable: ep.NormalMessage.from_prior(
+                variable.prior
+            )
+            for variable
+            in self._prior_variables
         }
 
-    def _node_for_prior_model(
+    def _node_for_likelihood_model(
             self,
-            prior_model
+            likelihood_model: "LikelihoodModel"
     ):
-        return ep.ModelFactor(
-            prior_model,
-            self._image_function,
-            self._prior_variables
-        )
-
-    def _graph_for_prior_model(
-            self,
-            prior_model,
-    ):
-        z = self._deterministic_variables[
-            prior_model
+        prior_variables = [
+            self._prior_variable_map[
+                prior
+            ]
+            for prior
+            in likelihood_model.prior_model.priors
         ]
-        likelihood_factor = ep.Factor(
-            self._likelihood_function,
-            z=z
+        return ep.ModelFactor(
+            likelihood_model.prior_model,
+            likelihood_model.likelihood_function,
+            prior_variables
         )
-        model_factor = self._node_for_prior_model(
-            prior_model
-        ) == z
-        return model_factor * likelihood_factor
 
     @property
     def graph(self):
@@ -118,12 +92,12 @@ class FactorModel:
         functions
         """
 
-        graph = self._graph_for_prior_model(
-            self._prior_models[0],
+        graph = self._node_for_likelihood_model(
+            self.likelihood_models[0],
         )
-        for prior_model in self._prior_models[1:]:
[…]
