+++ test_autofit/unit/tools/test_phase.py
@@ -6,7 +6,7 @@ def test__phase_settings_tag():
     assert settings.log_likelihood_cap_tag == ""
 
     settings = af.AbstractSettingsPhase(log_likelihood_cap=10.0)
-    assert settings.log_likelihood_cap_tag == "__lh_cap_10.0"
+    assert settings.log_likelihood_cap_tag == "_lh_cap_10.0"
 
     settings = af.AbstractSettingsPhase(log_likelihood_cap=200.0001)
-    assert settings.log_likelihood_cap_tag == "__lh_cap_200.0"
+    assert settings.log_likelihood_cap_tag == "_lh_cap_200.0"
