+++ requirements.txt
@@ -1,4 +1,4 @@
-autoconf==0.6.1
+autoconf==0.6.2
 corner>=2.0.1
 decorator>=4.2.1
 dill>=0.3.1.1
