### 2020-09-09 e48956c
prior model based factor nodes, callable with prior_id named variables
