### 2020-09-09 0ffbbd5
taking mean -> resolve function shape issue without change to source doe
