+++ autofit/expectation_propagation/factor_graphs/factor.py
@@ -1,4 +1,3 @@
-import logging
 from abc import ABC
 from inspect import getfullargspec
 from itertools import chain, repeat
@@ -9,10 +8,6 @@ import numpy as np
 from autofit.expectation_propagation.factor_graphs.abstract import AbstractNode, accept_variable_dict
 from autofit.expectation_propagation.factor_graphs.numerical import numerical_jacobian
 from autofit.expectation_propagation.factor_graphs.variable import Variable
-
-logger = logging.getLogger(
-    __name__
-)
 
 
 class FactorValue(NamedTuple):
@@ -323,18 +318,12 @@ class Factor(AbstractFactor):
         Object encapsulating the result of the function call
         """
         val = self._call_factor(**kwargs)
-        try:
-            val = val.reshape(
+        return FactorValue(
+            val.reshape(
                 self._function_shape(
                     **kwargs
                 )
-            )
-        except ValueError:
-            logger.debug(
-                "Could not reshape value"
-            )
-        return FactorValue(
-            val, {}
+            ), {}
         )
 
     def broadcast_variable(
