### 2020-09-09 65e1b24
put assertions in the test
