+++ autofit/expectation_propagation/mean_field.py
@@ -6,8 +6,7 @@ from typing import (
 
 import numpy as np
 
-from autofit.expectation_propagation import Variable
-from autofit.expectation_propagation.factor_graphs import Factor
+from autofit.expectation_propagation.factor_graphs import Factor, Variable
 from autofit.expectation_propagation.factor_graphs.graph import FactorGraph
 from autofit.expectation_propagation.messages import FixedMessage
 from autofit.expectation_propagation.messages.abstract import AbstractMessage
