+++ test_autofit/unit/expectation_propagation/gaussian/test_declarative.py
@@ -1,133 +1,9 @@
-from typing import List
-
 import numpy as np
 import pytest
 
 import autofit as af
 import autofit.expectation_propagation as ep
 from test_autofit.unit.expectation_propagation.gaussian.model import Gaussian, make_data, _likelihood
-
-
-class FactorModel:
-    def __init__(
-            self,
-            likelihood_models: List["LikelihoodModel"]
-    ):
-        self.likelihood_models = likelihood_models
-        self._unique_priors = {
-            prior: path
-            for prior_model
-            in self.prior_models
-            for path, prior
-            in prior_model.path_priors_tuples
-        }
-        self._prior_variables = [
-            ep.declarative.PriorVariable(
-                f"prior_{prior.id}",
-                prior
-            )
-            for prior, path in self._unique_priors.items()
-        ]
-        self._prior_variable_map = {
-            prior_variable.prior: prior_variable
-            for prior_variable in self._prior_variables
-        }
-
-    @property
-    def prior_variables(self):
-        return self._prior_variables
-
-    @property
-    def prior_models(self):
-        return [
-            model.prior_model
-            for model
-            in self.likelihood_models
-        ]
-
-    @property
-    def prior_factors(self):
-        return [
-            ep.Factor(
-                variable.prior,
-                x=variable
-            )
-            for variable
-            in self._prior_variables
-        ]
-
-    @property
-    def message_dict(self):
-        return {
-            variable: ep.NormalMessage.from_prior(
-                variable.prior
-            )
-            for variable
-            in self._prior_variables
-        }
-
-    def _node_for_likelihood_model(
-            self,
-            likelihood_model: "LikelihoodModel"
-    ):
-        prior_variables = [
-            self._prior_variable_map[
-                prior
-            ]
-            for prior
-            in likelihood_model.prior_model.priors
-        ]
-        return ep.ModelFactor(
-            likelihood_model.prior_model,
-            likelihood_model.likelihood_function,
-            prior_variables
-        )
-
-    @property
-    def graph(self):
-        """
-        - Test graph with associated fitness function
-        - Test running an actual fit for a Gaussian
-        - Test creating a graph with multiple gaussians and shared priors
-        - Also need to generate a dictionary mapping each of the prior variables to an initial message
-        - Can multiple instances of this class be combined? That would allow customisation of image and likelihood
-        functions
-        """
-
-        graph = self._node_for_likelihood_model(
-            self.likelihood_models[0],
-        )
-
-        for likelihood_model in self.likelihood_models[1:]:
-            graph *= self._node_for_likelihood_model(
-                likelihood_model
-            )
-        for prior_factor in self.prior_factors:
-            graph *= prior_factor
-        return graph
-
-    @property
-    def mean_field_approximation(self):
-        return ep.MeanFieldApproximation.from_kws(
-            self.graph,
-            self.message_dict
-        )
-
-    def __mul__(self, other: "FactorModel"):
-        return FactorModel(
-            other.likelihood_models + self.likelihood_models
-        )
-
-
-class LikelihoodModel(FactorModel):
-    def __init__(
-            self,
-            prior_model,
-            likelihood_function
-    ):
-        self.prior_model = prior_model
-        self.likelihood_function = likelihood_function
-        super().__init__([self])
 
 
 def test_gaussian():
@@ -167,7 +43,7 @@ def test_gaussian():
             )
         )
 
-    factor_model = LikelihoodModel(
+    factor_model = ep.LikelihoodModel(
         prior_model,
         likelihood_function=likelihood_function
     )
@@ -212,7 +88,7 @@ def make_factor_model(
     ):
         return 1
 
-    return LikelihoodModel(
+    return ep.LikelihoodModel(
         prior_model,
         likelihood_function=likelihood_function
     )
