+++ autofit/expectation_propagation/declarative.py
@@ -1,13 +1,16 @@
 from typing import Callable
+from typing import List
 
-import autofit as af
-from autofit import expectation_propagation as ep
+from autofit.expectation_propagation.factor_graphs.factor import Factor, Variable
+from autofit.expectation_propagation.mean_field import MeanFieldApproximation
+from autofit.expectation_propagation.messages import NormalMessage
+from autofit.mapper.prior_model.prior_model import Prior, PriorModel
 
 
-class ModelFactor(ep.Factor):
+class ModelFactor(Factor):
     def __init__(
             self,
-            prior_model: af.PriorModel,
+            prior_model: PriorModel,
             likelihood_function: Callable,
             prior_variables
     ):
@@ -38,11 +41,133 @@ class ModelFactor(ep.Factor):
         self.prior_model = prior_model
 
 
-class PriorVariable(ep.Variable):
+class PriorVariable(Variable):
     def __init__(
             self,
             name: str,
-            prior: af.Prior
+            prior: Prior
     ):
         super().__init__(name)
         self.prior = prior
+
+
+class FactorModel:
+    def __init__(
+            self,
+            likelihood_models: List["LikelihoodModel"]
+    ):
+        self.likelihood_models = likelihood_models
+        self._unique_priors = {
+            prior: path
+            for prior_model
+            in self.prior_models
+            for path, prior
+            in prior_model.path_priors_tuples
+        }
+        self._prior_variables = [
+            PriorVariable(
+                f"prior_{prior.id}",
+                prior
+            )
+            for prior, path in self._unique_priors.items()
+        ]
+        self._prior_variable_map = {
+            prior_variable.prior: prior_variable
+            for prior_variable in self._prior_variables
+        }
+
+    @property
+    def prior_variables(self):
+        return self._prior_variables
+
+    @property
+    def prior_models(self):
+        return [
+            model.prior_model
+            for model
+            in self.likelihood_models
+        ]
+
+    @property
+    def prior_factors(self):
+        return [
+            Factor(
+                variable.prior,
+                x=variable
+            )
+            for variable
+            in self._prior_variables
+        ]
+
+    @property
+    def message_dict(self):
+        return {
+            variable: NormalMessage.from_prior(
+                variable.prior
+            )
+            for variable
+            in self._prior_variables
+        }
+
+    def _node_for_likelihood_model(
+            self,
+            likelihood_model: "LikelihoodModel"
+    ):
+        prior_variables = [
+            self._prior_variable_map[
+                prior
+            ]
+            for prior
+            in likelihood_model.prior_model.priors
+        ]
+        return ModelFactor(
+            likelihood_model.prior_model,
+            likelihood_model.likelihood_function,
+            prior_variables
+        )
+
+    @property
+    def graph(self):
+        """
+        - Test graph with associated fitness function
+        - Test running an actual fit for a Gaussian
+        - Test creating a graph with multiple gaussians and shared priors
+        - Also need to generate a dictionary mapping each of the prior variables to an initial message
+        - Can multiple instances of this class be combined? That would allow customisation of image and likelihood
+        functions
+        """
+
+        graph = self._node_for_likelihood_model(
+            self.likelihood_models[0],
+        )
+
+        for likelihood_model in self.likelihood_models[1:]:
+            graph *= self._node_for_likelihood_model(
+                likelihood_model
+            )
+        for prior_factor in self.prior_factors:
+            graph *= prior_factor
+        return graph
+
+    @property
+    def mean_field_approximation(self):
+        return MeanFieldApproximation.from_kws(
+            self.graph,
+            self.message_dict
+        )
+
+    def __mul__(self, other: "FactorModel"):
+        return FactorModel(
[…]
