### 2020-09-09 d5ddc2d
docs and simplification
