### 2020-09-09 d896c5c
renamed expectation propagation to graphical
