+++ test_autofit/unit/graphical/regression/conftest.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from scipy import stats
 
-from autofit import expectation_propagation as mp
+from autofit import graphical as mp
 
 
 @pytest.fixture(
