+++ test_autofit/unit/graphical/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from scipy import stats
 
 
-from autofit import expectation_propagation as mp
+from autofit import graphical as mp
 
 
 @pytest.fixture(
