+++ test_autofit/unit/graphical/test_projection.py
@@ -2,15 +2,15 @@ import numpy as np
 import pytest
 from scipy import integrate
 
-import autofit.expectation_propagation.messages.normal
-from autofit import expectation_propagation as mp
+import autofit.graphical.messages.normal
+from autofit import graphical as mp
 
 
 @pytest.fixture(
     name="q_cavity"
 )
 def make_q_cavity():
-    return autofit.expectation_propagation.messages.normal.NormalMessage(-0.5, 0.5)
+    return autofit.graphical.messages.normal.NormalMessage(-0.5, 0.5)
 
 
 def test_integration(
@@ -32,7 +32,7 @@ def test_integration(
     ni_0, ni_1, ni_2 = (
         integrate.trapz(x ** i * tilted_distribution, x) for i in range(3))
 
-    q_numerical = autofit.expectation_propagation.messages.normal.NormalMessage.from_sufficient_statistics(
+    q_numerical = autofit.graphical.messages.normal.NormalMessage.from_sufficient_statistics(
         [ni_1 / ni_0, ni_2 / ni_0]
     )
 
@@ -69,7 +69,7 @@ def test_laplace_method(probit_factor, q_cavity, x):
     opt_probit = mp.OptFactor.from_approx(probit_approx)
     result = opt_probit.maximise({x: 0.})
 
-    q_probit_laplace = autofit.expectation_propagation.messages.normal.NormalMessage.from_mode(
+    q_probit_laplace = autofit.graphical.messages.normal.NormalMessage.from_mode(
         result.mode[x],
         covariance=result.inv_hessian[x]
     )
