### 2020-09-09 353cbe6
priors are a kind of variable
