### 2020-09-09 17c2fdc
eliminated prior variables
