+++ autofit/graphical/__init__.py
@@ -1,6 +1,5 @@
-from .declarative import ModelFactor, LikelihoodModel
+from .declarative import ModelFactor, LikelihoodModelCollection
 from .factor_graphs import Factor, FactorGraph, AbstractFactor, FactorValue
-from autofit.mapper.variable import Variable, Plate
 from .mean_field import FactorApproximation, MeanFieldApproximation
 from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, lstsq_laplace_factor_approx
