### 2020-09-11 9ddb19f
removing setting log_norm for project_on_to_factor_approx
