### 2020-09-11 ba15947
adding Variable and Plate to autofit.graphical import
