+++ autofit/graphical/__init__.py
@@ -4,3 +4,4 @@ from .mean_field import FactorApproximation, MeanFieldApproximation
 from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, lstsq_laplace_factor_approx
 from .sampling import ImportanceSampler, project_factor_approx_sample
+from ..mapper.variable import Variable, Plate
