### 2020-09-11 4504101
fixing bug with Factor._function_shape where function shape would be incorrect if shape of inputs changed, added test to verify correctness
