### 2020-09-11 0b54f63
adding test of vectorisation for factors
