+++ howtofit/chapter_1_introduction/tutorial_6_complex_models/tutorial_6_complex_models.py
@@ -21,7 +21,7 @@ generate the model-data we generate it as the sum of all individual profiles in 
 
 from autoconf import conf
 import autofit as af
-from autofit_workspace.howtofit.chapter_1_introduction.tutorial_6_complex_models import (
+from howtofit.chapter_1_introduction.tutorial_6_complex_models import (
     src as htf,
 )
 
@@ -140,7 +140,7 @@ and Exponential profile.
 """
 
 # %%
-from autofit_workspace.howtofit.simulators.chapter_1 import gaussian_x1_exponential_x1
+from howtofit.simulators.chapter_1 import gaussian_x1_exponential_x1
 
 dataset = htf.Dataset(
     data=gaussian_x1_exponential_x1.data, noise_map=gaussian_x1_exponential_x1.noise_map
@@ -172,7 +172,7 @@ phase = htf.Phase(
 
 print(
     "Emcee has begun running - checkout the autofit_workspace/howtofit/chapter_1_introduction/output/phase_t5_gaussian_x1_exponential_x1"
-    "folder for live output of the results."
+    " folder for live output of the results."
     "This Jupyter notebook cell with progress once Emcee has completed - this could take a few minutes!"
 )
 
@@ -190,7 +190,7 @@ Lets fit a model composed of two Gaussians and and an Exponential, which will ha
 """
 
 # %%
-from autofit_workspace.howtofit.simulators.chapter_1 import gaussian_x2_exponential_x1
+from howtofit.simulators.chapter_1 import gaussian_x2_exponential_x1
 
 dataset = htf.Dataset(
     data=gaussian_x2_exponential_x1.data, noise_map=gaussian_x2_exponential_x1.noise_map
@@ -208,7 +208,7 @@ phase = htf.Phase(
 
 print(
     "Emcee has begun running - checkout the autofit_workspace/howtofit/chapter_1_introduction/output/phase_t5_gaussian_x2_exponential_x1"
-    "folder for live output of the results."
+    " folder for live output of the results."
     "This Jupyter notebook cell with progress once Emcee has completed - this could take a few minutes!"
 )
 
@@ -257,7 +257,7 @@ We can now fit this model using a phase as per usual.
 """
 
 # %%
-from autofit_workspace.howtofit.simulators.chapter_1 import gaussian_x3
+from howtofit.simulators.chapter_1 import gaussian_x3
 
 dataset = htf.Dataset(data=gaussian_x3.data, noise_map=gaussian_x3.noise_map)
 
@@ -265,7 +265,7 @@ phase = htf.Phase(phase_name="phase_t6_gaussian_x3", profiles=model, search=af.E
 
 print(
     "Emcee has begun running - checkout the autofit_workspace/howtofit/chapter_1_introduction/output/phase_t5_gaussian_x3"
-    "folder for live output of the results."
+    " folder for live output of the results."
     "This Jupyter notebook cell with progress once Emcee has completed - this could take a few minutes!"
 )
 
