+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/tutorial_7_phase_customization.py
@@ -25,7 +25,7 @@ from autoconf import conf
 import autofit as af
 import numpy as np
 
-from autofit_workspace.howtofit.chapter_1_introduction.tutorial_7_phase_customization import (
+from howtofit.chapter_1_introduction.tutorial_7_phase_customization import (
     src as htf,
 )
 
@@ -89,14 +89,14 @@ Import the simulator module, set up the _Dataset_ and mask and set up the _Datas
 """
 
 # %%
-from autofit_workspace.howtofit.simulators.chapter_1 import gaussian_x1
+from howtofit.simulators.chapter_1 import gaussian_x1
 
 dataset = htf.Dataset(data=gaussian_x1.data, noise_map=gaussian_x1.noise_map)
 mask = np.full(fill_value=False, shape=dataset.data.shape)
 
 print(
     "Emcee has begun running - checkout the autofit_workspace/howtofit/chapter_1_introduction/output/phase_t6"
-    "folder for live output of the results."
+    " folder for live output of the results."
     "This Jupyter notebook cell with progress once Emcee has completed - this could take a few minutes!"
 )
 
@@ -140,7 +140,7 @@ phase = htf.Phase(
 
 print(
     "Emcee has begun running - checkout the autofit_workspace/howtofit/chapter_1_introduction/output/phase_t6"
-    "folder for live output of the results."
+    " folder for live output of the results."
     "This Jupyter notebook cell with progress once Emcee has completed - this could take a few minutes!"
 )
 
