+++ docs/conf.py
@@ -17,9 +17,12 @@ workspace_path = str(here())
 
 import os
 import sys
+import shutil
+
 sys.path.insert(0, os.path.abspath('.'))
 
 os.environ['WORKSPACE'] = f"{str(workspace_path)}"
+shutil.copytree(f"{workspace_path}/howtolens/config", workspace_path)
 
 # -- Project information -----------------------------------------------------
 
