+++ docs/conf.py
@@ -16,7 +16,7 @@ from pyprojroot import here
 workspace_path = str(here())
 
 import os
-import sys
+import sys 
 import shutil
 
 sys.path.insert(0, os.path.abspath('.'))
