+++ docs/conf.py
@@ -16,13 +16,11 @@ from pyprojroot import here
 workspace_path = str(here())
 
 import os
-import sys 
-import shutil
+import sys
 
 sys.path.insert(0, os.path.abspath('.'))
 
 os.environ['WORKSPACE'] = f"{str(workspace_path)}"
-shutil.copytree(f"{workspace_path}/../howtolens/config", workspace_path)
 
 # -- Project information -----------------------------------------------------
 
