+++ howtofit/chapter_1_introduction/tutorial_6_complex_models/tutorial_6_complex_models.py
@@ -39,8 +39,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 
 # %%
 conf.instance = conf.Config(
-    config_path=f"{workspace_path}/howtofit/config",
-    output_path=f"{workspace_path}/howtofit/output/chapter_1",
+    config_path=f"{workspace_path}/config",
+    output_path=f"{workspace_path}/output/chapter_1",
 )
 
 # %%
