+++ howtofit/chapter_1_introduction/tutorial_2_model_fitting/tutorial_2_model_fitting.py
@@ -24,7 +24,7 @@ These setup the configs as we did in the previous tutorial.
 """
 
 # %%
-conf.instance = conf.Config(config_path=f"{workspace_path}/howtofit/config")
+conf.instance = conf.Config(config_path=f"{workspace_path}/config")
 
 # %%
 """
