+++ docs/conf.py
@@ -20,7 +20,7 @@ import sys
 
 sys.path.insert(0, os.path.abspath('.'))
 
-os.environ['WORKSPACE'] = f"{str(workspace_path)}"
+os.environ['WORKSPACE'] = f"{str(workspace_path)}/howtofit"
 
 # -- Project information -----------------------------------------------------
 
