### 2020-09-14 #2550 @arm61
@dfm I was wondering are things like the "statement of need" (that is mentioned in the Documentation portion of this checklist) is expected to be in the code docs and the paper?
