### 2020-09-14 rhayes777/PyAutoFit#126 @arm61
How do I run tests?
For JOSS, I think there should be details of how I (the user) would run the tests if I were to install PyAutoFit from the source.
