### 2020-09-14 rhayes777/PyAutoFit#127 @arm61
Github Issues as well as Slack
Opening a github issue is probably a lower barrier to entry for most than using (and making an account for) Slack if someone has a bug.
