### 2020-09-14 #2550 @arm61
Once issues [127](https://github.com/rhayes777/PyAutoFit/issues/127) and [126](https://github.com/rhayes777/PyAutoFit/issues/126) are closed, @dfm has clarified my comments, and the referencing problems are solved I think I am happy.

An only small point in the "state of the field" bit for the paper, is that it might be worth mentioning [`uravu`](uravu.readthedocs.io) which has similar functionality to PyAutoFit. This is not a hard issue because `uravu` is my own package and I don't wanna be an awful person.  However, I think that the similarity between the packages mean that it is worth a mention, however, I will leave that up to @Jammy2211.
