### 2020-09-15 #2550 @dfm
@arm61: Thanks for the update! The statement of need should be in *both* the paper and the documentation. Let me know if you need more clarification.
