### 2020-09-15 #2550 @Jammy2211
Great - I think I pretty much agree with all your suggestions and will get everything sorted this week :). You've definitely brought up a couple of aspects of the docs that need some polishing!

No problem on the length of time, I think everyone's stretched pretty thin at the moment and just trying to get everything done whenever its possible.
