### 2020-09-16 d897390
Adding class  to mean_field.py to encode mean field approximations to factors
