+++ autofit/graphical/__init__.py
@@ -1,6 +1,6 @@
 from .declarative import ModelFactor, ModelFactorCollection
 from .factor_graphs import Factor, FactorGraph, AbstractFactor, FactorValue
-from .mean_field import FactorApproximation, MeanFieldApproximation
+from .mean_field import FactorApproximation, MeanField, MeanFieldApproximation
 from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, lstsq_laplace_factor_approx
 from .sampling import ImportanceSampler, project_factor_approx_sample
