### 2020-09-18 84e032e
getting a kind of result out with instance and model but without shared priors
