### 2020-09-18 b0e9a88
optimisation gives back a prior model collection
