### 2020-09-18 6ce64b0
docs and slightly more concise
