+++ test_autofit/unit/text/test_formatter.py
@@ -13,119 +13,149 @@ text_path = "{}/files/text/".format(os.path.dirname(os.path.realpath(__file__)))
 
 @pytest.fixture(scope="session", autouse=True)
 def do_something():
-    conf.instance = conf.Config(
-        config_path="{}/files/config/text".format(directory)
+    conf.instance = conf.Config(config_path="{}/files/config/text".format(directory))
+
+
+def test__add_whitespace_between_strings():
+    str0 = af.text.formatter.add_whitespace(str0="param0", str1="mass", whitespace=10)
+
+    assert str0 == "param0    mass"
+
+    str0 = af.text.formatter.add_whitespace(str0="param0", str1="mass", whitespace=20)
+
+    assert str0 == "param0              mass"
+
+
+def test__parameter_name_and_value_string():
+    str0 = af.text.formatter.parameter_result_string_from(
+        parameter_name="param0", value=2.0, whitespace=10
+    )
+    assert str0 == "param0    2.00"
+
+    str0 = af.text.formatter.parameter_result_string_from(
+        parameter_name="param11", value=3.00, whitespace=20
     )
 
+    assert str0 == "param11             3.0000"
 
-def test__label_and_label_string():
-    string0 = af.text.formatter.label_and_label_string(
-        label0="param0", label1="mass", whitespace=10
-    )
-    string1 = af.text.formatter.label_and_label_string(
-        label0="param00", label1="mass0", whitespace=10
-    )
-    string2 = af.text.formatter.label_and_label_string(
-        label0="param000", label1="mass111", whitespace=10
+    str0 = af.text.formatter.parameter_result_string_from(
+        parameter_name="param12", value=3.00, whitespace=15
     )
 
-    assert string0 == "param0    mass"
-    assert string1 == "param00   mass0"
-    assert string2 == "param000  mass111"
+    assert str0 == "param12        3.00e+00"
 
-    string0 = af.text.formatter.label_and_label_string(
-        label0="param0", label1="mass", whitespace=20
+    str0 = af.text.formatter.parameter_result_string_from(
+        parameter_name="param0", value=3.00, whitespace=15, name_to_label=True
     )
 
-    assert string0 == "param0              mass"
+    assert str0 == "p0             3.00"
 
 
-def test__label_and_value_string():
-    string0 = af.text.formatter.label_and_value_string(
-        label="param0", value=2.0, whitespace=10
+def test__parameter_name_value_and_limits_string():
+    str0 = af.text.formatter.parameter_result_string_from(
+        parameter_name="param0", value=2.0, values_at_sigma=(1.5, 2.5), whitespace=10
     )
-    string1 = af.text.formatter.label_and_value_string(
-        label="param00", value=2.0, whitespace=10
-    )
-    string2 = af.text.formatter.label_and_value_string(
-        label="param000", value=2.0, whitespace=10
-    )
+    assert str0 == "param0    2.00 (1.50, 2.50)"
 
-    assert string0 == "param0    2.00"
-    assert string1 == "param00   2.00"
-    assert string2 == "param000  2.00"
-
-    string = af.text.formatter.label_and_value_string(
-        label="param11", value=3.00, whitespace=20
-    )
-
-    assert string == "param11             3.0000"
-
-    string = af.text.formatter.label_and_value_string(
-        label="param12", value=3.00, whitespace=15
-    )
-
-    assert string == "param12        3.00e+00"
-
-
-def test__label_value_and_limits_string():
-    string0 = af.text.formatter.label_value_and_limits_string(
[…]
