+++ howtofit/chapter_2_results/scripts/tutorial_1_results_and_samples.py
@@ -159,7 +159,7 @@ for labeling figures.
 """
 
 # %%
-print(samples.parameter_names)
+print(samples.model_component_and_parameter_names)
 print(samples.parameter_labels)
 print("\n")
 
