+++ test_autofit/unit/config/notation/label_format.ini
@@ -7,4 +7,4 @@ param12 = {:.2e}
 one = {:.2f}
 two = {:.2f}
 radius = radius_value
-mass = mass_value
+mass = {:.2f}
