### 2020-09-18 8be9f89
removed workspace assertion
