### 2020-09-18 7d689cc
removed paths functionality from phase and fixed all the tests
