+++ autofit/non_linear/mock/mock_search.py
@@ -14,7 +14,6 @@ class MockSearch(NonLinearSearch):
 
     @convert_paths
     def __init__(self, paths=None, samples=None, fit_fast=True):
-
         super().__init__(paths=paths)
 
         self.fit_fast = fit_fast
