+++ test_autofit/mock.py
@@ -91,7 +91,8 @@ class MockSamples(af.PDFSamples):
 
 
 class MockSearch(af.NonLinearSearch):
-    def __init__(self, paths=None, samples=None):
+    def __init__(self, paths=None, samples=None, phase_name=None):
+        self.phase_name = phase_name
         super().__init__(paths=paths)
 
         self.samples = samples or MockSamples()
@@ -110,7 +111,6 @@ class MockSearch(af.NonLinearSearch):
 
                 # Return Chi squared
                 return -2 * log_likelihood
-
 
         fitness_function = Fitness(model.instance_from_vector)
         fitness_function(model.prior_count * [0.5])
