### 2020-09-19 5e337ff
refactored sampes text
