+++ autofit/text/samples_text.py
@@ -1,12 +1,17 @@
 import logging
 
-from autoconf import conf
 from autofit.text import formatter as frm
 
 logger = logging.getLogger(__name__)
 
 
-def parameter_results_at_sigma_summary(samples, sigma) -> str:
+def values_from_samples(samples, median_pdf_model):
+
+    if median_pdf_model:
+        return samples.median_pdf_vector
+    return samples.max_log_likelihood_vector
+
+def summary(samples, sigma=3.0, median_pdf_model=True, name_to_label=True) -> str:
     """ Create a string summarizing the results of the non-linear search at an input sigma value.
 
     This function is used for creating the model.results files of a non-linear search.
@@ -15,6 +20,8 @@ def parameter_results_at_sigma_summary(samples, sigma) -> str:
     ----------
     sigma : float
         The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF)."""
+
+    values = values_from_samples(samples=samples, median_pdf_model=median_pdf_model)
     values_at_sigma = samples.vector_at_sigma(sigma=sigma)
 
     sigma_formatter = frm.TextFormatter()
@@ -23,8 +30,9 @@ def parameter_results_at_sigma_summary(samples, sigma) -> str:
 
         value = frm.parameter_result_string_from(
             parameter_name=samples.model.parameter_names[i],
-            value=samples.median_pdf_vector[i],
+            value=values[i],
             values_at_sigma=values_at_sigma[i],
+            name_to_label=name_to_label,
         )
 
         sigma_formatter.add((prior_path, value))
@@ -34,7 +42,7 @@ def parameter_results_at_sigma_summary(samples, sigma) -> str:
     )
 
 
-def parameter_results_at_sigma_table(samples, sigma, name_to_label=True) -> str:
+def table(samples, median_pdf_model=True, sigma=3.0, name_to_label=True) -> str:
     """ Create a string summarizing the results of the non-linear search at an input sigma value.
 
     This function is used for creating the model.results files of a non-linear search.
@@ -44,6 +52,7 @@ def parameter_results_at_sigma_table(samples, sigma, name_to_label=True) -> str:
     sigma : float
         The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF)."""
 
+    values = values_from_samples(samples=samples, median_pdf_model=median_pdf_model)
     values_at_sigma = samples.vector_at_sigma(sigma=sigma)
 
     table = []
@@ -52,7 +61,7 @@ def parameter_results_at_sigma_table(samples, sigma, name_to_label=True) -> str:
 
         label_value = frm.parameter_result_string_from(
             parameter_name=samples.model.parameter_names[i],
-            value=samples.median_pdf_vector[i],
+            value=values[i],
             values_at_sigma=values_at_sigma[i],
             subscript=samples.model.subscripts[i],
             name_to_label=name_to_label,
@@ -64,7 +73,7 @@ def parameter_results_at_sigma_table(samples, sigma, name_to_label=True) -> str:
     return "\n\nMedian PDF model Table ({} sigma limits):\n\n{}".format(sigma, table)
 
 
-def parameter_results_at_sigma_latex(samples, sigma, name_to_label=True) -> str:
+def latex(samples, median_pdf_model=True, sigma=3.0, name_to_label=True) -> str:
     """ Create a string summarizing the results of the non-linear search at an input sigma value.
 
     This function is used for creating the model.results files of a non-linear search.
@@ -74,6 +83,7 @@ def parameter_results_at_sigma_latex(samples, sigma, name_to_label=True) -> str:
     sigma : float
[…]
