+++ autofit/text/formatter.py
@@ -109,22 +109,12 @@ def add_whitespace(str0, str1, whitespace):
     return f"{str0}{str1.rjust(whitespace - len(str0) + len(str1))}"
 
 
-def parameter_result_string_from(
-    parameter_name,
-    value,
-    values_at_sigma=None,
-    subscript=None,
-    unit=None,
-    format_string=None,
-    name_to_label=False,
+def value_result_string_from(
+    parameter_name, value, values_at_sigma=None, unit=None, format_string=None
 ):
 
     format_str = format_string or format_string_for_parameter_name(parameter_name)
     value = format_str.format(value)
-
-    parameter_tag = convert_name_to_label(
-        parameter_name=parameter_name, name_to_label=name_to_label
-    )
 
     if unit is not None:
         unit = f" {unit}"
@@ -132,15 +122,11 @@ def parameter_result_string_from(
         unit = ""
 
     if values_at_sigma is None:
-        parameter_result = f"{value}{unit}"
+        return f"{value}{unit}"
     else:
         lower_value_at_sigma = format_str.format(values_at_sigma[0])
         upper_value_at_sigma = format_str.format(values_at_sigma[1])
-        parameter_result = f"{value} ({lower_value_at_sigma}, {upper_value_at_sigma}){unit}"
-
-    if subscript is None:
-        return f"{parameter_tag} {parameter_result}"
-    return f"{parameter_tag}_{subscript} {parameter_result}"
+        return f"{value} ({lower_value_at_sigma}, {upper_value_at_sigma}){unit}"
 
 
 def parameter_result_latex_from(
@@ -176,8 +162,6 @@ def parameter_result_latex_from(
         lower_value_at_sigma = format_str.format(errors[0])
         upper_value_at_sigma = format_str.format(errors[1])
         return f"{str0}{subscript} = {value}^{{+{upper_value_at_sigma}}}_{{-{lower_value_at_sigma}}}{unit} & "
-
-
 
 
 def output_list_of_strings_to_file(file, list_of_strings):
