+++ autofit/text/samples_text.py
@@ -11,7 +11,10 @@ def values_from_samples(samples, median_pdf_model):
         return samples.median_pdf_vector
     return samples.max_log_likelihood_vector
 
-def summary(samples, sigma=3.0, median_pdf_model=True, name_to_label=True) -> str:
+
+def summary(
+    samples, sigma=3.0, median_pdf_model=True, indent=1, line_length=None
+) -> str:
     """ Create a string summarizing the results of the non-linear search at an input sigma value.
 
     This function is used for creating the model.results files of a non-linear search.
@@ -24,53 +27,24 @@ def summary(samples, sigma=3.0, median_pdf_model=True, name_to_label=True) -> st
     values = values_from_samples(samples=samples, median_pdf_model=median_pdf_model)
     values_at_sigma = samples.vector_at_sigma(sigma=sigma)
 
-    sigma_formatter = frm.TextFormatter()
+    parameter_names = samples.model.parameter_names
+
+    if line_length is None:
+        line_length = len(max(parameter_names, key=len)) + 8
+
+    sigma_formatter = frm.TextFormatter(indent=indent, line_length=line_length)
 
     for i, prior_path in enumerate(samples.model.unique_prior_paths):
 
-        value = frm.parameter_result_string_from(
-            parameter_name=samples.model.parameter_names[i],
+        value_result = frm.value_result_string_from(
+            parameter_name=parameter_names[i],
             value=values[i],
             values_at_sigma=values_at_sigma[i],
-            name_to_label=name_to_label,
         )
 
-        sigma_formatter.add((prior_path, value))
+        sigma_formatter.add((prior_path, value_result))
 
-    return "\n\nMedian PDF model Summary ({} sigma limits):\n\n{}".format(
-        sigma, sigma_formatter.text
-    )
-
-
-def table(samples, median_pdf_model=True, sigma=3.0, name_to_label=True) -> str:
-    """ Create a string summarizing the results of the non-linear search at an input sigma value.
-
-    This function is used for creating the model.results files of a non-linear search.
-
-    Parameters
-    ----------
-    sigma : float
-        The sigma within which the PDF is used to estimate errors (e.g. sigma = 1.0 uses 0.6826 of the PDF)."""
-
-    values = values_from_samples(samples=samples, median_pdf_model=median_pdf_model)
-    values_at_sigma = samples.vector_at_sigma(sigma=sigma)
-
-    table = []
-
-    for i in range(samples.model.prior_count):
-
-        label_value = frm.parameter_result_string_from(
-            parameter_name=samples.model.parameter_names[i],
-            value=values[i],
-            values_at_sigma=values_at_sigma[i],
-            subscript=samples.model.subscripts[i],
-            name_to_label=name_to_label,
-        )
-        table.append(f"{label_value} & ")
-
-    table = "".join(table)[:-3]
-
-    return "\n\nMedian PDF model Table ({} sigma limits):\n\n{}".format(sigma, table)
+    return "\n\nSummary ({} sigma limits):\n\n{}".format(sigma, sigma_formatter.text)
 
 
 def latex(samples, median_pdf_model=True, sigma=3.0, name_to_label=True) -> str:
@@ -100,10 +74,6 @@ def latex(samples, median_pdf_model=True, sigma=3.0, name_to_label=True) -> str:
 
         table.append(f"{label_value}")
 
-
     table = "".join(table)[:-3]
 
-    return "\n\nMedian PDF model Table ({} sigma limits):\n\n{}".format(sigma, table)
-
-
-
+    return "{}".format(table)
