### 2020-09-23 6ab42f9
separated out aggregator tests
