+++ test_autofit/unit/test_aggregator.py
@@ -1,368 +1,0 @@
-import os
-
-import pytest
-
-import autofit as af
-
-
-class MedianPDFInstance:
-    def __init__(self, name):
-        self.name = name
-
-
-class MockPhaseOutput:
-    def __init__(self, directory, pipeline, phase, dataset):
-        self.directory = directory
-        self.pipeline = pipeline
-        self.phase = phase
-        self.dataset = dataset
-
-    @property
-    def median_pdf_instance(self):
-        return MedianPDFInstance(
-            self.phase
-        )
-
-    @property
-    def output(self):
-        return self
-
-
-@pytest.fixture(name="aggregator")
-def make_aggregator():
-    aggregator = af.Aggregator("")
-    aggregator.phases = [
-        MockPhaseOutput(
-            "directory/number/one",
-            "pipeline1",
-            "phase1",
-            "dataset1"
-        ),
-        MockPhaseOutput(
-            "directory/number/two",
-            "pipeline1",
-            "phase2",
-            "dataset1"
-        ),
-        MockPhaseOutput(
-            "directory/letter/a",
-            "pipeline2",
-            "phase2",
-            "dataset2"
-        ),
-    ]
-    return aggregator
-
-
-@pytest.fixture(
-    name="aggregator_directory"
-)
-def make_aggregator_directory():
-    directory = os.path.dirname(
-        os.path.realpath(__file__)
-    )
-    return f"{directory}/tools/files/aggregator"
-
-
-@pytest.fixture(
-    name="path_aggregator"
-)
-def make_path_aggregator(
-        aggregator_directory
-):
-    aggregator = af.Aggregator(
-        aggregator_directory
-    )
-    yield aggregator
-    aggregator.remove_unzipped()
-
-
-def test_completed_aggregator(
-        aggregator_directory
-):
-    aggregator = af.Aggregator(
-        aggregator_directory,
-        completed_only=True
-    )
-    assert len(aggregator) == 1
-    assert "completed" in aggregator[0].directory
-
-
-class TestLoading:
-    def test_unzip(self, path_aggregator):
-        assert len(path_aggregator) == 2
-
-    def test_pickles(self, path_aggregator):
-        assert list(path_aggregator.values(
-            "dataset"
-        ))[0]["name"] == "dataset"
-        assert list(path_aggregator.values(
-            "model"
-        ))[0]["name"] == "model"
-        assert list(path_aggregator.values(
-            "non_linear"
-        ))[0]["name"] == "optimizer"
-        assert list(path_aggregator.values(
-            "nonsense"
-        ))[0] is None
-
-
-@pytest.fixture(
-    name="ages_for_predicate"
-)
-def make_ages_for_predicate(path_aggregator):
-    def ages_for_predicate(
-            predicate
-    ):
-        result = path_aggregator.filter(
-            predicate
-        )
-        return [child.age for child in result.values(
-            "child"
-        )]
-
-    return ages_for_predicate
-
-
-class TestNumericalFiltering:
-    def test_equality(self, path_aggregator):
-        predicate = path_aggregator.child.age == 17
-        result = path_aggregator.filter(
-            predicate
-        )
-        assert len(result) == 1
-        assert list(result.values("child"))[0].age == 17
-
-    def test_greater_than(
-            self,
-            ages_for_predicate,
-            path_aggregator
-    ):
-        ages = ages_for_predicate(
-            path_aggregator.child.age > 10
-        )
-        assert ages == [17]
-
-    def test_greater_than_equal(
-            self,
-            ages_for_predicate,
-            path_aggregator
-    ):
-        ages = ages_for_predicate(
-            path_aggregator.child.age >= 10
-        )
-        assert set(ages) == {10, 17}
-
-    def test_less_than(
-            self,
[…]
