### 2020-09-23 964fe3d
separated out aggregator tests
