### 2020-09-23 a643831
homogenise function
