+++ autofit/aggregator/aggregator.py
@@ -16,7 +16,7 @@ import os
 import zipfile
 from collections import defaultdict
 from shutil import rmtree
-from typing import List, Union, Iterator
+from typing import List, Union, Iterator, Tuple
 
 from .phase_output import PhaseOutput
 from .predicate import AttributePredicate
@@ -188,6 +188,45 @@ class AbstractAggregator:
             self.phases
         )
 
+    def homogenize(
+            self,
+            aggregator: "AbstractAggregator",
+            on: str
+    ) -> Tuple[
+        "AbstractAggregator",
+        "AbstractAggregator"
+    ]:
+        """
+        Filter this aggregator and another aggregator such that each only
+        contain results where at least one result in the other aggregator
+        has the same value for a given property.
+
+        Parameters
+        ----------
+        aggregator
+            Another aggregator to homogenize with this one.
+        on
+            The attribute of the underlying results which should match.
+
+        Returns
+        -------
+        A pair of aggregators with only matching results.
+        """
+        def _homogenize(a, b):
+            values = set(b.values(on))
+            return AbstractAggregator([
+                phase for phase in a.phases
+                if getattr(phase, on) in values
+            ])
+
+        return _homogenize(
+            self,
+            aggregator
+        ), _homogenize(
+            aggregator,
+            self
+        )
+
     def map(self, func):
         """
         Map some function onto the aggregated output objects.
@@ -279,7 +318,5 @@ class Aggregator(AbstractAggregator):
         if len(phases) == 0:
             print(f"\nNo phases found in {directory}\n")
         else:
-        #    paths_string = "\n".join(phase.directory for phase in phases)
-        #    print(f"\nPhases were found in these directories:\n\n{paths_string}\n")
             print(f"\n A total of {str(len(phases))} phases and results were found.")
         super().__init__(phases)
