### 2020-09-23 d5ce76f
homogenise function
