+++ test_autofit/mock.py
@@ -242,3 +242,26 @@ class MockComponents:
 
 class HyperGalaxy:
     pass
+
+
+class MedianPDFInstance:
+    def __init__(self, name):
+        self.name = name
+
+
+class MockPhaseOutput:
+    def __init__(self, directory, pipeline, phase, dataset):
+        self.directory = directory
+        self.pipeline = pipeline
+        self.phase = phase
+        self.dataset = dataset
+
+    @property
+    def median_pdf_instance(self):
+        return MedianPDFInstance(
+            self.phase
+        )
+
+    @property
+    def output(self):
+        return self
