### 2020-09-23 rhayes777/PyAutoFit#127 @Jammy2211
Agreed and updated README.
