### 2020-09-23 b4f735f
fixed test default and migrated first config parsing to new method
