### 2020-09-23 47b3f78
updating config references
