+++ autofit/non_linear/nest/multi_nest.py
@@ -112,22 +112,22 @@ class MultiNest(abstract_nest.AbstractNest):
         """
 
         self.n_live_points = (
-            self._config("search", "n_live_points", int)
+            self._config("search", "n_live_points")
             if n_live_points is None
             else n_live_points
         )
         self.sampling_efficiency = (
-            self._config("search", "sampling_efficiency", float)
+            self._config("search", "sampling_efficiency")
             if sampling_efficiency is None
             else sampling_efficiency
         )
         self.const_efficiency_mode = (
-            self._config("search", "const_efficiency_mode", bool)
+            self._config("search", "const_efficiency_mode")
             if const_efficiency_mode is None
             else const_efficiency_mode
         )
         self.evidence_tolerance = (
-            self._config("search", "evidence_tolerance", float)
+            self._config("search", "evidence_tolerance")
             if evidence_tolerance is None
             else evidence_tolerance
         )
@@ -136,53 +136,53 @@ class MultiNest(abstract_nest.AbstractNest):
             self.evidence_tolerance = 0.8
 
         self.multimodal = (
-            multimodal or self._config("search", "multimodal", bool)
+            multimodal or self._config("search", "multimodal")
             if multimodal is None
             else multimodal
         )
         self.importance_nested_sampling = (
-            self._config("search", "importance_nested_sampling", bool)
+            self._config("search", "importance_nested_sampling")
             if importance_nested_sampling is None
             else importance_nested_sampling
         )
         self.max_modes = (
-            self._config("search", "max_modes", int) if max_modes is None else max_modes
+            self._config("search", "max_modes") if max_modes is None else max_modes
         )
         self.mode_tolerance = (
-            self._config("search", "mode_tolerance", float)
+            self._config("search", "mode_tolerance")
             if mode_tolerance is None
             else mode_tolerance
         )
         self.max_iter = (
-            self._config("search", "max_iter", int) if max_iter is None else max_iter
+            self._config("search", "max_iter") if max_iter is None else max_iter
         )
         self.n_iter_before_update = (
-            self._config("settings", "n_iter_before_update", int)
+            self._config("settings", "n_iter_before_update")
             if n_iter_before_update is None
             else n_iter_before_update
         )
         self.null_log_evidence = (
-            self._config("settings", "null_log_evidence", float)
+            self._config("settings", "null_log_evidence")
             if null_log_evidence is None
             else null_log_evidence
         )
-        self.seed = self._config("settings", "seed", int) if seed is None else seed
+        self.seed = self._config("settings", "seed") if seed is None else seed
         self.verbose = (
-            self._config("settings", "verbose", bool) if verbose is None else verbose
+            self._config("settings", "verbose") if verbose is None else verbose
         )
         self.resume = (
-            self._config("settings", "resume", bool) if resume is None else resume
+            self._config("settings", "resume") if resume is None else resume
         )
         self.context = (
-            self._config("settings", "context", int) if context is None else context
+            self._config("settings", "context") if context is None else context
         )
         self.write_output = (
-            self._config("settings", "write_output", bool)
+            self._config("settings", "write_output")
             if write_output is None
             else write_output
         )
         self.log_zero = (
-            self._config("settings", "log_zero", float)
+            self._config("settings", "log_zero")
             if log_zero is None
             else log_zero
         )
