+++ autofit/non_linear/mock/mock_search.py
@@ -76,7 +76,7 @@ class MockSearch(NonLinearSearch):
 
     @property
     def config_type(self):
-        return conf.instance.mock
+        return conf.instance["non_linear"]["mock"]
 
     @property
     def tag(self):

[…]
