+++ autofit/tools/phase.py
@@ -1,16 +1,14 @@
 import logging
 import pickle
 from abc import ABC, abstractmethod
-from copy import deepcopy
 from typing import Dict
+
 import dill
 
 from autoconf import conf
 from autofit.mapper.model_mapper import ModelMapper
-from autofit.non_linear.paths import convert_paths
 from autofit.mapper.prior.promise import PromiseResult
 from autofit.non_linear import grid_search
-from autofit.non_linear.paths import Paths
 
 logger = logging.getLogger(__name__)
 
@@ -247,7 +245,6 @@ def as_grid_search(phase_class, parallel=False):
                 number_of_steps=4,
                 **kwargs,
         ):
-
             super().__init__(search=search, **kwargs)
 
             self.search = grid_search.GridSearch(
@@ -278,7 +275,6 @@ def as_grid_search(phase_class, parallel=False):
             )
 
         def run_analysis(self, analysis, **kwargs):
-
             self.search.search.paths = self.paths
             self.search.paths = self.paths
 
@@ -313,5 +309,5 @@ class AbstractSettingsPhase:
         """
         if self.log_likelihood_cap is None:
             return ""
-        return f"__{conf.instance.settings_tag.get('phase', 'log_likelihood_cap')}" \
-                + "_{0:.1f}".format(self.log_likelihood_cap)
+        return f"__{conf.instance['notation']['settings_tags']['phase']['log_likelihood_cap']}" \
+               + "_{0:.1f}".format(self.log_likelihood_cap)
