+++ autofit/text/formatter.py
@@ -68,7 +68,7 @@ def format_string_for_parameter_name(parameter_name: str) -> str:
     format
         The format string (e.g {:.2f})
     """
-    label_conf = conf.instance.label_format
+    label_conf = conf.instance["notation"]["label_format"]
 
     try:
         # noinspection PyProtectedMember
@@ -82,7 +82,7 @@ def format_string_for_parameter_name(parameter_name: str) -> str:
     except KeyError:
         pass
     raise configparser.NoSectionError(
-        "Could not find an entry for the parameter {} in the label_format.iniconfig at path {}".format(
+        "Could not find an entry for the parameter {} in the label_format.ini config at path {}".format(
             parameter_name, label_conf.path
         )
     )
