+++ test_autofit/unit/tools/test_phase.py
@@ -1,7 +1,7 @@
 import autofit as af
 
-def test__phase_settings_tag():
 
+def test__phase_settings_tag():
     settings = af.AbstractSettingsPhase(log_likelihood_cap=None)
     assert settings.log_likelihood_cap_tag == ""
 
