+++ paper/paper.md
@@ -31,15 +31,15 @@ Probabilistic programming languages (PPLs) have enabled contemporary statistical
 
 # Software Description
 
-`PyAutoFit` is a PPL whose core design is providing a direct interface with the model, data, fitting procedure and results, allowing it to provide more complete management of the model-fitting task than other PPLs and making it suited to longer term software projects. Model components are written as Python classes, allowing `PyAutoFit` to define the model and associated parameters in an expressive way that is tied to the modeling software's API. A model fit then only requires that a `PyAutoFit` _Analysis_ class is writen, which combines the data, model and likelihood function and defines how the model-fit is performed using a non-linear search (e.g. `dynesty` [@dynesty], `emcee` [@emcee] or `PySwarms` [@pyswarms]). 
+`PyAutoFit` is a PPL whose core design is providing a direct interface with the model, data, fitting procedure and results, allowing it to provide more complete management of the model-fitting task than other PPLs and making it suited to longer term software projects. Model components are written as Python classes, allowing `PyAutoFit` to define the model and associated parameters in an expressive way that is tied to the modeling software's API. A model fit then only requires that a `PyAutoFit` `Analysis` class is writen, which combines the data, model and likelihood function and defines how the model-fit is performed using a non-linear search (e.g. `dynesty` [@dynesty], `emcee` [@emcee] or `PySwarms` [@pyswarms]). 
 
-The _Analysis_ class provides a model specific interface between `PyAutoFit` and the modeling software, allowing it to handle the 'heavy lifting' that comes with writing model-fitting software. This includes interfacing with the non-linear search, outputting results in a structured path format and model-specific visualization during and after the non-linear search. Results are output in a database structure with metadata that allows the _Aggregator_ tool to load results post-analysis via a Python script or Jupyter notebook. This includes methods for summarizing the results of every fit, filtering results to inspect subsets of model fits and visualizing results. Results are loaded as `Python` generators, ensuring the _Aggregator_ can be used to interpret large result datasets in a memory efficient way. `PyAutoFit` is therefore suited to 'big data' problems where independent fits to large homogeneous data-sets using an identical model-fitting procedure are performed. 
+The `Analysis` class provides a model specific interface between `PyAutoFit` and the modeling software, allowing it to handle the 'heavy lifting' that comes with writing model-fitting software. This includes interfacing with the non-linear search, outputting results in a structured path format and model-specific visualization during and after the non-linear search. Results are output in a database structure with metadata that allows the `Aggregator` tool to load results post-analysis via a Python script or Jupyter notebook. This includes methods for summarizing the results of every fit, filtering results to inspect subsets of model fits and visualizing results. Results are loaded as `Python` generators, ensuring the `Aggregator` can be used to interpret large result datasets in a memory efficient way. `PyAutoFit` is therefore suited to 'big data' problems where independent fits to large homogeneous data-sets using an identical model-fitting procedure are performed. 
 
 # Model Abstraction and Composition
 
 For many modeling problems the model comprises abstract model components representing objects or processes in a physical system. For example, galaxy morphology studies in astrophysics where model components represent the light profile of stars [@Haussler2013] [@Nightingale2019]. For these problems the likelihood function is typically a sequence of numerical processes (e.g. convolutions, Fourier transforms, linear algebra) and extensions to the model often requires the addition of new model components in a way that is non-trivially included in the fitting process and likelihood function. Existing PPLs have tools for these problems, for example `black-box' likelihood functions in PyMC3. However, these solutions decouple model composition from the data and fitting procedure, making the model less expressive, restricting model customization and reducing flexibility in how the model-fit is performed.
 
[…]
