+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/src/phase/analysis.py
@@ -15,7 +15,7 @@ class Analysis(af.Analysis):
 
         super().__init__()
 
-        """The _MaskedDataset_ and visualizer are created in the same way as tutorial 6."""
+        """The `MaskedDataset` and visualizer are created in the same way as tutorial 6."""
 
         self.masked_dataset = masked_dataset
 
