+++ howtofit/chapter_1_introduction/tutorial_6_complex_models/tutorial_6_complex_models.py
@@ -3,15 +3,15 @@
 Tutorial 6: Complex Models
 ==========================
 
-Up to now, we've fitted a very simple model - a single 1D _Gaussian_ with just 3 free parameters. In this tutorial,
+Up to now, we've fitted a very simple model - a single 1D `Gaussian` with just 3 free parameters. In this tutorial,
 we'll look at how PyAutoFit allows us to compose and fit models of arbitrary complexity.
 
 To begin, you should check out the module 'tutorial_6_complex_models/model/profiles.py'. In previous tutorials this
-module was called 'gaussian.py' and it contained only the _Gaussian_ class we fitted to data. The module now includes
-a second profile, 'Exponential', which like the _Gaussian_ class is a model-component that can be fitted to data.
+module was called 'gaussian.py' and it contained only the `Gaussian` class we fitted to data. The module now includes
+a second profile, 'Exponential', which like the `Gaussian` class is a model-component that can be fitted to data.
 
-Up to now, our data has always been generated using a single _Gaussian_ profile. Thus, we have only needed to fit
-it with a single Gaussian. In this tutorial, our _Dataset_ is now a superpositions of multiple profiles (e.g a
+Up to now, our data has always been generated using a single `Gaussian` profile. Thus, we have only needed to fit
+it with a single Gaussian. In this tutorial, our `Dataset` is now a superpositions of multiple profiles (e.g a
 Gaussians + Exponential). The models we compose and fit are therefore composed of multiple profiles, such that when we
 generate the model-data we generate it as the sum of all individual profiles in our model.
 """
@@ -45,14 +45,14 @@ conf.instance = conf.Config(
 
 # %%
 """
-Lets quickly recap tutorial 1, where using PriorModels we created a _Gaussian_ as a model component and used it to map a
+Lets quickly recap tutorial 1, where using PriorModels we created a `Gaussian` as a model component and used it to map a
 list of parameters to a model instance.
 """
 
 # %%
 model = af.PriorModel(htf.profiles.Gaussian)
 
-print("PriorModel _Gaussian_ object: \n")
+print("PriorModel `Gaussian` object: \n")
 print(model)
 
 instance = model.instance_from_vector(vector=[0.1, 0.2, 0.3])
@@ -78,8 +78,8 @@ model = af.CollectionPriorModel(
 
 # %%
 """
-A CollectionPriorModel behaves like a _PriorModel_ but contains a collection of model components. For example, it can
-create a model instance by mapping a list of parameters, which in this case is 6 (3 for the _Gaussian_ [centre,
+A CollectionPriorModel behaves like a `PriorModel` but contains a collection of model components. For example, it can
+create a model instance by mapping a list of parameters, which in this case is 6 (3 for the `Gaussian` [centre,
 intensity, sigma] and 3 for the Exponential [centre, intensity, rate]).
 """
 
@@ -135,7 +135,7 @@ Checkout 'phase.py' and 'analysis.py' now, for a description of how this has bee
 
 # %%
 """
-Import the simulator module and set up the _Dataset_. This uses a new _Dataset_ that is generated as a sum of a _Gaussian_ 
+Import the simulator module and set up the _Dataset_. This uses a new `Dataset` that is generated as a sum of a `Gaussian` 
 and Exponential profile.
 """
 
@@ -218,11 +218,11 @@ print("Emcee has finished run - you may now continue the notebook.")
 
 # %%
 """
-We can fully custommize the model that we fit. Lets suppose we have a _Dataset_ that consists of three _Gaussian_ line
[…]

[…]
