+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/tutorial_7_phase_customization.py
@@ -3,19 +3,19 @@
 Tutorial 7: Phase Customization
 ===============================
 
-In this tutorial, we're going to add input parameters to a Phase object that customizes the analysis. We'll use the
-specific example of two input parameters that trim our data-set from the left and right before fitting it. This
+In this tutorial, we`re going to add input parameters to a `Phase` object that customizes the analysis. we'll use the
+specific example of two input parameters that trim our `Dataset` from the left and right before fitting it. This
 example is somewhat trivial (we could achieve almost the same effect with masking), but it will serve to illustrate
-phase customization.
+`Phase` customization.
 
-When we customize a phase, we'll use these 'phase settings' to perform phase tagging. Here, we 'tag' the output
-path of the phase's results, such that every time a phase is run with a different customization a new set of
-unique results are stored for those settings. For a given data-set we are thus able to fit it multiple times using
-different settings to compare the results.
+When we customize a `Phase`, we'll use `SettingsPhase` to perform `tagging`. Here, we `tag` the output
+path of the `Phase``s `Result``., such that every time a `Phase` is run with a different customization a new set of
+unique `Result``s are stored for those `PhaseSetting``.. For a given `Dataset` we are thus able to fit it multiple
+times using different settings to compare the results.
 
-These new features have led to an additional module in the 'phase' package called 'settings.py', as well as extensions
-to the 'dataset.py' module. Before looking at these modules, lets first perform a series of Emcee fits to see how
-they change the behaviour of PyAutoFit.
+These new features have led to an additional module in the `phase` package called `settings.py`, as well as extensions
+to the `dataset.py` module. Before looking at these modules, lets first perform a series of `Emcee` fits to see how
+they change the behaviour of **PyAutoFit**.
 """
 
 # %%
@@ -36,7 +36,7 @@ print("Workspace Path: ", workspace_path)
 
 # %%
 """
-Setup the configs as we did in the previous tutorial, as well as the output folder for our non-linear search.
+Setup the configs as we did in the previous tutorial, as well as the output folder for our `NonLinearSearch`.
 """
 
 # %%
@@ -47,26 +47,26 @@ conf.instance = conf.Config(
 
 # %%
 """
-We're now going to perform multiple fits, where each fit trims the data-set that is fitted.
+We`re now going to perform multiple fits, where each fit trims the `Dataset` that is fitted.
 
-To do this, we'll set up phases with a new class called `SettingsMaskedDataset`, which contains the settings that 
-customize how a `MaskedDataset` is created. This has two inputs, 'data_trim_left' and 'data_trim_right':
+To do this, we'll set up `Phase`_s using a new class called `SettingsMaskedDataset`, which contains the settings that 
+customize how a `MaskedDataset` is created. This has two inputs, `data_trim_left` and `data_trim_right`:
 
-- data_trim_left:
+- `data_trim_left`:
 
-  The dataset's image and noise-map are trimmed and removed from the left (e.g. 1d index values from 0).
+  The `Dataset``s `image` and `noise_map` are trimmed and removed from the left (e.g. 1d index values from 0).
   
-  For example, if the `Dataset` has shape (100,) and we set data_trim_left=10, the `MaskedDataset` that is fitted will 
-  have shape (90,). The mask is trimmed in the same way.
+  For example, if the `Dataset` has shape (100,) and we set `data_trim_left=10`, the `MaskedDataset` that is fitted 
+  will have shape (90,). The `mask` is trimmed in the same way.
 
-- data_trim_right:
+- `data_trim_right`:
 
-  This behaves the same as data_trim_left, however data is removed from the right (e.g. 1D index values from the
+  This behaves the same as `data_trim_left`, however `data` is removed from the right (e.g. 1D index values from the
   shape[0] value of the 1D data).
 
[…]
