+++ howtofit/chapter_1_introduction/tutorial_2_model_fitting/tutorial_2_model_fitting.py
@@ -28,7 +28,7 @@ conf.instance = conf.Config(config_path=f"{workspace_path}/config")
 
 # %%
 """
-To create the Dataset, we import the simulator module and use it to generate the Dataset's data and noise-map. 
+To create the Dataset, we import the simulator module and use it to generate the Dataset`s data and noise-map. 
 """
 
 # %%
@@ -48,7 +48,7 @@ print(xvalues)
 
 # %%
 """
-Lets plot the data and noise-map we're going to fit.
+Lets plot the data and noise-map we`re going to fit.
 """
 
 # %%
@@ -68,15 +68,15 @@ plt.clf()
 """
 So, how do we actually go about fitting our `Gaussian` model to this data? First, we need to be able to generate
 an image of our 1D `Gaussian` model. As we did in tutorial 1, we define the `Gaussian` as a Python class with the format 
-required for PyAutoFit to use it as a model-component.
+required for **PyAutoFit** to use it as a model-component.
 """
 
 # %%
 class Gaussian:
     def __init__(
         self,
-        centre=0.0,  # <- PyAutoFit recognises these constructor arguments
-        intensity=0.1,  # <- are the Gaussian's model parameters.
+        centre=0.0,  # <- **PyAutoFit** recognises these constructor arguments
+        intensity=0.1,  # <- are the Gaussian`s model parameters.
         sigma=0.01,
     ):
         self.centre = centre
@@ -103,11 +103,11 @@ class Gaussian:
 
 # %%
 """
-We've extended the `Gaussian` class to have a method "profile_from_xvalues". Given an input set of x coordinates
-this computes the intensity of the `Gaussian` at every point. We've already seen how the data contains the xvalues we 
+We`ve extended the `Gaussian` class to have a method `profile_from_xvalues`. Given an input set of x coordinates
+this computes the intensity of the `Gaussian` at every point. We`ve already seen how the data contains the xvalues we 
 use, which are a 1D NumPy array spanning values 0 to 100.
 
-If we pass these values to an instance of the `Gaussian` class, we can create a line of the gaussian's values.
+If we pass these values to an instance of the `Gaussian` class, we can create a line of the gaussian`s values.
 """
 
 # %%
@@ -128,18 +128,18 @@ plt.clf()
 
 # %%
 """
-Different values of centre, intensity and sigma change the Gaussian's apperance - have a go at editing some of the
+Different values of centre, intensity and sigma change the Gaussian`s apperance - have a go at editing some of the
 values above to see this behaviour.
 """
 
 # %%
 """
-Okay, so lets recap. We've defined a model which is a 1D `Gaussian` and given a set of parameters for that model
-(x, I, sigma) we can create 'model_data' of the Gaussian. And, we have some data of a 1D `Gaussian` we want to
-fit this model with, so as to determine the values of (x, I, sigma) from which it was created. So how do we do that?
+Okay, so lets recap. We`ve defined a model which is a 1D `Gaussian` and given a set of parameters for that model
+$(x, I, \sigma)$ we can create `model_data` of the `Gaussian`. And, we have some data of a 1D `Gaussian` we want to
+fit this model with, so as to determine the values of $(x, I, \sigma)$ from which it was created. So how do we do that?
 
 Simple, we take the image from our data and our model_image of the `Gaussian` and subtract the two to get a
-'residual-map'.
+`residual-map`.
 """
 
 # %%
@@ -152,13 +152,13 @@ plt.clf()
 
 # %%
 """
-Clearly, this model isn't a good fit to the data - which was to be expected as they looked nothing alike!
+Clearly, this model isn`t a good fit to the data - which was to be expected as they looked nothing alike!
 
[…]
