+++ howtofit/chapter_1_introduction/tutorial_5_visualization_masking/src/model/gaussian.py
@@ -5,11 +5,11 @@ class Gaussian:
     def __init__(
         self,
         centre=0.0,  # <- PyAutoFit recognises these constructor arguments
-        intensity=0.1,  # <- are the Gaussian's model parameters.
+        intensity=0.1,  # <- are the Gaussian`s model parameters.
         sigma=0.01,
     ):
         """Represents a 1D `Gaussian` profile, which may be treated as a model-component of PyAutoFit the
-        parameters of which are fitted for by a non-linear search.
+        parameters of which are fitted for by a `NonLinearSearch`.
 
         Parameters
         ----------
