+++ autofit/__init__.py
@@ -71,4 +71,4 @@ from .tools.phase_property import PhaseProperty
 from .tools.pipeline import Pipeline
 from .tools.pipeline import ResultsCollection
 
-__version__ = '0.67.6'
+__version__ = '0.67.7'
