### 2020-09-24 b1f8a23
pool ids fixed for DynestyStatic
