+++ docs/general/workspace.rst
@@ -26,7 +26,7 @@ section of the ``readthedocs`` for a complete description of every configuration
 Examples
 --------
 
-Example scripts using the 1D data fitting model are provided here, including scripts for creating an ``Analysis`` class,
+Example scripts using the 1D ``data`` fitting model are provided here, including scripts for creating an ``Analysis`` class,
 performing a model-fit, inspecting results and using the ``Aggregator``. Two examples, illustrating ``simple`` and
 ``complex`` model-fits are provided.
 
@@ -43,7 +43,7 @@ full description of the lectures and online examples of every notebook.
 Dataset
 -------
 
-The folder where data for your model-fitting problem is stored. Example data for the 1D data fitting problem
+The folder where ``data`` for your model-fitting problem is stored. Example ``data`` for the 1D ``data`` fitting problem
 are provided in the workspace.
 
 Output

[…]
