+++ autofit/tools/phase.py
@@ -4,7 +4,6 @@ from abc import ABC, abstractmethod
 from copy import deepcopy
 from typing import Dict
 import dill
-import os
 
 from autoconf import conf
 from autofit.mapper.model_mapper import ModelMapper
@@ -78,10 +77,7 @@ class AbstractPhase:
         Save metadata associated with the phase, such as the name of the pipeline, the
         name of the phase and the name of the dataset being fit
         """
-        file_path = "{}/metadata".format(self.paths.make_path())
-        if os.path.exists(file_path):
-            os.remove(file_path)
-        with open(file_path, "w+") as f:
+        with open("{}/metadata".format(self.paths.make_path()), "w+") as f:
             f.write(
                 self.make_metadata_text(
                     dataset.name
@@ -92,28 +88,19 @@ class AbstractPhase:
         """
         Save the dataset associated with the phase
         """
-        file_path = f"{self.paths.pickle_path}/dataset.pickle"
-        if os.path.exists(file_path):
-            os.remove(file_path)
-        with open(file_path, "wb") as f:
+        with open(f"{self.paths.pickle_path}/dataset.pickle", "wb") as f:
             pickle.dump(dataset, f)
 
     def save_mask(self, mask):
         """
         Save the mask associated with the phase
         """
-        file_path = f"{self.paths.pickle_path}/mask.pickle"
-        if os.path.exists(file_path):
-            os.remove(file_path)
-        with open(file_path, "wb") as f:
+        with open(f"{self.paths.pickle_path}/mask.pickle", "wb") as f:
             dill.dump(mask, f)
 
     def save_settings(self, settings):
-        file_path = f"{self.paths.pickle_path}/settings.pickle"
-        if os.path.exists(file_path):
-            os.remove(file_path)
         with open(
-                file_path,
+                f"{self.paths.pickle_path}/settings.pickle",
                 "wb+"
         ) as f:
             pickle.dump(
@@ -121,11 +108,8 @@ class AbstractPhase:
             )
 
     def save_phase_attributes(self, phase_attributes):
-        file_path = f"{self.paths.pickle_path}/phase_attributes.pickle"
-        if os.path.exists(file_path):
-            os.remove(file_path)
         with open(
-                file_path,
+                f"{self.paths.pickle_path}/phase_attributes.pickle",
                 "wb+"
         ) as f:
             pickle.dump(
@@ -148,6 +132,7 @@ class AbstractPhase:
         return PromiseResult(self)
 
     def run_analysis(self, analysis, info=None, pickle_files=None):
+
         return self.search.fit(model=self.model, analysis=analysis, info=info, pickle_files=pickle_files)
 
     def make_phase_attributes(self, analysis):
