+++ howtofit/chapter_1_introduction/tutorial_5_visualization_masking/src/plot/line_plots.py
@@ -9,7 +9,10 @@ plot function below has been extended to provide this functionality.
 def line(
     xvalues,
     line,
+    title=None,
     ylabel=None,
+    errors=None,
+    color="k",
     output_path=None,
     output_filename=None,
     output_format="show",
@@ -35,8 +38,11 @@ def line(
         Determines where the plot is displayed on your screen ("show") or output to the hard-disk as a png ("png").
     """
 
-    plt.plot(xvalues, line)
-    plt.xlabel("x")
+    plt.errorbar(
+        x=xvalues, y=line, yerr=errors, color=color, ecolor="k", elinewidth=1, capsize=2
+    )
+    plt.title(title)
+    plt.xlabel("x value of profile")
     plt.ylabel(ylabel)
     if "show" in output_format:
         plt.show()
