+++ howtofit/chapter_1_introduction/tutorial_4_source_code/src/plot/fit_plots.py
@@ -19,7 +19,13 @@ def data(fit):
     Fit : fit.Fit
         The observed Fit `Dataset` whose data is plotted.
     """
-    line_plots.line(xvalues=fit.xvalues, line=fit.data, ylabel="Data Values")
+    line_plots.line(
+        xvalues=fit.xvalues,
+        line=fit.data,
+        title="Data",
+        color="k",
+        ylabel="Data Values",
+    )
 
 
 def noise_map(fit):
@@ -30,7 +36,13 @@ def noise_map(fit):
     Fit : fit.Fit
         The observed Fit whose noise-map is plotted.
     """
-    line_plots.line(xvalues=fit.xvalues, line=fit.noise_map, ylabel="Noise Map")
+    line_plots.line(
+        xvalues=fit.xvalues,
+        line=fit.noise_map,
+        title="Noise-Map",
+        color="k",
+        ylabel="Noise Map",
+    )
 
 
 def signal_to_noise_map(fit):
@@ -42,7 +54,11 @@ def signal_to_noise_map(fit):
         The observed Fit whose signal-to-noise-map is plotted.
     """
     line_plots.line(
-        xvalues=fit.xvalues, line=fit.signal_to_noise_map, ylabel="Signal-To-Noise Map"
+        xvalues=fit.xvalues,
+        line=fit.signal_to_noise_map,
+        title="Signal-To-Noise Map",
+        color="k",
+        ylabel="Signal-To-Noise Map",
     )
 
 
@@ -54,7 +70,13 @@ def model_data(fit):
     Fit : fit.Fit
         The Fit model data.
     """
-    line_plots.line(xvalues=fit.xvalues, line=fit.model_data, ylabel="Model Data")
+    line_plots.line(
+        xvalues=fit.xvalues,
+        line=fit.model_data,
+        title="Model Data",
+        color="r",
+        ylabel="Model Data",
+    )
 
 
 def residual_map(fit):
@@ -65,7 +87,13 @@ def residual_map(fit):
     Fit : fit.Fit
         The observed Fit whose residual-map is plotted.
     """
-    line_plots.line(xvalues=fit.xvalues, line=fit.residual_map, ylabel="Residual Map")
+    line_plots.line(
+        xvalues=fit.xvalues,
+        line=fit.residual_map,
+        title="Residual-Map",
+        color="r",
+        ylabel="Residual Map",
+    )
 
 
 def normalized_residual_map(fit):
@@ -79,6 +107,8 @@ def normalized_residual_map(fit):
     line_plots.line(
         xvalues=fit.xvalues,
         line=fit.normalized_residual_map,
+        title="Normalized Residual Map",
+        color="r",
         ylabel="Normalized Residual Map",
     )
 
@@ -92,5 +122,9 @@ def chi_squared_map(fit):
         The observed Fit whose chi-squared-map is plotted.
     """
     line_plots.line(
-        xvalues=fit.xvalues, line=fit.chi_squared_map, ylabel="Chi-Squared Map"
+        xvalues=fit.xvalues,
+        line=fit.chi_squared_map,
+        title="Chi-Squared Map",
+        color="r",
+        ylabel="Chi-Squared Map",
     )
