+++ howtofit/chapter_2_results/scripts/tutorial_4_filtering.py
@@ -47,7 +47,7 @@ are in the middle of a `NonLinearSearch` will be omitted and not loaded in the `
 """
 
 # %%
-output_path = f"{workspace_path}/output"
+output_path = f"{workspace_path}/howtofit/output/chapter_2/aggregator"
 
 agg = af.Aggregator(directory=str(output_path), completed_only=True)
 
