+++ howtofit/simulators/chapter_1/profiles.py
@@ -78,7 +78,7 @@ class Gaussian(Profile):
 
         Parameters
         ----------
-        values : ndarray
+        values : np.ndarray
             The x coordinates in the original reference frame of the grid.
         """
         transformed_xvalues = np.subtract(xvalues, self.centre)
@@ -120,7 +120,7 @@ class Exponential(Profile):
 
         Parameters
         ----------
-        values : ndarray
+        values : np.ndarray
             The x coordinates in the original reference frame of the grid.
         """
         transformed_xvalues = np.subtract(xvalues, self.centre)
