+++ howtofit/chapter_1_introduction/tutorial_5_visualization_masking/src/phase/analysis.py
@@ -36,7 +36,8 @@ class Analysis(af.Analysis):
         )
 
     def log_likelihood_function(self, instance):
-        """Determine the log likelihood of a fit of a `Gaussian` to the dataset, using the model instance of a Gaussian.
+        """
+    Returns the log likelihood of a fit of a `Gaussian` to the dataset, using the model instance of a Gaussian.
 
         Parameters
         ----------
