+++ howtofit/chapter_1_introduction/tutorial_5_visualization_masking/src/plot/line_plots.py
@@ -24,9 +24,9 @@ def line(
 
     Parameters
     ----------
-    xvalues : ndarray
+    xvalues : np.ndarray
         The x-coordinates the profile is defined on.
-    line : ndarray
+    line : np.ndarray
         The intensity values of the profile which are plotted.
     ylabel : str
         The y-label of the plot.
