+++ howtofit/chapter_1_introduction/tutorial_2_model_fitting/tutorial_2_model_fitting.py
@@ -88,7 +88,7 @@ class Gaussian:
 
         Parameters
         ----------
-        xvalues : ndarray
+        xvalues : np.ndarray
             The x coordinates in the original reference frame of the data.
         """
         transformed_xvalues = np.subtract(xvalues, self.centre)
