+++ howtofit/chapter_1_introduction/tutorial_1_model_mapping/tutorial_1_model_mapping.py
@@ -160,7 +160,7 @@ class Gaussian:
 
         Parameters
         ----------
-        xvalues : ndarray
+        xvalues : np.ndarray
             The x coordinates in the original reference frame of the data.
         """
         transformed_xvalues = np.subtract(xvalues, self.centre)
