+++ paper/paper.md
@@ -23,33 +23,104 @@ bibliography: paper.bib
 
 # Summary
 
-A major trend in academia and data science is the rapid adoption of Bayesian statistics for data analysis and modelling, leading to the development of probabilistic programming languages (PPL). A PPL provides a framework that allows users to easily specify a probabilistic model and perform inference automatically. `PyAutoFit` is a Python-based PPL aimed at model fitting problems tackled by long-term software development projects. By interfacing with all aspects of the modelling (e.g. the model, data, fitting procedure, visulization, results, etc.), `PyAutoFit` provides a more complete management of modeling than other PPLs. This includes composing high-dimensionality models from individual model components, customizing the fitting procedure and performing data augmentation before a model-fit. Advanced features include database tools for analysing large suites of modeling results and exploiting domain-specific knowledge of a problem via transdimensional model-fitting pipelines. Accompanying `PyAutoFit` is the [autofit workspace](https://github.com/Jammy2211/autofit_workspace), which includes example scripts and the `HowToFit` lecture series which introduces non-experts to model-fitting and provides a guide on how to write a software project using `PyAutoFit`. To get started readers should go to our [readthedocs](https://pyautofit.readthedocs.io/en/latest/) and contact us to join the [PyAutoFit Slack channel](https://pyautofit.slack.com/) where we are building our online community.
+A major trend in academia and data science is the rapid adoption of Bayesian statistics for data analysis and modelling, 
+leading to the development of probabilistic programming languages (PPL). A PPL provides a framework that allows users 
+to easily specify a probabilistic model and perform inference automatically. `PyAutoFit` is a Python-based PPL aimed 
+at model fitting problems tackled by long-term software development projects. By interfacing with all aspects of the 
+modelling (e.g. the model, data, fitting procedure, visulization, results, etc.), `PyAutoFit` provides a more complete 
+management of modeling than other PPLs. This includes composing high-dimensionality models from individual model 
+components, customizing the fitting procedure and performing data augmentation before a model-fit. Advanced features 
+include database tools for analysing large suites of modeling results and exploiting domain-specific knowledge of a 
+problem via transdimensional model-fitting pipelines. Accompanying `PyAutoFit` is 
+the [autofit workspace](https://github.com/Jammy2211/autofit_workspace), which includes example scripts and 
+the `HowToFit` lecture series which introduces non-experts to model-fitting and provides a guide on how to write a 
+software project using `PyAutoFit`. To get started readers should go to 
+our [readthedocs](https://pyautofit.readthedocs.io/en/latest/) and contact us to join 
+the [PyAutoFit Slack channel](https://pyautofit.slack.com/) where we are building our online community.
 
 # Background of Probabilistic Programming
 
-Probabilistic programming languages (PPLs) have enabled contemporary statistical inference techniques to be applied to a diverse range of problems across academia and industry. Packages such as PyMC3 [@Salvatier2016], Pyro [@Bingham2019] and STAN [@Carpenter2017] offer general-purpose frameworks where users can specify a generative model and fit it to data using a variety of non-linear fitting techniques. Each package is specialized to problems of a certain nature, with many focused on problems like generalized linear modeling or determining the distribution(s) from which the data was drawn. For these problems the model is typically composed of linear equations which are easily expressed syntactically, such that the PPL API offers an expressive way to define the model and extensions can be implemented in an intuitive and straightforward way.
+Probabilistic programming languages (PPLs) have enabled contemporary statistical inference techniques to be applied 
+to a diverse range of problems across academia and industry. Packages such as PyMC3 [@Salvatier2016], 
+Pyro [@Bingham2019] and STAN [@Carpenter2017] offer general-purpose frameworks where users can specify a generative 
[…]
