+++ paper/paper.bib
@@ -276,3 +276,15 @@ year = {2009}
  publisher = {CreateSpace},
  address = {Scotts Valley, CA}
 }
+@article{uravu,
+  doi = {10.21105/joss.02214},
+  url = {https://doi.org/10.21105/joss.02214},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {50},
+  pages = {2214},
+  author = {Andrew R. McCluskey and Tim Snow},
+  title = {uravu: Making Bayesian modelling easy(er)},
+  journal = {Journal of Open Source Software}
+}
