### 2020-09-26 rhayes777/PyAutoFit#126 @Jammy2211
Good shout, I've added instructions for installing from source (which include how to run tests) and a contibuting section to the github.

https://pyautofit.readthedocs.io/en/latest/general/installation.html#cloning-forking
https://github.com/rhayes777/PyAutoFit/blob/master/CONTRIBUTING.md
