+++ docs/index.rst
@@ -1,4 +1,4 @@
-Probibalistic Programming
+Probabilistic Programming
 =========================
 
 Probabilistic programming languages (PPLs) have enabled contemporary statistical inference techniques to be applied
@@ -17,21 +17,99 @@ Why PyAutoFit?
 **PyAutoFit** is a PPL whose core design is providing a direct interface with the model, data, fitting procedure and 
 results, allowing it to provide more complete management of the model-fitting task than other PPLs and making it 
 suited to longer term software projects. Model components are written as Python classes, allowing **PyAutoFit** to 
-define the model and associated parameters in an expressive way that is tied to the modeling software's API. A 
-model fit then only requires that a **PyAutoFit** ``Analysis`` class is writen, which combines the data, model and 
+define the model and associated parameters in an expressive way that is tied to the modeling software's API. Here is
+a simple example of how a model representing a 1D Gaussian is written:
+
+.. code-block:: python
+
+    class Gaussian:
+
+        def __init__(
+            self,
+            centre=0.0,     # <- PyAutoFit recognises these
+            intensity=0.1,  # <- constructor arguments are
+            sigma=0.01,     # <- the Gaussian's parameters.
+        ):
+            self.centre = centre
+            self.intensity = intensity
+            self.sigma = sigma
+
+    """
+    An instance of the Gaussian class will be available during model fitting.
+
+    This method will be used to fit the model to ``data`` and compute a likelihood.
+    """
+
+    def profile_from_xvalues(self, xvalues):
+
+        transformed_xvalues = xvalues - self.centre
+
+        return (self.intensity / (self.sigma * (2.0 * np.pi) ** 0.5)) * \
+                np.exp(-0.5 * transformed_xvalues / self.sigma)
+
+A model fit then only requires that a **PyAutoFit** ``Analysis`` class is writen, which combines the data, model and
 likelihood function and defines how the model-fit is performed using a non-linear search 
 (e.g. `dynesty <https://github.com/joshspeagle/dynesty>`_, `emcee <https://github.com/dfm/emcee>`_
-or `PySwarms <https://pyswarms.readthedocs.io/en/latest/>`_).
+or `PySwarms <https://pyswarms.readthedocs.io/en/latest/>`_). Lets take a look at an example ``Analysis`` class:
 
-The ``Analysis`` class provides a model specific interface between **PyAutoFit** and the modeling software, allowing it 
-to handle the 'heavy lifting' that comes with writing model-fitting software. This includes interfacing with the 
+.. code-block:: python
+
+    class Analysis(af.Analysis):
+
+        def __init__(self, data, noise_map):
+
+            self.data = data
+            self.noise_map = noise_map
+
+        def log_likelihood_function(self, instance):
+
+            """
+            The 'instance' that comes into this method is an instance of the Gaussian class
+            above, with the parameters set to values chosen by the non-linear search.
+            """
+
+            print("Gaussian Instance:")
+            print("Centre = ", instance.centre)
+            print("Intensity = ", instance.intensity)
+            print("Sigma = ", instance.sigma)
+
+            """
+            We fit the ``data`` with the Gaussian instance, using its
+            "profile_from_xvalues" function to create the model data.
+            """
+
+            xvalues = np.arange(self.data.shape[0])
+
+            model_data = instance.profile_from_xvalues(xvalues=xvalues)
+            residual_map = self.data - model_data
+            chi_squared_map = (residual_map / self.noise_map) ** 2.0
+            log_likelihood = -0.5 * sum(chi_squared_map)
+
+            return log_likelihood
+
+The ``Analysis`` class provides a model specific interface between **PyAutoFit** and the modeling software, allowing
[…]
