### 2020-09-26 #2550 @Jammy2211
I have responded to all suggestions :). Pretty much every single one improved the quality of the software and docs significantly, so thank you!

I have added uravu to the software list at the bottom, but not in the main body of text. uravu looks like a really cool project and shares a lot of similarity with PyAutoFit given how we both focus on emcee and dynesty. However, it doesn't take me as a probabilistic programming language (i.e. no model composition aspect), which is what the text's discussion is focused on (and is our main focus with PyAutoFit).

I think I have updated the references on the paper (I updated the .bib / .md files as necessary) but haven't been able to check to be 100% sure so let me know if there are still problems.
