### 2020-09-26 #2550 @Jammy2211
Also, if you have any recommendations for logo designers please do let me know - its definitely one of the most important things we're missing. I have the artistic talents of a sponge :(.
