### 2020-09-26 3ce1f60
moved aggregator saving to analysi class from phase
