+++ autofit/non_linear/abstract_search.py
@@ -37,7 +37,7 @@ class NonLinearSearch(ABC):
         Parameters
         ------------
         paths : af.Paths
-            Manages all paths, e.g. where the search outputs are stored, the samples, backups, etc.
+            Manages all paths, e.g. where the search outputs are stored, the samples, etc.
         prior_passer : af.PriorPasser
             Controls how priors are passed from the results of this non-linear search to a subsequent non-linear search.
         initializer : non_linear.initializer.Initializer
@@ -84,7 +84,6 @@ class NonLinearSearch(ABC):
             self.iterations_per_update = conf.instance.general.get("hpc", "iterations_per_update", float)
 
         self.log_every_update = self._config("updates", "log_every_update", int)
-        self.backup_every_update = self._config("updates", "backup_every_update", int)
         self.visualize_every_update = self._config(
             "updates", "visualize_every_update", int
         )
@@ -94,7 +93,6 @@ class NonLinearSearch(ABC):
 
         self.iterations = 0
         self.should_log = IntervalCounter(self.log_every_update)
-        self.should_backup = IntervalCounter(self.backup_every_update)
         self.should_visualize = IntervalCounter(self.visualize_every_update)
         self.should_output_model_results = IntervalCounter(
             self.model_results_every_update
@@ -110,7 +108,7 @@ class NonLinearSearch(ABC):
         self._in_phase = False
 
     class Fitness:
-        def __init__(self, paths, model, analysis, samples_from_model, pool_ids=None):
+        def __init__(self, paths, model, analysis, samples_from_model, log_likelihood_cap=None, pool_ids=None):
 
             self.paths = paths
             self.max_log_likelihood = -np.inf
@@ -119,15 +117,16 @@ class NonLinearSearch(ABC):
             self.model = model
             self.samples_from_model = samples_from_model
 
+            self.log_likelihood_cap = log_likelihood_cap
             self.pool_ids = pool_ids
 
         def fit_instance(self, instance):
 
             log_likelihood = self.analysis.log_likelihood_function(instance=instance)
 
-            if self.analysis.log_likelihood_cap is not None:
-                if log_likelihood > self.analysis.log_likelihood_cap:
-                    log_likelihood = self.analysis.log_likelihood_cap
+            if self.log_likelihood_cap is not None:
+                if log_likelihood > self.log_likelihood_cap:
+                    log_likelihood = self.log_likelihood_cap
 
             if log_likelihood > self.max_log_likelihood:
 
@@ -185,7 +184,7 @@ class NonLinearSearch(ABC):
              should be given a likelihood so low that it is discard."""
             return -np.inf
 
-    def fit(self, model, analysis: "Analysis", info=None, pickle_files=None) -> "Result":
+    def fit(self, model, analysis: "Analysis", info=None, pickle_files=None, log_likelihood_cap=None) -> "Result":
         """ Fit a model, M with some function f that takes instances of the
         class represented by model M and gives a score for their fitness.
 
@@ -249,12 +248,12 @@ class NonLinearSearch(ABC):
             samples = self.samples_via_csv_json_from_model(model=model)
             self.save_samples(samples=samples)
 
-        self.paths.backup_zip_remove()
+        self.paths.zip_remove()
 
         return Result(samples=samples, previous_model=model, search=self)
 
     @abstractmethod
-    def _fit(self, model, analysis):
+    def _fit(self, model, analysis, log_likelihood_cap=None):
         pass
 
     @property
@@ -312,11 +311,10 @@ class NonLinearSearch(ABC):
         non-linear search. The update performs the following tasks:
 
         1) Visualize the maximum log likelihood model.
-        2) Backup the samples.
[…]
