### 2020-09-27 641e20f
Now remove dynesty state files to reduce file sizes
