+++ autofit/non_linear/paths.py
@@ -188,6 +188,13 @@ class Paths:
         return f"{self.output_path}/samples"
 
     @property
+    def samples_from_sym_path(self) -> str:
+        """
+        The path to the samples folder.
+        """
+        return f"{self.output_path}/samples_from_sym"
+
+    @property
     def zip_path(self) -> str:
         return f"{self.output_path}.zip"
 
@@ -308,19 +315,19 @@ class Paths:
 
     @property
     def file_summary(self) -> str:
-        return "{}/{}".format(self.path, "multinestsummary.txt")
+        return "{}/{}".format(self.samples_from_sym_path, "multinestsummary.txt")
 
     @property
     def file_weighted_samples(self):
-        return "{}/{}".format(self.path, "multinest.txt")
+        return "{}/{}".format(self.samples_from_sym_path, "multinest.txt")
 
     @property
     def file_phys_live(self) -> str:
-        return "{}/{}".format(self.path, "multinestphys_live.points")
+        return "{}/{}".format(self.samples_from_sym_path, "multinestphys_live.points")
 
     @property
     def file_resume(self) -> str:
-        return "{}/{}".format(self.path, "multinestresume.dat")
+        return "{}/{}".format(self.samples_from_sym_path, "multinestresume.dat")
 
     @property
     def file_search_summary(self) -> str:
@@ -329,6 +336,24 @@ class Paths:
     @property
     def file_results(self):
         return "{}/{}".format(self.output_path, "model.results")
+
+    def copy_from_sym(self):
+        """
+        Copy files from the sym-linked search folder to the samples folder.
+        """
+        try:
+            shutil.rmtree(self.samples_from_sym_path)
+        except FileNotFoundError:
+            pass
+
+     #   print(os.listdir(self.path))
+
+     #   stop
+
+        try:
+            shutil.copytree(self.path, self.samples_from_sym_path)
+        except shutil.Error as e:
+            logger.exception(e)
 
     def zip_remove(self):
         """
