+++ test_autofit/integration/config/non_linear/mcmc/Emcee.ini
@@ -18,6 +18,7 @@ iterations_per_update=2500
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
+remove_state_files_at_end=True
 
 [printing]
 silence=False
