+++ test_autofit/unit/non_linear/nest/files/dynesty/config/non_linear/nest/DynestyStatic.ini
@@ -28,6 +28,7 @@ iterations_per_update=500
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
+remove_state_files_at_end=True
 
 [printing]
 silence=False
