+++ test_autofit/unit/tools/files/config/non_linear/optimize/DownhillSimplex.ini
@@ -12,6 +12,7 @@ iterations_per_update = 11
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
+remove_state_files_at_end=True
 
 [printing]
 silence=False
