### 2020-09-28 #2550 @arm61
100 % agree with your identification of `uravu` as not a PPL.

Full ticks from me now! As for a logo designer, unfortunately I have always done my own (which speak volumes to my skill considering that they are usually text over a matplotlib figure...)
