+++ test_autofit/unit/config/non_linear/optimize/DownhillSimplex.ini
@@ -7,6 +7,9 @@ full_output = 0
 disp = 1
 retall = 0
 
+[initialize]
+method=prior
+
 [updates]
 iterations_per_update = 11
 visualize_every_update=1
