+++ test_autofit/unit/non_linear/files/nlo/config/non_linear/mock/MockSearch.ini
@@ -5,6 +5,9 @@ backup_every_update=1
 model_results_every_update=1
 log_every_update=1
 
+[initialize]
+method=prior
+
 [printing]
 silence=False
 
