### 2020-09-30 fa73a60
cast config path to string
