+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/src/phase/settings.py
@@ -10,7 +10,7 @@ This module create tags for phases settings that customize the analysis. We tag 
     1) Tags describes the phase settings, making it explicit what analysis was used to create the results.
 
     2) Tags create unique output paths, ensuring that if you run multiple phases on the same data with different settings
-       each `NonLinearSearch` (e.g. Emcee) won`t inadvertently use results generated via a different analysis method.
+       each `NonLinearSearch` (e.g. Emcee) won't inadvertently use results generated via a different analysis method.
        
 The settings of a `SettingsPhase` use the settings objects of individual parts of the code. For example, below, it uses
 as input a `SettingsMaskedDataset` object, which in the module `dataset/dataset.py` can be seen contains the inputs and
