### 2020-10-05 d47c315
fixing AbstractNode._broadcast to account for plate reordering
