### 2020-10-06 43537a2
fixing routine to represent messages as strings
