### 2020-10-06 ba22d71
ensuring that str and repr works for FixedMessage
