### 2020-10-06 caf32ea
adding test to check str representation of Messages
