### 2020-10-06 f0162c8
refactoring FactorApproximation to remove deterministic_dist attribute
