### 2020-10-06 b886152
fixing lotka_volterra test to be more stable
