### 2020-10-07 4248fc3
Refactor MeanField
