+++ autofit/text/text_util.py
@@ -11,7 +11,7 @@ def results_to_file(samples, filename, during_analysis):
     Parameters
     ----------
     during_analysis : bool
-        Whether the model.results are being written during the analysis or after the non-linear search has finished.
+        Whether the model.results are being written during the analysis or after the `NonLinearSearch` has finished.
     """
 
     results = []
