+++ citations.bib
@@ -103,7 +103,7 @@ year = {2014}
 }
 @misc{pyautofit,
   author = {James Nightingale and Richard Hayes},
-  title = {PyAutoFit: A classy Probablistic Programming Language for Complex Model-fitting},
+  title = {PyAutoFit: A classy Probabilistic Programming Language for Complex Model-fitting},
   year = {2020},
   publisher = {​GitHub},
   journal = {​GitHub repository},

[…]
