+++ test_autofit/unit/non_linear/nest/files/multinest/config/notation/label.ini
@@ -1,24 +1,0 @@
-[label]
-one=one_label
-two=two_label
-three=three_label
-four=four_label
-centre_0=x
-centre_1=y
-gamma=\gamma
-contribution_factor=\omega0
-redshift=z
-axis_ratio=q
-phi=phi
-intensity=I
-effective_radius=R
-sersic_index=n
-mass_to_light_ratio=Psi
-
-[subscript]
-MockClassx4=a
-MockClassx5=a
-MockClassx6=b
-EllipticalProfile=l
-GeometryProfile=gp
-Galaxy=g
