+++ howtofit/config/notation/label_format.ini
@@ -1,5 +1,0 @@
-[format]
-centre={:.2f}
-intensity={:.2f}
-sigma={:.2f}
-rate={:.2f}
