+++ howtofit/config/general.ini
@@ -1,15 +1,0 @@
-[output]
-log_to_file=False
-log_file=output.log
-log_level=INFO
-model_results_decimal_places=3
-remove_files=False
-skip_completed=True
-force_pickle_overwrite=False
-
-[fits]
-flip_for_ds9=False
-
-[hpc]
-hpc_mode=False
-iterations_per_update=5000
