+++ test_autofit/unit/tools/files/config/non_linear/optimize/DownhillSimplex.ini
@@ -1,26 +1,0 @@
-[search]
-xtol = 1e-4
-ftol = 1e-4
-maxiter = None
-maxfun = None
-full_output = 0
-disp = 1
-retall = 0
-
-[updates]
-iterations_per_update = 11
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=False
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[tag]
-name=dhs
