+++ test_autofit/unit/non_linear/nest/files/dynesty_unconverged/config/non_linear/nest/DynestyStatic.ini
@@ -1,41 +1,0 @@
-[search]
-n_live_points=150
-bound=multi
-sample=auto
-bootstrap=0
-enlarge=-1
-update_interval=-1.0
-vol_dec=0.5
-vol_check=2.0
-walks=5
-sampling_efficiency=0.5
-slices=5
-fmove=0.9
-max_move=100
-maxiter=0
-maxcall=0
-logl_max=inf
-n_effective=0
-
-[settings]
-stagger_resampling_likelihood=False
-terminate_at_acceptance_ratio=False
-acceptance_ratio_threshold=1.0
-
-[updates]
-iterations_per_update=2500
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=False
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[parallel]
-number_of_cores=1
