+++ test_autofit/unit/config/non_linear/mcmc/Emcee.ini
@@ -1,36 +1,0 @@
-[search]
-nwalkers=10
-nsteps=20
-
-[initialize]
-method = prior
-ball_lower_limit = 0.49
-ball_upper_limit = 0.51
-
-[auto_correlation]
-check_for_convergence = False
-check_size = 100
-required_length = 50
-change_threshold=0.01
-
-[updates]
-iterations_per_update=2500
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=False
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[parallel]
-number_of_cores=1
-
-[tag]
-name=emcee
-nwalkers=nwalkers
