+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/src/fit/fit.py
@@ -106,7 +106,7 @@ def residual_map_from_data_model_data_and_mask(data, mask, model_data):
     data : np.ndarray
         The observed data that is fitted.
     mask : np.ndarray
-        The mask applied to the dataset, where ``False`` entries are included in the calculation.
+        The mask applied to the dataset, where `False` entries are included in the calculation.
     model_data : np.ndarray
         The model data used to fit the observed data.
     """
@@ -130,7 +130,7 @@ def normalized_residual_map_from_residual_map_noise_map_and_mask(
     noise_map : np.ndarray
         The noise-map of the observed `Dataset`.
     mask : np.ndarray
-        The mask applied to the residual-map, where ``False`` entries are included in the calculation.
+        The mask applied to the residual-map, where `False` entries are included in the calculation.
     """
     return np.divide(
         residual_map,
@@ -153,7 +153,7 @@ def chi_squared_map_from_residual_map_noise_map_and_mask(residual_map, noise_map
     noise_map : np.ndarray
         The noise-map of the observed `Dataset`.
     mask : np.ndarray
-        The mask applied to the residual-map, where ``False`` entries are included in the calculation.
+        The mask applied to the residual-map, where `False` entries are included in the calculation.
     """
     return np.square(
         np.divide(
@@ -175,7 +175,7 @@ def chi_squared_from_chi_squared_map_and_mask(chi_squared_map, mask):
     chi_squared_map : np.ndarray
         The chi-squared-map of values of the model-simulator fit to the observed `Dataset`.
     mask : np.ndarray
-        The mask applied to the chi-squared-map, where ``False`` entries are included in the calculation.
+        The mask applied to the chi-squared-map, where `False` entries are included in the calculation.
     """
     return np.sum(chi_squared_map[np.asarray(mask) == 0])
 
@@ -191,7 +191,7 @@ def noise_normalization_from_noise_map_and_mask(noise_map, mask):
     noise_map : np.ndarray
         The masked noise-map of the observed `Dataset`.
     mask : np.ndarray
-        The mask applied to the noise-map, where ``False`` entries are included in the calculation.
+        The mask applied to the noise-map, where `False` entries are included in the calculation.
     """
     return np.sum(np.log(2 * np.pi * noise_map[np.asarray(mask) == 0] ** 2.0))
 
