+++ test_autofit/unit/tools/test_metadata.py
@@ -1,7 +1,7 @@
 import pytest
 
 import autofit as af
-from test_autofit import mock
+from autofit import mock
 
 
 @pytest.fixture(
@@ -11,7 +11,7 @@ def make_phase():
     return af.AbstractPhase(
         search=mock.MockSearch(
             paths=af.Paths(
-                name="phase_name",
+                name="name",
                 tag="phase_tag"
             )
         )
@@ -22,7 +22,7 @@ def test_metadata_dictionary(phase):
     phase.pipeline_name = "pipeline_name"
     phase.pipeline_tag = "pipeline_tag"
     assert phase._default_metadata == {
-        "phase": "phase_name",
+        "phase": "name",
         "phase_tag": "phase_tag",
         "pipeline": "pipeline_name",
         "pipeline_tag": "pipeline_tag",

[…]
