+++ howtofit/chapter_2_results/scripts/tutorial_2_fitting_multiple_datasets.py
@@ -30,8 +30,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 
 # %%
 conf.instance = conf.Config(
-    config_path=f"{workspace_path}/config",
-    output_path=f"{workspace_path}/output/chapter_2",
+    config_path=f"config",
+    output_path=f"output/chapter_2",
 )
 
 # %%
