+++ howtofit/chapter_1_introduction/tutorial_4_source_code/tutorial_4_source_code.py
@@ -37,8 +37,8 @@ print("Workspace Path: ", workspace_path)
 
 # %%
 conf.instance = conf.Config(
-    config_path=f"{workspace_path}/config",
-    output_path=f"{workspace_path}/output/chapter_1",  # <- This sets up where the `NonLinearSearch`'s outputs go.
+    config_path=f"config",
+    output_path=f"output/chapter_1",  # <- This sets up where the `NonLinearSearch`'s outputs go.
 )
 
 # %%
