+++ howtofit/chapter_2_results/scripts/tutorial_5_data_and_models.py
@@ -28,8 +28,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 
 # %%
 conf.instance = conf.Config(
-    config_path=f"{workspace_path}/config",
-    output_path=f"{workspace_path}/output/chapter_2",
+    config_path=f"config",
+    output_path=f"output/chapter_2",
 )
 
 # %%
@@ -39,7 +39,7 @@ our path straight to the `Aggregator` results ensuring we don't need to filter o
 """
 
 # %%
-output_path = f"{workspace_path}/output/chapter_2/aggregator"
+output_path = f"output/chapter_2/aggregator"
 agg = af.Aggregator(directory=str(output_path))
 
 # %%
