+++ test_autofit/unit/mapper/promise/test_iteration.py
@@ -50,8 +50,8 @@ class TestIteration:
         promise_0 = phase.result.model.collection[0]
         promise_1 = phase.result.model.collection[1]
 
-        assert isinstance(promise_0, af.prior.Promise)
-        assert isinstance(promise_1, af.prior.Promise)
+        assert isinstance(promise_0, af.Promise)
+        assert isinstance(promise_1, af.Promise)
 
     def test_index_populate_model(self, phase, prior_0, prior_1, results_collection):
         promise_0 = phase.result.model.collection[0]
@@ -77,4 +77,4 @@ class TestIteration:
         promises = list(phase.result.model.collection)
 
         assert len(promises) == 2
-        assert all([isinstance(promise, af.prior.Promise) for promise in promises])
+        assert all([isinstance(promise, af.Promise) for promise in promises])
