+++ test_autofit/unit/mapper/promise/test_promise.py
@@ -206,13 +206,13 @@ class TestCase:
         assert model.prior_count == 0
 
     def test_model_promise(self, model_promise, phase):
-        assert isinstance(model_promise, af.prior.Promise)
+        assert isinstance(model_promise, af.Promise)
         assert model_promise.path == ("one", "parameter")
         assert model_promise.is_instance is False
         assert model_promise._phase is phase
 
     def test_grid_search_promise(self, grid_search_promise, phase):
-        assert isinstance(grid_search_promise, af.prior.Promise)
+        assert isinstance(grid_search_promise, af.Promise)
         assert grid_search_promise.path == ("one", "parameter")
         assert grid_search_promise.is_instance is False
         assert grid_search_promise._phase is not phase
@@ -228,7 +228,7 @@ class TestCase:
         assert result is None
 
     def test_instance_promise(self, instance_promise, phase):
-        assert isinstance(instance_promise, af.prior.Promise)
+        assert isinstance(instance_promise, af.Promise)
         assert instance_promise.path == ("one", "parameter")
         assert instance_promise.is_instance is True
         assert instance_promise._phase is phase
@@ -277,13 +277,13 @@ class TestCase:
     def test_embedded_results(self, phase, collection):
         hyper_result = phase.result.hyper_result
 
-        assert isinstance(hyper_result, af.prior.PromiseResult)
+        assert isinstance(hyper_result, af.PromiseResult)
 
         model_promise = hyper_result.model
         instance_promise = hyper_result.instance
 
-        assert isinstance(model_promise.hyper_galaxy, af.prior.Promise)
-        assert isinstance(instance_promise.hyper_galaxy, af.prior.Promise)
+        assert isinstance(model_promise.hyper_galaxy, af.Promise)
+        assert isinstance(instance_promise.hyper_galaxy, af.Promise)
 
         model = model_promise.populate(collection)
         instance = instance_promise.populate(collection)
