+++ test_autofit/unit/mapper/test_assertion.py
@@ -119,29 +119,29 @@ def make_model(collection):
 class TestPromiseAssertion:
     def test_less_than(self, promise_model, collection, model):
         promise = promise_model.one < promise_model.two
-        assert isinstance(promise, af.prior.GreaterThanLessThanAssertion)
+        assert isinstance(promise, af.GreaterThanLessThanAssertion)
 
         assertion = promise.populate(collection)
-        assert isinstance(assertion, af.prior.GreaterThanLessThanAssertion)
+        assert isinstance(assertion, af.GreaterThanLessThanAssertion)
 
     def test_greater_than(self, promise_model, collection, model):
         promise = promise_model.one > promise_model.two
-        assert isinstance(promise, af.prior.GreaterThanLessThanAssertion)
+        assert isinstance(promise, af.GreaterThanLessThanAssertion)
 
     def test_greater_than_equal(self, promise_model, collection, model):
         promise = promise_model.one >= promise_model.two
-        assert isinstance(promise, af.prior.GreaterThanLessThanEqualAssertion)
+        assert isinstance(promise, af.GreaterThanLessThanEqualAssertion)
 
     def test_integer_promise_assertion(self, promise_model, collection, model):
         promise = promise_model.one > 1.0
-        assert isinstance(promise, af.prior.GreaterThanLessThanAssertion)
+        assert isinstance(promise, af.GreaterThanLessThanAssertion)
 
     def test_compound_assertion(self, promise_model, collection, model):
         promise = (1.0 < promise_model.one) < 1.0
-        assert isinstance(promise, af.prior.CompoundAssertion)
+        assert isinstance(promise, af.CompoundAssertion)
 
         assertion = promise.populate(collection)
-        assert isinstance(assertion, af.prior.CompoundAssertion)
+        assert isinstance(assertion, af.CompoundAssertion)
 
 
 class TestModel:
