### 2020-10-14 e73d929
updating graphical optimisation algorithms
