### 2020-10-14 271ecff
Bugfixes for graphical.optimize
