+++ autofit/graphical/optimise.py
@@ -7,16 +7,8 @@ from scipy.optimize import minimize, OptimizeResult, least_squares
 from autofit.graphical import FixedMessage
 from autofit.mapper.variable import Variable
 from autofit.graphical.factor_graphs import Factor
-from .mean_field import FactorApproximation, MeanFieldApproximation, Status
-from .utils import propagate_uncertainty, FlattenArrays
-
-
-class OptResult(NamedTuple):
-    mode: Dict[str, np.ndarray]
-    inv_hessian: Dict[str, np.ndarray]
-    log_norm: float
-    result: OptimizeResult
-    status: Optional[Status]
+from autofit.graphical.mean_field import FactorApproximation, MeanFieldApproximation, Status
+from autofit.graphical.utils import propagate_uncertainty, FlattenArrays, OptResult
 
 
 class OptFactor:
@@ -99,9 +91,11 @@ class OptFactor:
             "optimise.find_factor_mode: "
             f"nfev={result.nfev}, nit={result.nit}, "
             f"status={result.status}, message={message}",)
+        mode = self.param_shapes.unflatten(result.x)
+        mode.update(self.fixed_kws)
+        covar = self.param_shapes.unflatten(result.hess_inv.todense())
         return OptResult(
-            self.param_shapes.unflatten(result.x),
-            self.param_shapes.unflatten(result.hess_inv.todense()),
+            mode, covar,
             -result.fun, # minimized negative logpdf of factor approximation
             result,
             Status(success, messages))
@@ -121,7 +115,8 @@ class OptFactor:
             arrays_dict={},
             bounds=None,
             constraints=(), tol=None, callback=None,
-            options=None
+            options=None,
+            status=None,
     ):
         self.sign = -1
         p0 = {
@@ -132,12 +127,12 @@ class OptFactor:
             bounds=bounds, constraints=constraints, tol=tol,
             callback=callback, options=options)
         self.sign = 1
-        return self._parse_result(res)
+        return self._parse_result(res, status=status)
 
     minimize = minimise
     maximize = maximise
 
-def update_det_inv_hessian(
+def update_det_cov(
         res: OptResult,
         jacobian: Dict[Variable, np.ndarray]):
     """Calculates the inv hessian of the deterministic variables
@@ -147,8 +142,7 @@ def update_det_inv_hessian(
     covars = res.inv_hessian
     for (det, v), jac in jacobian.deterministic_values.items():
         cov = covars[v]
-        det_cov = propagate_uncertainty(cov, jac)
-        covars[det] += covars.get(det, 0.) + det_cov
+        covars[det] = covars.get(det, 0.) + propagate_uncertainty(cov, jac)
 
     return res
 
@@ -165,62 +159,32 @@ def maximise_factor_approx(
 
 maximize_factor_approx = maximise_factor_approx
 
-
 def find_factor_mode(
         factor_approx: FactorApproximation,
         return_cov: bool = True,
         status: Optional[Status] = None,
         min_iter: int = 2,
+        opt_kws: Optional[dict] = None,
         **kwargs
-):
+    ) -> OptResult:
     """
     """
-    success, messages = Status() if status is None else status
-    factor = factor_approx.factor
+    opt_kws = {} if opt_kws is None else opt_kws 
 
     opt = OptFactor.from_approx(factor_approx, **kwargs)
-    free_vars = opt.free_vars
-    det_vars = factor_approx.deterministic_dist
-    p0 = {
-        v: kwargs.pop(v, factor_approx.model_dist[v].sample(1)[0])
-        for v in free_vars}
-
-    # find the mode of the factor approximation
-    res = opt.maximise(p0)
-    mode = {**res.mode, **opt.fixed_kws}
-
[…]
