+++ autofit/non_linear/mcmc/emcee.py
@@ -11,11 +11,14 @@ from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.mcmc.abstract_mcmc import AbstractMCMC
 from autofit.non_linear.samples import MCMCSamples
 from autofit.mapper.model_mapper import ModelMapper
+from autofit.non_linear.paths import convert_paths
 
 from autofit.non_linear.log import logger
 
 
 class Emcee(AbstractMCMC):
+
+    @convert_paths
     def __init__(
         self,
         paths=None,
