+++ autofit/non_linear/abstract_search.py
@@ -249,18 +249,6 @@ class NonLinearSearch(ABC):
             samples = self.samples_via_csv_json_from_model(model=model)
             self.save_samples(samples=samples)
 
-            if self.remove_state_files_at_end:
-                try:
-                    self.remove_state_files()
-                except FileNotFoundError:
-                    pass
-
-                try:
-                    shutil.rmtree(f"{self.paths.samples_path}_backup")
-                except FileNotFoundError:
-                    pass
-
-
         self.paths.zip_remove()
 
         return Result(samples=samples, previous_model=model, search=self)
@@ -674,7 +662,7 @@ class PriorPasser:
             3) The sigma of the Gaussian will use the maximum of two values:
 
                     (i) the 1D error of the parameter computed at an input sigma value (default sigma=3.0).
-                    (ii) The value specified for the profile in the 'config/json_priors/*.json' config
+                    (ii) The value specified for the profile in the 'config/priors/*.json' config
                          file's 'width_modifer' field (check these files out now).
 
                The idea here is simple. We want a value of sigma that gives a GaussianPrior wide enough to search a
@@ -725,7 +713,7 @@ class PriorPasser:
         to 4.0 +- 0.5, the sigma of the Gaussian prior would instead be 0.5.
 
         If the error on the parameter in phase 1 had been really small, lets say, 0.01, we would instead use the value
-        of the parameter width in the json_priors config file to set sigma instead. Lets imagine the prior config file
+        of the parameter width in the priors config file to set sigma instead. Lets imagine the prior config file
         specifies that we use an "Absolute" value of 0.8 to link this prior. Then, the GaussianPrior in phase 2 would
         have a mean=4.0 and sigma=0.8.
 
