+++ autofit/config/non_linear/nest/MultiNest.ini
@@ -1,0 +1,51 @@
+[search]
+n_live_points=50
+sampling_efficiency=0.2
+const_efficiency_mode=False
+evidence_tolerance=0.5
+multimodal=False
+importance_nested_sampling=False
+max_modes=100
+mode_tolerance=-1e90
+max_iter=0
+
+[settings]
+n_iter_before_update=5
+null_log_evidence=-1e90
+seed=-1
+verbose=False
+resume=True
+context=0
+write_output=True
+log_zero=-1e100
+init_MPI=False
+stagger_resampling_likelihood=False
+terminate_at_acceptance_ratio=False
+acceptance_ratio_threshold=1.0
+
+[updates]
+iterations_per_update=2500
+visualize_every_update=1
+model_results_every_update=1
+log_every_update=1
+remove_state_files_at_end=False
+
+[initialize]
+method=prior
+
+[printing]
+silence=False
+
+[prior_passer]
+sigma=3.0
+use_errors=True
+use_widths=True
+
+[tag]
+name=multinest
+n_live_points=nlive
+sampling_efficiency=eff
+const_efficiency_mode=const
+multimodal=mm
+importance_nested_sampling=is
+stagger_resampling_value=1e80
