+++ autofit/config/general.ini
@@ -3,7 +3,7 @@ log_to_file=False
 log_file=output.log
 log_level=INFO
 model_results_decimal_places=3
-remove_files=True
+remove_files=False
 skip_completed=True
 force_pickle_overwrite=False
 
