+++ test_autofit/unit/non_linear/nest/files/multinest/config/general.ini
@@ -8,9 +8,6 @@ remove_files=True
 skip_completed=True
 force_pickle_overwrite=False
 
-[fits]
-flip_for_ds9=False
-
 [hpc]
 hpc_mode=False
 iterations_per_update=5000
