+++ test_autofit/unit/non_linear/files/nlo/config/general.ini
@@ -6,9 +6,6 @@ remove_files = True
 skip_completed=True
 force_pickle_overwrite=False
 
-[fits]
-flip_for_ds9=False
-
 [hpc]
 hpc_mode=False
 iterations_per_update=5000
