+++ autofit/config/non_linear/optimize/PySwarmsGlobal.ini
@@ -1,0 +1,40 @@
+[search]
+n_particles=50
+iters=2000
+cognitive=0.5
+social=0.3
+inertia=0.9
+ftol=-inf
+
+[initialize]
+method=prior
+ball_lower_limit=0.49
+ball_upper_limit=0.51
+
+[updates]
+iterations_per_update=500
+visualize_every_update=1
+model_results_every_update=1
+log_every_update=1
+remove_state_files_at_end=False
+
+[initialize]
+method=prior
+
+[printing]
+silence=False
+
+[prior_passer]
+sigma=3.0
+use_errors=True
+use_widths=True
+
+[parallel]
+number_of_cores=1
+
+[tag]
+name=pyswarms_global
+n_particles=particles
+cognitive=c
+social=s
+inertia=i
