+++ test_autofit/unit/non_linear/nest/test_dynesty.py
@@ -1,11 +1,12 @@
 import os
+import pickle
 import sys
+
+import numpy as np
 import pytest
 
-from autoconf import conf
 import autofit as af
-import pickle
-import numpy as np
+from autoconf import conf
 from autofit import mock
 
 directory = os.path.dirname(os.path.realpath(__file__))
@@ -14,10 +15,9 @@ pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
 
 @pytest.fixture(autouse=True)
 def set_config_path():
-    conf.instance = conf.Config(
-        config_path=os.path.join(directory, "files/dynesty/config"),
-        output_path=os.path.join(directory, "files/dynesty/output"),
-    )
+    conf.instance.push(os.path.join(directory, "files/dynesty/config"),
+                       output_path=os.path.join(directory, "files/dynesty/output"),
+                       )
 
 
 class MockDynestyResults:
@@ -37,7 +37,6 @@ class MockDynestySampler:
 
 class TestDynestyConfig:
     def test__loads_from_config_file_if_not_input(self):
-
         dynesty = af.DynestyStatic(
             prior_passer=af.PriorPasser(sigma=2.0, use_errors=False, use_widths=False),
             n_live_points=151,
@@ -198,14 +197,13 @@ class TestDynestyConfig:
         assert dynesty.number_of_cores == 4
 
     def test__tag(self):
-
         dynesty = af.DynestyStatic(
             n_live_points=40, bound="none", sample="auto", enlarge=1.0
         )
 
         assert (
-            dynesty.tag
-            == "dynesty_static[nlive_40__bound_none__enlarge_1.0__sample_auto]"
+                dynesty.tag
+                == "dynesty_static[nlive_40__bound_none__enlarge_1.0__sample_auto]"
         )
 
         dynesty = af.DynestyStatic(
@@ -218,8 +216,8 @@ class TestDynestyConfig:
         )
 
         assert (
-            dynesty.tag
-            == "dynesty_static[nlive_41__bound_multi_vol_dec_2.0_vol_check_3.0__enlarge_1.0__sample_unif]"
+                dynesty.tag
+                == "dynesty_static[nlive_41__bound_multi_vol_dec_2.0_vol_check_3.0__enlarge_1.0__sample_unif]"
         )
 
         dynesty = af.DynestyStatic(
@@ -232,8 +230,8 @@ class TestDynestyConfig:
         )
 
         assert (
-            dynesty.tag
-            == "dynesty_static[nlive_43__bound_single__enlarge_1.0__sample_rwalk_walks_1_facc_0.5]"
+                dynesty.tag
+                == "dynesty_static[nlive_43__bound_single__enlarge_1.0__sample_rwalk_walks_1_facc_0.5]"
         )
 
         dynesty = af.DynestyStatic(
@@ -246,8 +244,8 @@ class TestDynestyConfig:
         )
 
         assert (
-            dynesty.tag
-            == "dynesty_static[nlive_44__bound_balls__enlarge_1.0__sample_hslice_slices_1_max_move_2]"
+                dynesty.tag
+                == "dynesty_static[nlive_44__bound_balls__enlarge_1.0__sample_hslice_slices_1_max_move_2]"
         )
 
         dynesty = af.DynestyStatic(
@@ -255,8 +253,8 @@ class TestDynestyConfig:
         )
 
         assert (
-            dynesty.tag
-            == "dynesty_static[nlive_44__bound_balls__enlarge_1.0__sample_slice_slices_1]"
+                dynesty.tag
+                == "dynesty_static[nlive_44__bound_balls__enlarge_1.0__sample_slice_slices_1]"
         )
 
[…]
