+++ howtofit/chapter_1_introduction/tutorial_7_phase_customization/tutorial_7_phase_customization.py
@@ -40,8 +40,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 """
 
 # %%
-conf.instance = conf.Config(
-    config_path=f"config",
+conf.instance.push(
+f"config",
     output_path=f"output/chapter_1",
 )
 
