+++ howtofit/chapter_2_results/scripts/tutorial_3_aggregator.py
@@ -29,8 +29,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 """
 
 # %%
-conf.instance = conf.Config(
-    config_path=f"config",
+conf.instance.push(
+f"config",
     output_path=f"output/chapter_2",
 )
 
