+++ test_autofit/integration/tests/searches/pipeline_composition.py
@@ -47,8 +47,8 @@ def make_pipeline_1():
 
 integration_path = "{}/../..".format(os.path.dirname(os.path.realpath(__file__)))
 
-conf.instance = conf.Config(
-    config_path=f"{integration_path}/config",
+conf.instance.push(
+f"{integration_path}/config",
     output_path=f"{integration_path}/output/{test_type}"
 )
 
