+++ howtofit/chapter_2_results/scripts/tutorial_1_results_and_samples.py
@@ -34,8 +34,8 @@ Setup the configs as we did in the previous tutorial, as well as the output fold
 """
 
 # %%
-conf.instance = conf.Config(
-    config_path=f"config",
+conf.instance.push(
+f"config",
     output_path=f"output/chapter_2",
 )
 
