### 2020-10-15 518bc0d
refactoring LaplaceOptimiser
