### 2020-10-15 bc1c6b9
fixing bug in test_lotka_volterra
