+++ test_autofit/unit/non_linear/test_abstract_search.py
@@ -7,8 +7,8 @@ import pytest
 
 import autofit as af
 from autoconf import conf
-from autofit.non_linear.mock.mock_search import MockSamples
 from autofit import mock
+from autofit.non_linear.mock.mock_search import MockSamples
 
 directory = os.path.dirname(os.path.realpath(__file__))
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
@@ -16,7 +16,7 @@ pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
 
 @pytest.fixture(autouse=True)
 def set_config_path():
-    conf.instance = conf.Config(
+    conf.instance.push(
         os.path.join(directory, "files/nlo/config"),
         output_path=os.path.join(directory, "files/nlo/output"),
     )
