### 2020-10-15 3af3365
fixing more bugs in optimise
