+++ autofit/__init__.py
@@ -11,23 +11,23 @@ from .mapper.model import path_instances_of_class
 from .mapper.model_mapper import ModelMapper
 from .mapper.model_mapper import ModelMapper as Mapper
 from .mapper.model_object import ModelObject
-from .mapper.prior.prior import WidthModifier
-from .mapper.prior.prior import RelativeWidthModifier
-from .mapper.prior.prior import AbsoluteWidthModifier
+from .mapper.prior.assertion import ComparisonAssertion
+from .mapper.prior.assertion import ComparisonAssertion
 from .mapper.prior.assertion import GreaterThanLessThanAssertion
 from .mapper.prior.assertion import GreaterThanLessThanEqualAssertion
-from .mapper.prior.assertion import ComparisonAssertion
-from .mapper.prior.promise import AbstractPromise
+from .mapper.prior.deferred import DeferredArgument
+from .mapper.prior.deferred import DeferredInstance
+from .mapper.prior.prior import AbsoluteWidthModifier
 from .mapper.prior.prior import GaussianPrior
 from .mapper.prior.prior import LogUniformPrior
 from .mapper.prior.prior import Prior
+from .mapper.prior.prior import RelativeWidthModifier
 from .mapper.prior.prior import UniformPrior
-from .mapper.prior.assertion import ComparisonAssertion
+from .mapper.prior.prior import WidthModifier
+from .mapper.prior.promise import AbstractPromise
 from .mapper.prior.promise import Promise
 from .mapper.prior.promise import PromiseResult
 from .mapper.prior.promise import last
-from .mapper.prior.deferred import DeferredArgument
-from .mapper.prior.deferred import DeferredInstance
 from .mapper.prior_model.abstract import AbstractPriorModel
 from .mapper.prior_model.annotation import AnnotationPriorModel
 from .mapper.prior_model.attribute_pair import AttributeNameValue
@@ -43,38 +43,40 @@ from .mapper.prior_model.prior_model import PriorModel as Model
 from .mapper.prior_model.util import PriorModelNameValue
 from .non_linear.abstract_search import Analysis
 from .non_linear.abstract_search import NonLinearSearch
-from .non_linear.abstract_search import Result
 from .non_linear.abstract_search import PriorPasser
-from .non_linear.optimize.downhill_simplex import DownhillSimplex
-from .non_linear.optimize.pyswarms import PySwarmsLocal
-from .non_linear.optimize.pyswarms import PySwarmsGlobal
+from .non_linear.abstract_search import Result
+from .non_linear.grid_search import GridSearch as NonLinearSearchGridSearch
+from .non_linear.grid_search import GridSearchResult
+from .non_linear.initializer import InitializerBall
+from .non_linear.initializer import InitializerPrior
 from .non_linear.mcmc.emcee import Emcee
-from .non_linear.mock.mock_search import MockSearch
 from .non_linear.mock.mock_search import MockResult
+from .non_linear.mock.mock_search import MockSearch
 from .non_linear.nest.dynesty import DynestyDynamic
 from .non_linear.nest.dynesty import DynestyStatic
 from .non_linear.nest.multi_nest import MultiNest
-from .non_linear.initializer import InitializerPrior
-from .non_linear.initializer import InitializerBall
+from .non_linear.optimize.downhill_simplex import DownhillSimplex
+from .non_linear.optimize.pyswarms import PySwarmsGlobal
+from .non_linear.optimize.pyswarms import PySwarmsLocal
 from .non_linear.paths import Paths
 from .non_linear.paths import convert_paths
 from .non_linear.paths import make_path
-from .non_linear.samples import OptimizerSamples
-from .non_linear.samples import PDFSamples
 from .non_linear.samples import MCMCSamples
[…]
