### 2020-10-15 cc59f60
remove the pickled child.
