+++ autofit/aggregator/__init__.py
@@ -1,6 +1,1 @@
 from .aggregator import *
-
-
-class PickledChild:
-    def __init__(self, age):
-        self.age = age
