+++ test_autofit/unit/non_linear/nest/files/dynesty/config/non_linear/nest/DynestyStatic.ini
@@ -1,59 +1,0 @@
-[search]
-n_live_points=150
-sampling_efficiency=0.5
-evidence_tolerance=-1.0
-bound=multi
-sample=auto
-bootstrap=0
-enlarge=-1
-update_interval=-1.0
-vol_dec=0.5
-vol_check=2.0
-walks=5
-slices=5
-fmove=0.9
-max_move=100
-maxiter=0
-maxcall=0
-logl_max=inf
-n_effective=0
-
-[settings]
-stagger_resampling_likelihood=False
-terminate_at_acceptance_ratio=True
-acceptance_ratio_threshold=2.0
-
-[updates]
-iterations_per_update=500
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=False
-
-[initialize]
-method=prior
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[parallel]
-number_of_cores=1
-
-[tag]
-name=dynesty_static
-n_live_points=nlive
-sample=sample
-bound=bound
-enlarge=enlarge
-vol_dec=vol_dec
-vol_check=vol_check
-walks=walks
-slices=slices
-max_move=max_move
-facc=facc
-sampling_efficiency=eff
