+++ test_autofit/unit/non_linear/nest/files/dynesty/config/non_linear/nest/DynestyDynamic.ini
@@ -1,59 +1,0 @@
-[search]
-n_live_points=5
-sampling_efficiency=0.6
-evidence_tolerance=0.01
-bound=balls
-sample=rwalk
-bootstrap=1
-enlarge=2
-update_interval=2.0
-vol_dec=0.4
-vol_check=3.0
-walks=26
-slices=6
-fmove=0.8
-max_move=101
-maxiter=0
-maxcall=0
-logl_max=inf
-n_effective=0
-
-[settings]
-stagger_resampling_likelihood=False
-terminate_at_acceptance_ratio=True
-acceptance_ratio_threshold=2.0
-
-[updates]
-iterations_per_update=501
-visualize_every_update=1
-model_results_every_update=1
-log_every_update=1
-remove_state_files_at_end=False
-
-[initialize]
-method=prior
-
-[printing]
-silence=False
-
-[prior_passer]
-sigma=3.0
-use_errors=True
-use_widths=True
-
-[parallel]
-number_of_cores=4
-
-[tag]
-name=dynesty_dynamic
-n_live_points=nlive
-sample=sample
-bound=bound
-enlarge=enlarge
-vol_dec=vol_dec
-vol_check=vol_check
-walks=walks
-slices=slices
-max_move=max_move
-facc=facc
-sampling_efficiency=eff
