### 2020-10-15 c31ce2a
Revert "remove the pickled child."
This reverts commit cc59f6082834c17d09f347d2d0ab6112d984d8ad.
