+++ test_autofit/unit/non_linear/nest/files/multinest/config/non_linear/nest/MultiNest.ini
@@ -30,9 +30,6 @@ model_results_every_update=1
 log_every_update=1
 remove_state_files_at_end=False
 
-[initialize]
-method=prior
-
 [printing]
 silence=False
 
