+++ test_autofit/unit/config/notation/label_format.ini
@@ -1,0 +1,10 @@
+[format]
+param0 = {:.2f}
+param00 = {:.2f}
+param000 = {:.2f}
+param11 = {:.4f}
+param12 = {:.2e}
+one = {:.2f}
+two = {:.2f}
+radius = radius_value
+mass = {:.2f}
