+++ test_autofit/unit/config/notation/label.ini
@@ -1,0 +1,31 @@
+[label]
+parameter=p
+param0=p0
+one=one_label
+two=two_label
+three=three_label
+four=four_label
+one_tuple_0=one_tuple_0_label
+one_tuple_1=one_tuple_1_label
+centre_0=x
+centre_1=y
+gamma=\gamma
+contribution_factor=\omega0
+redshift=z
+axis_ratio=q
+phi=phi
+intensity=I
+effective_radius=R
+sersic_index=n
+mass_to_light_ratio=Psi
+
+[subscript]
+MockClassx2=a
+MockClassx4=a
+MockClassx2Tuple=a
+MockComponents=a
+MockClassRelativeWidth=r
+EllipticalProfile=l
+GeometryProfile=gp
+Galaxy=g
+Class=c
