+++ test_autofit/unit/non_linear/nest/files/dynesty_unconverged/config/non_linear/nest/DynestyDynamic.ini
@@ -1,0 +1,16 @@
+[search]
+bound=balls
+sample=rwalk
+bootstrap=1
+enlarge=2
+update_interval=2.0
+vol_dec=0.4
+vol_check=3.0
+walks=26
+facc=0.6
+slices=6
+fmove=0.8
+max_move=101
+
+[settings]
+iterations_per_update=501
