+++ test_autofit/unit/non_linear/mcmc/files/emcee/config/non_linear/mcmc/Emcee.ini
@@ -1,0 +1,36 @@
+[search]
+nwalkers=50
+nsteps=2000
+
+[initialize]
+method = prior
+ball_lower_limit = 0.49
+ball_upper_limit = 0.51
+
+[auto_correlation]
+check_for_convergence = True
+check_size = 100
+required_length = 50
+change_threshold=0.01
+
+[updates]
+iterations_per_update=2500
+visualize_every_update=1
+model_results_every_update=1
+log_every_update=1
+remove_state_files_at_end=False
+
+[printing]
+silence=False
+
+[prior_passer]
+sigma=3.0
+use_errors=True
+use_widths=True
+
+[parallel]
+number_of_cores=1
+
+[tag]
+name=emcee
+nwalkers=nwalkers
