+++ howtofit/chapter_2_results/scripts/tutorial_3_aggregator.py
@@ -14,7 +14,6 @@ API for using `Result`'s follow closely tutorial 1 of this chapter.
 #%matplotlib inline
 
 from pyprojroot import here
-
 workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
