+++ requirements.txt
@@ -1,12 +1,10 @@
-autoconf==0.6.2
+autoconf==0.7.0
 corner>=2.0.1
 decorator>=4.2.1
 dill>=0.3.1.1
 dynesty==1.0.1
 typing-inspect>=0.4.0
 emcee==3.0.2
-nbsphinx>=0.7.0
-nbsphinx-link>=1.3.0
 numpydoc>=1.0.0
 pyprojroot==0.2.0
 pyswarms==1.1.0
