+++ howtofit/chapter_1_introduction/tutorial_2_model_fitting/tutorial_2_model_fitting.py
@@ -25,7 +25,7 @@ Load the dataset we will fit from the `autofit_workspace/dataset` folder.
 """
 
 # %%
-dataset_path = "dataset/howtofit/chapter_1/gaussian_x1"
+dataset_path = "dataset/chapter_1/gaussian_x1"
 data = af.util.numpy_array_from_json(file_path=f"{dataset_path}/data.json")
 noise_map = af.util.numpy_array_from_json(file_path=f"{dataset_path}/noise_map.json")
 
