+++ howtofit/chapter_1_introduction/tutorial_4_source_code/tutorial_4_source_code.py
@@ -72,7 +72,7 @@ To create the `Dataset`, we load it from the `autofit_workspace/dataset` folder 
 """
 
 # %%
-dataset_path = "dataset/howtofit/chapter_1/gaussian_x1"
+dataset_path = "dataset/chapter_1/gaussian_x1"
 data = af.util.numpy_array_from_json(file_path=f"{dataset_path}/data.json")
 noise_map = af.util.numpy_array_from_json(file_path=f"{dataset_path}/noise_map.json")
 
