### 2020-10-18 cf5ea27
config files in manifest
