+++ MANIFEST.in
@@ -5,12 +5,11 @@ exclude .travis.yml
 exclude docs
 include README.rst
 include setup.cfg
-include autofit/config
+recursive-include autofit/config *
 prune .cache
 prune .git
 prune build
 prune dist
-recursive-include autolens/config *
 recursive-exclude *.egg-info *
 recursive-include tests *
 include requirements.txt
