### 2020-10-18 02b4e06
Fixing calculation of log_normalisation of messages, adding tests in test_evidence.py
