+++ autofit/graphical/mean_field.py
@@ -6,8 +6,8 @@ from typing import (
 
 import numpy as np
 
-from autofit.graphical.factor_graphs import Factor
-from autofit.graphical.factor_graphs.graph import FactorGraph
+from autofit.graphical.factor_graphs import (
+    Factor, AbstractNode, FactorGraph)
 from autofit.graphical.messages import FixedMessage, map_dists
 from autofit.graphical.messages.abstract import AbstractMessage
 from autofit.graphical.utils import prod, add_arrays, OptResult, Status
@@ -132,10 +132,11 @@ class MeanField(Dict[Variable, AbstractMessage], Factor):
         return all(d.is_valid for d in self.values())
     
     def prod(self, *approxs: 'MeanField') -> 'MeanField':
-        return type(self)({
-            k: prod((m.get(k, 1) for m in approxs), m) 
+        return MeanField({
+            k: prod((m.get(k, 1.) for m in approxs), m) 
             for k, m in self.items()},
-            sum((m.log_norm for m in approxs), self.log_norm))
+            sum((m.log_norm for m in approxs), 
+            getattr(self, "log_norm", 0.)))
 
     __mul__ = prod
     
@@ -180,10 +181,8 @@ class MeanField(Dict[Variable, AbstractMessage], Factor):
 class FactorApproximation(NamedTuple):
     factor: Factor
     cavity_dist: MeanField
-    # deterministic_dist: MeanField
     factor_dist: MeanField
     model_dist: MeanField
-    # log_norm: float = 0.
 
     @property
     def deterministic_variables(self):
@@ -206,7 +205,6 @@ class FactorApproximation(NamedTuple):
     def is_valid(self) -> bool:
         dists = chain(
             self.cavity_dist.values(),
-            self.deterministic_dist.values(),
             self.factor_dist.values(),
             self.model_dist.values())
         return all(d.is_valid for d in dists if isinstance(d, AbstractMessage))
@@ -379,3 +377,131 @@ class MeanFieldApproximation:
                     self._variable_factor_dist.values()
                     for dist in factors.values())
                 and all(dist.is_valid for dist in self.approx.values()))
+
+
+class EPMeanField(AbstractNode):
+    
+    '''
+    TODO: rename this EP approximation
+    '''
+    def __init__(
+            self,
+            factor_graph: FactorGraph,
+            factor_mean_field: Dict[Factor, MeanField]
+    ):
+        self._factor_graph = factor_graph
+        self._factor_mean_field = factor_mean_field
+        variable_factor = {}
+        for factor, vs in factor_graph.factor_all_variables.items():
+            for v in vs:
+                variable_factor.setdefault(v, set()).add(factor)
+        self._variable_factor = variable_factor
+        
+        super().__init__(**self.factor_graph._kwargs)
+
+    @property 
+    def name(self):
+        return f"EP_{self.factor_graph.name}"
+
+    @property
+    def variables(self): 
+        return self.factor_graph.variables
+
+    @property
+    def deterministic_variables(self): 
+        return self.factor_graph.deterministic_variables
+
+    @property
+    def variable_names(self) -> Dict[str, Variable]: 
+        return self.factor_graph.variable_names
+ 
+    @property
+    def factor_mean_field(self) -> Dict[Factor, MeanField]:
+        return self._factor_mean_field.copy()
+
+    @property
+    def factor_graph(self) -> FactorGraph:
+        return self._factor_graph
+
+    @classmethod
+    def from_approx_dists(
+            cls,
+            factor_graph: FactorGraph,
+            approx_dists: Dict[Variable, AbstractMessage],
+    ) -> "EPMeanField":
+        factor_mean_field = {
+            factor: MeanField({
[…]
