### 2020-10-18 3691a2b
fixing some bugs in test_evidence
