### 2020-10-19 c5a1eaa
fixed pipelines paths
