### 2020-10-19 fb1f107
refactoring some log_normalisation calculations
