+++ test_autofit/unit/graphical/test_evidence.py
@@ -5,25 +5,30 @@ from scipy import stats, integrate
 import autofit.graphical.messages.normal
 from autofit import graphical as graph
 
+def test_message_norm():
+    messages = [
+        tuple(
+            map(graph.NormalMessage, 
+                [0.5, 0.1], [0.2, 0.3])),
+        tuple(
+            map(graph.NormalMessage, 
+                [0.5, 0.1], [0.2, 0.3], [-0.5, 1.3])),
+        tuple(
+            map(graph.GammaMessage, 
+                [0.5, 1.1], [0.2, 1.3])),
+        tuple(
+            map(graph.GammaMessage, 
+                [0.5, 1.1], [0.2, 1.3], [2, 1])),
+    ]
+    for ms in messages:
+        m1, *m2s = ms
+        A = np.exp(m1.log_normalisation(*m2s))
 
+        # Calculate normalisation numerically
+        i12, ierr = integrate.quad(
+            lambda x: np.exp(sum(m.logpdf(x) for m in ms)), 
+            *m1._support[0])
 
-def test_normal_log_norm():
-    m1 = graph.NormalMessage(0.5, 1.3)
-    m2 = graph.NormalMessage(-0.1, 0.2)
-
-    A = np.exp(m1.log_normalisation(m2))
-    i12, ierr = integrate.quad(
-        lambda x: m1.pdf(x) * m2.pdf(x), -np.inf, np.inf)
-
-    assert np.abs(A - i12) < ierr < 1e-6
-
-
-def test_gamma_log_norm():
-    m1 = graph.GammaMessage(1.5, 1.3)
-    m2 = graph.GammaMessage(1.3, 1.2)
-
-    A = np.exp(m1.log_normalisation(m2))
-    i12, ierr = integrate.quad(
-        lambda x: m1.pdf(x) * m2.pdf(x), 0, np.inf)
-
-    assert np.abs(A - i12) < ierr < 1e-6
+        # verify within tolerance
+        print(ms)
+        assert np.abs(A - i12) < ierr < 1e-6
