### 2020-10-19 5c04ecb
adding log_evidence property to EPMeanField
