+++ test_autofit/unit/non_linear/nest/files/dynesty/config/non_linear/nest/DynestyDynamic.ini
@@ -28,7 +28,7 @@ iterations_per_update=501
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
-remove_state_files_at_end=False
+remove_state_files_at_end=True
 
 [initialize]
 method=prior
