+++ test_autofit/unit/non_linear/optimize/files/pyswarms/config/non_linear/optimize/PySwarmsGlobal.ini
@@ -16,7 +16,7 @@ iterations_per_update = 11
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
-remove_state_files_at_end=False
+remove_state_files_at_end=True
 
 [printing]
 silence=False
