+++ test_autofit/unit/non_linear/files/nlo/config/non_linear/mock/MockSearch.ini
@@ -3,7 +3,7 @@ iterations_per_update = 11
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
-remove_state_files_at_end=False
+remove_state_files_at_end=True
 
 [initialize]
 method=prior
