+++ test_autofit/unit/config/non_linear/optimize/DownhillSimplex.ini
@@ -15,7 +15,7 @@ iterations_per_update = 11
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
-remove_state_files_at_end=False
+remove_state_files_at_end=True
 
 [printing]
 silence=False
