+++ test_autofit/unit/non_linear/nest/files/dynesty_unconverged/config/non_linear/nest/DynestyStatic.ini
@@ -27,7 +27,7 @@ iterations_per_update=2500
 visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
-remove_state_files_at_end=False
+remove_state_files_at_end=True
 
 [printing]
 silence=False
