+++ howtofit/chapter_2_results/scripts/tutorial_5_data_and_models.py
@@ -13,13 +13,12 @@ inspect, interpret and plot the results of the model-fit, including refitting th
 #%matplotlib inline
 
 from pyprojroot import here
-
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
-from autofit_workspace.howtofit.chapter_2_results import src as htf
+from howtofit.chapter_2_results import src as htf
 
 
 # %%
@@ -186,7 +185,7 @@ fits = [
 
 # %%
 """
-We can now plot different components of the `Fit` (again benefiting from autofit_workspace.how we set up the `fit_plots.py` module)!
+We can now plot different components of the `Fit` (again benefiting from how we set up the `fit_plots.py` module)!
 """
 
 # %%
