+++ howtofit/chapter_1_introduction/tutorial_4_source_code/tutorial_4_source_code.py
@@ -57,7 +57,7 @@ more intuitive.
 """
 
 # %%
-from autofit_workspace.howtofit.chapter_1_introduction.tutorial_4_source_code import (
+from howtofit.chapter_1_introduction.tutorial_4_source_code import (
     src as htf,
 )
 
