### 2020-10-20 b106649
refactor of jacobian to take variables optionally
