+++ autofit/graphical/factor_graphs/numerical.py
@@ -1,33 +1,16 @@
-from typing import Tuple, Dict, NamedTuple
+from typing import Tuple, Dict, NamedTuple, Optional
 
 import numpy as np
 
 from autofit.mapper.variable import Variable
 
-
-class FactorValue(NamedTuple):
-    """
-    The return value associated with a factor
-    """
-    log_value: np.ndarray
-    deterministic_values: Dict[Variable, np.ndarray]
-
-
-class JacobianValue(NamedTuple):
-    log_value: Dict[Variable, np.ndarray]
-    deterministic_values: Dict[Tuple[Variable, Variable], np.ndarray]
-
-
-HessianValue = Dict[Variable, np.ndarray]
-
-
 def numerical_func_jacobian(
-        factor,
-        args: Tuple[Variable, ...],
-        kwargs: Dict[Variable, np.array],
+        factor: "AbstractNode",
+        values: Dict[Variable, np.array],
+        variables: Optional[Tuple[Variable, ...]],
         _eps: float = 1e-6,
         _calc_deterministic: bool = True,
-) -> JacobianValue:
+) -> Tuple["FactorValue", "JacobianValue"]:
     """Calculates the numerical Jacobian of the passed factor
 
     the arguments passed correspond to the variable that we want
@@ -58,27 +41,29 @@ def numerical_func_jacobian(
         log_value={'x': array([[0.], [0.]])},
         deterministic_values={('x', 'y'): array([[0., 2.], [1., 3.]])})
     """
+    if variables is None:
+        variables = factor.variables
     # copy the input array
-    p0 = {v: np.array(x, dtype=float) for v, x in kwargs.items()}
+    p0 = {v: np.array(x, dtype=float) for v, x in values.items()}
     f0 = factor(p0)
     log_f0 = f0.log_value
     det_vars0 = f0.deterministic_values
 
     jac_f = {
-        v: np.empty(np.shape(kwargs[v]) + np.shape(log_f0))
-        for v in args}
+        v: np.empty(np.shape(values[v]) + np.shape(log_f0))
+        for v in variables}
     if _calc_deterministic:
         jac_det = {
             (det, v): np.empty(
-                np.shape(val) + np.shape(kwargs[v]))
-            for v in args
+                np.shape(val) + np.shape(values[v]))
+            for v in variables
             for det, val in det_vars0.items()}
         det_slices = {
-            v: (slice(None),) * np.ndim(a) for v, a in kwargs.items()}
+            v: (slice(None),) * np.ndim(a) for v, a in values.items()}
     else:
         jac_det = {}
 
-    for v in args:
+    for v in variables:
         x0 = p0[v]
         if x0.shape:
             inds = tuple(a.ravel() for a in np.indices(x0.shape))
@@ -108,50 +93,42 @@ def numerical_func_jacobian(
     return f0, JacobianValue(jac_f, jac_det)
 
 
-def numerical_jacobian(
-        factor,
-        args: Tuple[Variable, ...],
-        kwargs: Dict[Variable, np.array],
-        _eps: float = 1e-6,
-        _calc_deterministic: bool = True,
-) -> 'FactorValue':
-    return numerical_func_jacobian(
-        factor, args, kwargs, _eps, _calc_deterministic)[1]
-
-
 def numerical_func_jacobian_hessian(
-        factor,
-        args: Tuple[Variable, ...],
-        kwargs: Dict[Variable, np.array],
+        factor: "AbstractNode",
+        values: Dict[Variable, np.array],
+        variables: Optional[Tuple[Variable, ...]] = None,
         _eps: float = 1e-6,
         _calc_deterministic: bool = True
[…]
