+++ autofit/graphical/factor_graphs/__init__.py
@@ -1,5 +1,4 @@
 from autofit.graphical.factor_graphs.abstract import AbstractNode
 from autofit.graphical.factor_graphs.factor import FactorValue, Factor, AbstractFactor
 from autofit.graphical.factor_graphs.graph import DeterministicFactorNode, FactorGraph
-from autofit.graphical.factor_graphs.numerical import numerical_jacobian
 from autofit.mapper.variable import Variable, Plate
