### 2020-10-20 974d7a8
some bugfixes for MeanField
