### 2020-10-22 c7dac22
cleaned up mock configs in tests
