### 2020-10-24 f7b168a
removed skip completed option
