+++ test_autofit/unit/non_linear/optimize/test_abstract_optimize.py
@@ -5,7 +5,7 @@ from autoconf import conf
 import autofit as af
 
 from autofit.non_linear.samples import OptimizerSamples
-from autofit.mock import MockClassx4
+from autofit.mock.mock import MockClassx4
 
 directory = os.path.dirname(os.path.realpath(__file__))
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
