+++ test_autofit/unit/mapper/test_from_data_names.py
@@ -1,7 +1,7 @@
 import pytest
 
 import autofit as af
-from autofit import mock
+from autofit.mock import mock
 
 names = ["one", "two", "three"]
 
