+++ test_autofit/unit/mapper/test_assertion.py
@@ -2,7 +2,7 @@ import pytest
 
 import autofit as af
 from autofit import exc
-from autofit import mock
+from autofit.mock import mock
 from autofit.mapper.prior.assertion import CompoundAssertion
 
 
