+++ test_autofit/unit/non_linear/nest/test_multi_nest.py
@@ -6,7 +6,7 @@ import pytest
 
 import autofit as af
 from autoconf import conf
-from autofit import mock
+from autofit.mock import mock
 from autofit.non_linear.nest import multi_nest as mn
 
 directory = os.path.dirname(os.path.realpath(__file__))
