+++ test_autofit/unit/mapper/model/test_prior_model.py
@@ -3,8 +3,8 @@ import copy
 import pytest
 
 import autofit as af
-from autofit import mock
-from autofit import mock_real
+from autofit.mock import mock
+from autofit.mock import mock_real
 
 
 @pytest.fixture(name="instance_prior_model")
