+++ autofit/config/notation/label.ini
@@ -2,7 +2,13 @@
 parameter0=a
 parameter1=b
 parameter2=c
+centre=x
+intensity=I
+sigma=sigma
+rate=\lambda
 
 [subscript]
 ModelComponent0=M0
 ModelComponent1=M1
+Gaussian=g
+Exponential=e
