+++ citation.tex
@@ -12,7 +12,7 @@ This work uses the following software packages:
 
 \item
 \href{https://github.com/dfm/corner.py}{\textt{corner.py}}
-\citep{corner2016}
+\citep{corner}
 
 \item
 \href{https://github.com/joshspeagle/dynesty}{\textt{dynesty}}
