+++ autofit/config/non_linear/nest/MultiNest.ini
@@ -12,7 +12,7 @@ max_iter=0
 [settings]
 n_iter_before_update=5
 null_log_evidence=-1e90
-seed=-1
+seed=0
 verbose=False
 resume=True
 context=0
