+++ autofit/config/non_linear/nest/MultiNest.ini
@@ -29,6 +29,7 @@ visualize_every_update=1
 model_results_every_update=1
 log_every_update=1
 remove_state_files_at_end=True
+should_update_sym=250
 
 [initialize]
 method=prior
