+++ howtofit/chapter_1_introduction/scripts/tutorial_8_aggregator.py
@@ -1,0 +1,208 @@
+# %%
+"""
+Tutorial 3: Aggregator
+======================
+
+In the previous tutorial, we fitted 3 datasets with an identical `NonLinearSearch`, outputting the results of each to a
+unique folder on our hard disk.
+
+In this tutorial, we'll use the `Aggregator` to load the `Result`'s and manipulate / plot them using our Jupyter
+notebook. The API for using `Result`'s follow closely tutorial 1 of this chapter.
+"""
+
+# %%
+#%matplotlib inline
+
+from pyprojroot import here
+
+workspace_path = str(here())
+%cd $workspace_path
+print(f"Working Directory has been set to `{workspace_path}`")
+
+import autofit as af
+
+# %%
+"""
+To load the results of the previous tutorial into the `Aggregator`, we simply point the `Aggregator` class to the path 
+of the results we want it to load.
+"""
+
+# %%
+agg = af.Aggregator(directory="output/howtofit/chapter_1/aggregator")
+
+# %%
+"""
+To begin, let me quickly explain what a generator is in Python, for those unaware. A generator is an object that 
+iterates over a function when it is called. The `Aggregator` creates all objects as generators, rather than lists, or 
+dictionaries, or whatever.
+
+Why? Because lists and dictionaries store every entry in memory simultaneously. If you fit many `Dataset`s, you`ll 
+have lots of results and therefore use a lot of memory. This will crash your laptop! On the other hand, a generator 
+only stores the object in memory when it runs the function; it is free to overwrite it afterwards. Thus, your laptop 
+won't crash!
+
+There are two things to bare in mind with generators:
+
+ 1) A generator has no length, thus to determine how many entries of data it corresponds to you first must turn it to a 
+    list.
+
+ 2) Once we use a generator, we cannot use it again and we'll need to remake it. For this reason, we typically avoid 
+    storing the generator as a variable and instead use the `Aggregator` to create them on use.
+
+We can now create a `Samples` generator of every fit. This creates instances of the `Samples` class we manipulated in
+tutorial 1, which with the `Aggregator` now acts as an interface between the results of the non-linear fit on your 
+hard-disk and Python.
+"""
+
+# %%
+samples_gen = agg.values("samples")
+
+# %%
+"""
+When we print this list of outputs you should see over 3 different `NestSamples` instances, corresponding to the 3
+model-fits we performed in the previous tutorial.
+"""
+
+# %%
+print("Emcee Samples:\n")
+print(samples_gen)
+print("Total Samples Objects = ", len(list(samples_gen)), "\n")
+
+
+# %%
+"""
+We've encountered the `Samples` class in previous tutorials. As we saw in tutorial 1, the `Samples` class contains all 
+the accepted parameter samples of the `NonLinearSearch`, which is a list of lists where:
+
+ - The outer list is the size of the total number of samples.
+ - The inner list is the size of the number of free parameters in the fit.
+
+With the `Aggregator` we can now get information on the `Samples` of all 3 model-fits, as opposed to just 1 fit using 
+its `Result` object.
+"""
+
+# %%
+for samples in agg.values("samples"):
+    print("All parameters of the very first sample")
+    print(samples.parameters[0])
+    print("The tenth sample`s third parameter")
+    print(samples.parameters[9][2])
+    print()
+
+# %%
+"""
[…]
