+++ howtofit/chapter_1_introduction/tutorial_6_complex_models/src/fit/fit.py
@@ -1,214 +1,0 @@
-import numpy as np
-
-"""The `fit.py` module is unchanged from the previous tutorial."""
-
-
-class FitDataset:
-
-    # noinspection PyUnresolvedReferences
-    def __init__(self, masked_dataset, model_data):
-        """Class to fit a `Dataset` with model data.
-
-        Parameters
-        -----------
-        masked_data : np.ndarray
-            The observed `MaskedDataset` that is fitted.
-        model_data : np.ndarray
-            The model data the data is fitted with.
-
-        Attributes
-        -----------
-        residual_map : np.ndarray
-            The residual-map of the fit (data - model_data).
-        chi_squared_map : np.ndarray
-            The chi-squared-map of the fit ((data - model_data) / noise_maps ) **2.0
-        chi_squared : float
-            The overall chi-squared of the model`s fit to the dataset, summed over every data point.
-        reduced_chi_squared : float
-            The reduced chi-squared of the model`s fit to simulate (chi_squared / number of datas points), summed over \
-            every data point.
-        noise_normalization : float
-            The overall normalization term of the noise_map, summed over every data point.
-        log_likelihood : float
-            The overall log likelihood of the model`s fit to the dataset, summed over evey data point.
-        """
-        self.masked_dataset = masked_dataset
-        self.model_data = model_data
-
-    @property
-    def xvalues(self):
-        return self.masked_dataset.dataset.xvalues
-
-    @property
-    def mask(self):
-        return self.masked_dataset.mask
-
-    @property
-    def data(self):
-        return self.masked_dataset.data
-
-    @property
-    def noise_map(self):
-        return self.masked_dataset.noise_map
-
-    @property
-    def residual_map(self):
-        return residual_map_from_data_model_data_and_mask(
-            data=self.data, model_data=self.model_data, mask=self.mask
-        )
-
-    @property
-    def normalized_residual_map(self):
-        return normalized_residual_map_from_residual_map_noise_map_and_mask(
-            residual_map=self.residual_map, noise_map=self.noise_map, mask=self.mask
-        )
-
-    @property
-    def chi_squared_map(self):
-        return chi_squared_map_from_residual_map_noise_map_and_mask(
-            residual_map=self.residual_map, noise_map=self.noise_map, mask=self.mask
-        )
-
-    @property
-    def signal_to_noise_map(self):
-        """The signal-to-noise_map of the `Dataset` and noise-map which are fitted."""
-        signal_to_noise_map = np.divide(self.data, self.noise_map)
-        signal_to_noise_map[signal_to_noise_map < 0] = 0
-        return signal_to_noise_map
-
-    @property
-    def chi_squared(self):
-        return chi_squared_from_chi_squared_map_and_mask(
-            chi_squared_map=self.chi_squared_map, mask=self.mask
-        )
-
-    @property
-    def noise_normalization(self):
-        return noise_normalization_from_noise_map_and_mask(
-            noise_map=self.noise_map, mask=self.mask
-        )
-
-    @property
-    def log_likelihood(self):
-        return likelihood_from_chi_squared_and_noise_normalization(
-            chi_squared=self.chi_squared, noise_normalization=self.noise_normalization
-        )
-
-
-def residual_map_from_data_model_data_and_mask(data, mask, model_data):
-    """
-    Returns the residual-map between a `MaskedDataset` and model data, where:
-
-    Residuals = (Data - Model_Data).
-
-    Parameters
-    -----------
-    data : np.ndarray
-        The observed data that is fitted.
-    mask : np.ndarray
[…]
