+++ howtofit/chapter_1_introduction/tutorial_4_source_code/src/plot/line_plots.py
@@ -1,27 +1,0 @@
-import matplotlib.pyplot as plt
-
-"""This function plots a line of 1D data, and was already introduce in tutorial 2."""
-
-
-def line(xvalues, line, title=None, color="k", errors=None, ylabel=None):
-    """Plot a 1D line of data on a plot of x versus y, where the x-axis is the x coordinate of the line and the y-axis
-    is the intensity of the line at that coordinate.
-
-    Parameters
-    ----------
-    xvalues : np.ndarray
-        The x-coordinates the profile is defined on.
-    line : np.ndarray
-        The intensity values of the profile which are plotted.
-    ylabel : str
-        The y-label of the plot.
-    """
-
-    plt.errorbar(
-        x=xvalues, y=line, yerr=errors, color=color, ecolor="k", elinewidth=1, capsize=2
-    )
-    plt.title(title)
-    plt.xlabel("x value of profile")
-    plt.ylabel(ylabel)
-    plt.show()
-    plt.clf()
