+++ howtofit/chapter_1_introduction/scripts/tutorial_9_filtering.py
@@ -16,7 +16,7 @@ the `Aggregator`'s `filter` tool, which filters the results and provides us with
 from pyprojroot import here
 
 workspace_path = str(here())
-%cd $workspace_path
+#%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
