+++ howtofit/chapter_1_introduction/scripts/tutorial_1_model_mapping.py
@@ -36,7 +36,7 @@ locate the path to the workspace on your computer and use it to set the working 
 from pyprojroot import here
 
 workspace_path = str(here())
-%cd $workspace_path
+#%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 # %%
