+++ howtofit/chapter_1_introduction/scripts/tutorial_2_model_fitting.py
@@ -12,7 +12,7 @@ In this tutorial, we'll fit the 1D `Gaussian` model from the previous tutorial t
 from pyprojroot import here
 
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
