+++ howtofit/chapter_1_introduction/scripts/tutorial_10_data_and_models.py
@@ -15,7 +15,7 @@ inspect, interpret and plot the results of the model-fit, including refitting th
 from pyprojroot import here
 
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
