+++ howtofit/chapter_1_introduction/scripts/tutorial_8_aggregator.py
@@ -16,7 +16,7 @@ notebook. The API for using `Result`'s follow closely tutorial 1 of this chapter
 from pyprojroot import here
 
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
