+++ howtofit/chapter_1_introduction/scripts/tutorial_4_visualization.py
@@ -13,7 +13,7 @@ output on-the-fly during the `NonLinearSearch`.
 from pyprojroot import here
 
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
