+++ howtofit/chapter_1_introduction/scripts/tutorial_7_fitting_multiple_datasets.py
@@ -16,7 +16,7 @@ we'll fit 3 different dataset`s, each with a single `Gaussian` model.
 from pyprojroot import here
 
 workspace_path = str(here())
-#%cd $workspace_path
+%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
 import autofit as af
