+++ docs/conf.py
@@ -22,6 +22,10 @@ os.system("git clone https://github.com/Jammy2211/autofit_workspace --depth 1")
 os.system(f"cp -r autofit_workspace/dataset {workspace_path}")
 os.system(f"cp -r autofit_workspace/howtofit/chapter_1_introduction/*.py {workspace_path}")
 os.system(f"cp -r autofit_workspace/howtofit/chapter_phase_api/src {workspace_path}")
+os.system(f"mkdir {workspace_path}/howtofit")
+os.system(f"mkdir {workspace_path}/howtofit/chapter_phase_api")
+os.system(f"mkdir {workspace_path}/howtofit/chapter_phase_api/src")
+os.system(f"cp -r autofit_workspace/howtofit/chapter_phase_api/src/config {workspace_path}/howtofit/chapter_phase_api/src")
 os.system("rm -rf autofit_workspace")
 
 os.system("git clone https://github.com/Jammy2211/autolens_files --depth 1")
