### 2020-11-04 a4cdfc2
global prior model
