+++ test_autofit/unit/graphical/gaussian/test_global_fit.py
@@ -5,18 +5,13 @@ import autofit.graphical as g
 
 
 @pytest.fixture(
-    name="model"
+    name="model_factor"
 )
-def make_model():
-    return af.Collection(
+def make_model_factor():
+    model = af.Collection(
         one=af.UniformPrior()
     )
 
-
-@pytest.fixture(
-    name="model_factor"
-)
-def make_model_factor(model):
     def likelihood_function(
             instance
     ):
@@ -39,12 +34,11 @@ class TestGlobalLikelihood:
     def test_single_factor(
             self,
             model_factor,
-            model,
             unit_value,
             likelihood
     ):
         assert model_factor.global_likelihood(
-            model.instance_from_unit_vector(
+            model_factor.global_prior_model.instance_from_unit_vector(
                 [unit_value]
             )
         ) == likelihood
@@ -59,7 +53,6 @@ class TestGlobalLikelihood:
     def test_collection(
             self,
             model_factor,
-            model,
             unit_value,
             likelihood
     ):
@@ -68,7 +61,45 @@ class TestGlobalLikelihood:
             model_factor
         )
         assert collection.global_likelihood(
-            model.instance_from_unit_vector(
+            collection.global_prior_model.instance_from_unit_vector(
                 [unit_value]
+            )
+        ) == likelihood
+
+    @pytest.mark.parametrize(
+        "unit_vector, likelihood",
+        [
+            ([0.5, 0.0], 0.0),
+            ([1.0, 0.5], -0.0625)
+        ]
+    )
+    def test_two_factor(
+            self,
+            model_factor,
+            unit_vector,
+            likelihood
+    ):
+        model_2 = af.Collection(
+            one=af.UniformPrior()
+        )
+
+        def likelihood_function(
+                instance
+        ):
+            return -(instance.one - 0.0) ** 2
+
+        model_factor_2 = g.ModelFactor(
+            model_2,
+            likelihood_function
+        )
+
+        collection = g.ModelFactorCollection(
+            model_factor,
+            model_factor_2
+        )
+
+        assert collection.global_likelihood(
+            collection.global_prior_model.instance_from_unit_vector(
+                unit_vector
             )
         ) == likelihood
