+++ autofit/graphical/declarative.py
@@ -3,6 +3,7 @@ from typing import Callable, cast, Set, List, Dict
 
 import numpy as np
 
+from autofit import ModelInstance
 from autofit.graphical.factor_graphs.factor import Factor
 from autofit.graphical.factor_graphs.graph import FactorGraph
 from autofit.graphical.mean_field import MeanFieldApproximation
@@ -126,7 +127,23 @@ class AbstractModelFactor(ABC):
             arguments
         )
 
-    def global_likelihood(self, instance):
+    def global_likelihood(
+            self,
+            instance: ModelInstance
+    ) -> float:
+        """
+        Compute the combined likelihood of each factor from a collection of instances
+        with the same ordering as the factors.
+
+        Parameters
+        ----------
+        instance
+            A collection of instances, one corresponding to each factor
+
+        Returns
+        -------
+        The combined likelihood of all factors
+        """
         likelihood = abs(
             self.model_factors[0].likelihood_function(
                 instance[0]
@@ -144,7 +161,10 @@ class AbstractModelFactor(ABC):
         return -likelihood
 
     @property
-    def global_prior_model(self):
+    def global_prior_model(self) -> CollectionPriorModel:
+        """
+        A collection of prior models, with one model for each factor.
+        """
         return CollectionPriorModel([
             model_factor.prior_model
             for model_factor
