### 2020-11-04 665eccf
fixes to make tests pass
