+++ test_autofit/unit/graphical/gaussian/test_declarative.py
@@ -3,8 +3,7 @@ import pytest
 
 import autofit as af
 import autofit.graphical as ep
-from test_autofit.unit.graphical.gaussian.model import make_data, _likelihood
-from autofit.mock.mock import Gaussian
+from test_autofit.unit.graphical.gaussian.model import Gaussian, make_data, _likelihood
 
 
 def test_shared_intensity():
