### 2020-11-04 16223f2
some bugfixes to ensure that MeanField is passed around correctly
