+++ autofit/non_linear/grid_search.py
@@ -576,6 +576,6 @@ def make_lists(no_dimensions, step_size, centre_steps=True):
     sub_lists = make_lists(no_dimensions - 1, step_size, centre_steps=centre_steps)
     return [
         [step_size * value + (0.5 * step_size if centre_steps else 0)] + sub_list
-        for value in range(0, int((1 / step_size)))
+        for value in range(int((1 / step_size)))
         for sub_list in sub_lists
     ]
