### 2020-11-04 eb41d5b
extracting multiprocessing code
