### 2020-11-04 b4bc378
Fixing test_logistic_regression and sampling
