### 2020-11-05 3b7da12
adding gradients to messages and some refactoring
