+++ autofit/graphical/expectation_propagration.py
@@ -1,0 +1,229 @@
+
+from abc import ABC, abstractmethod
+from typing import (
+    Dict, Tuple, Optional, NamedTuple, Iterator, List,
+    Callable
+)
+
+import numpy as np
+
+from autofit.graphical.messages.abstract import AbstractMessage
+
+from autofit.graphical.factor_graphs import (
+    Factor, AbstractNode, FactorGraph)
+from autofit.mapper.variable import Variable
+from autofit.graphical.utils import Status
+from autofit.graphical.mean_field import MeanField, FactorApproximation 
+
+
+class EPMeanField(AbstractNode):
+    '''
+    '''
+    def __init__(
+            self,
+            factor_graph: FactorGraph,
+            factor_mean_field: Dict[Factor, MeanField]
+    ):
+        self._factor_graph = factor_graph
+        self._factor_mean_field = factor_mean_field
+        variable_factor = {}
+        for factor, vs in factor_graph.factor_all_variables.items():
+            for v in vs:
+                variable_factor.setdefault(v, set()).add(factor)
+        self._variable_factor = variable_factor
+        
+        super().__init__(**self.factor_graph._kwargs)
+
+    @property 
+    def name(self):
+        return f"EP_{self.factor_graph.name}"
+
+    @property
+    def variables(self): 
+        return self.factor_graph.variables
+
+    @property
+    def deterministic_variables(self): 
+        return self.factor_graph.deterministic_variables
+
+    @property
+    def variable_names(self) -> Dict[str, Variable]: 
+        return self.factor_graph.variable_names
+ 
+    @property
+    def factor_mean_field(self) -> Dict[Factor, MeanField]:
+        return self._factor_mean_field.copy()
+
+    @property
+    def factor_graph(self) -> FactorGraph:
+        return self._factor_graph
+
+    @classmethod
+    def from_approx_dists(
+            cls,
+            factor_graph: FactorGraph,
+            approx_dists: Dict[Variable, AbstractMessage],
+    ) -> "EPMeanField":
+        factor_mean_field = {
+            factor: MeanField({
+                v: approx_dists[v] for v in factor.variables})
+            for factor in factor_graph.factors}
+
+        return cls(
+            factor_graph,
+            factor_mean_field)
+
+    from_kws = from_approx_dists
+    
+    def factor_approximation(self, factor: Factor) -> FactorApproximation:
+        factor_mean_field = self.factor_mean_field
+        factor_dist = factor_mean_field.pop(factor)
+        cavity_dist = MeanField.prod(
+            {v: 1. for v in factor_dist.all_variables},
+            *(dist for fac, dist in factor_mean_field.items()))
+        # cavity_dist.log_norm = 0.
+        model_dist = factor_dist.prod(cavity_dist)
+
+        return FactorApproximation(
+            factor, cavity_dist, factor_dist, model_dist)
+
+    def project_factor_approx(
+        self, projection: FactorApproximation, status: Optional[Status] = None,
+    ) -> "EPMeanField":
+        """
+        """
+        factor_mean_field = self.factor_mean_field
+        factor_mean_field[projection.factor] = projection.factor_dist
+
+        new_approx = type(self)(
+            factor_graph=self._factor_graph,
+            factor_mean_field=factor_mean_field)
+        return new_approx, status
+
+    project = project_factor_approx
+
+    @property
+    def mean_field(self) -> MeanField:
+        return MeanField.prod(
+            {v: 1. for v in self.all_variables},
+            *self._factor_mean_field.values())
+
+    @property
+    def variable_factor_message(self
+    ) -> Dict[Variable, Dict[Factor, AbstractMessage]]:
+        variable_factor_message = {
[…]
