### 2020-11-05 ed0ca2d
adding tests to check gradients and hessians of Messages
