### 2020-11-05 2a068a3
refactored AbstractNode to not allow kwargs
