### 2020-11-05 edd87ba
adding axis keyword to factors
