### 2020-11-06 c502135
some bugfixes for gradients and FactorGraph calls
