### 2020-11-06 cce054d
minor refactoring
