+++ autofit/graphical/utils.py
@@ -160,6 +160,24 @@ def prod(iterable: Iterable[_M], *arg: Tuple[_M]) -> _M:
     """
     return reduce(mul, iterable, *arg)
 
+class CachedProperty(object):
+    ''' 
+    A property that is only computed once per instance and then replaces
+    itself with an ordinary attribute. Deleting the attribute resets the
+    property. 
+    
+    Source: https://github.com/bottlepy/bottle/commit/fa7733e075da0d790d809aa3d2f53071897e6f76
+    '''
+
+    def __init__(self, func):
+        self.func = func
+
+    def __get__(self, obj, cls):
+        if obj is None: return self
+        value = obj.__dict__[self.func.__name__] = self.func(obj)
+        return value
+
+cached_property = CachedProperty
 
 def propagate_uncertainty(
         cov: np.ndarray, jac: np.ndarray) -> np.ndarray:
