### 2020-11-06 71f6cbe
bugfixing axis call in numerical.py
