+++ autofit/graphical/factor_graphs/numerical.py
@@ -5,13 +5,13 @@ import numpy as np
 
 from autofit.mapper.variable import Variable
 from autofit.graphical.utils import \
-    FactorValue, JacobianValue, HessianValue, aggregate
+    FactorValue, JacobianValue, HessianValue, aggregate, Axis
 
 def numerical_func_jacobian(
         factor: "AbstractNode",
         values: Dict[Variable, np.array],
         variables: Optional[Tuple[Variable, ...]] = None,
-        axis: Optional[Union[bool, int, Tuple[int, ...]]] = False, 
+        axis: Axis = False, 
         _eps: float = 1e-6,
         _calc_deterministic: bool = True,
 ) -> Tuple[FactorValue, JacobianValue]:
@@ -48,12 +48,11 @@ def numerical_func_jacobian(
     if variables is None:
         variables = factor.variables
 
-    agg = partial(aggregate, axis=axis)
 
     # copy the input array
     p0 = {v: np.array(x, dtype=float) for v, x in values.items()}
-    f0 = factor(p0)
-    log_f0 = agg(f0.log_value)
+    f0 = factor(p0, axis=axis)
+    log_f0 = f0.log_value
     det_vars0 = f0.deterministic_values
 
     jac_f = {
@@ -77,10 +76,10 @@ def numerical_func_jacobian(
             for ind in zip(*inds):
                 x0[ind] += _eps
                 p0[v] = x0
-                f = factor(p0)
+                f = factor(p0, axis=axis)
                 x0[ind] -= _eps
 
-                jac_f[v][ind] = (agg(f.log_value) - log_f0) / _eps
+                jac_f[v][ind] = (f.log_value - log_f0) / _eps
                 if _calc_deterministic:
                     det_vars = f.deterministic_values
                     for det, val in det_vars.items():
@@ -88,7 +87,7 @@ def numerical_func_jacobian(
                             (val - det_vars0[det]) / _eps
         else:
             p0[v] += _eps
-            f = factor(p0)
+            f = factor(p0, axis=axis)
             p0[v] -= _eps
 
             jac_f[v] = (f.log_value - log_f0) / _eps
@@ -112,13 +111,13 @@ def numerical_func_jacobian_hessian(
     if variables is None:
         variables = factor.variables
 
-    agg = partial(aggregate, axis=axis)
+    # agg = partial(aggregate, axis=axis)
 
     p0 = {v: np.array(x, dtype=float) for v, x in values.items()}
     f0, jac_f0 = factor.func_jacobian(p0, variables, axis=axis)
     (log_f0, det_vars0), (grad_f0, jac_det_vars0) = f0, jac_f0
 
-    log_f0 = agg(log_f0)
+    log_f0 = log_f0
     f_shape = np.shape(log_f0)
     f_size = np.prod(f_shape, dtype=int)
     hess_f = {
@@ -174,7 +173,7 @@ def numerical_func_jacobian_hessian(
         else:
             p0[v] += _eps
             grad_f, _ = factor.jacobian(
-                p0, (v,), _calc_deterministic=False)
+                p0, (v,), axis=axis, _eps=_eps, _calc_deterministic=False)
             p0[v] -= _eps
             hess_f[v] = grad_f[v] - grad_f0[v]
             
