### 2020-11-06 4ae9099
refactored numerical_jacobian
