+++ test_autofit/unit/graphical/test_messages.py
@@ -32,7 +32,6 @@ def test_message_norm():
             *m1._support[0])
 
         # verify within tolerance
-        print(ms)
         assert np.abs(A - i12) < ierr < 1e-6
 
 def test_numerical_gradient_hessians():
@@ -70,3 +69,33 @@ def test_meanfield_gradients():
         v3: graph.NormalMessage(
             np.random.randn(n3, n1),
             np.random.exponential(size=(n3, n1)))})
+
+    values = mean_field.sample()
+    l0, _ = mean_field(values, axis=None)
+    logl, _ = mean_field(values, axis=False)
+    assert logl.sum() == pytest.approx(l0, abs=1e-5)
+    logl, _ = mean_field(values, axis=1)
+    assert logl.sum() == pytest.approx(l0, abs=1e-5)
+    logl, _ = mean_field(values, axis=2)
+    assert logl.sum() == pytest.approx(l0, abs=1e-5)
+    logl, _ = mean_field(values, axis=(0, 2))
+    assert logl.sum() == pytest.approx(l0, abs=1e-5)
+
+    njac0 = mean_field._numerical_func_jacobian(
+        values, axis=None, _eps=1e-8)[1][0]
+    njac1 = mean_field._numerical_func_jacobian(
+        values, axis=1, _eps=1e-8)[1][0]
+    njac2 = mean_field._numerical_func_jacobian(
+        values, axis=(0, 1), _eps=1e-8)[1][0]
+    njac = mean_field._numerical_func_jacobian(
+        values, axis=False, _eps=1e-8)[1][0]
+    grad = mean_field.logpdf_gradient(values, axis=False)[1]
+    for v in grad:
+        norm = np.linalg.norm(grad[v] - njac[v].sum((0, 1, 2)))
+        assert norm == pytest.approx(0, abs=1e-2)
+        norm = np.linalg.norm(grad[v] - njac0[v])
+        assert norm == pytest.approx(0, abs=1e-2)
+        norm = np.linalg.norm(grad[v] - njac1[v].sum((0, 1)))
+        assert norm == pytest.approx(0, abs=1e-2)
+        norm = np.linalg.norm(grad[v] - njac2[v].sum(0))
+        assert norm == pytest.approx(0, abs=1e-2)
