### 2020-11-06 876c2eb
fixed gradients counting problem
