### 2020-11-06 48f1510
minor bugfixes
