+++ docs/general/future.rst
@@ -5,8 +5,7 @@ Future
 
 The following features are planned for 2021 - 2022:
 
- - **Message Passing** - Composing complex models from many model components using Bayes factors and fitting them
-within a message passing framework.
+- **Message Passing** - Composing complex models from many model components using Bayes factors and fitting them within a message passing framework.
 
 - **Time series modelling** - Fit temporally varying models using fits which marginalize over time.
 
