### 2020-11-09 54d75d1
adding tests for checking jacobians and fixing FactorApproximation.func_jacobian
