+++ test_autofit/unit/graphical/regression/test_logistic_regression.py
@@ -67,6 +67,14 @@ def make_model_approx(
         }
     )
 
+def test_jacobians(
+        model_approx
+):
+    for factor in model_approx.factor_graph.factors:
+        factor_approx = model_approx.factor_approximation(factor)
+        opt = mp.optimise.OptFactor.from_approx(factor_approx)
+        assert opt.numerically_verify_jacobian(
+            100, rtol=1e-2, atol=1e-2), factor    
 
 def test_laplace(
         model_approx,
