+++ autofit/graphical/factor_graphs/factor.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from autofit.graphical.utils import aggregate, Axis
 from autofit.graphical.factor_graphs.abstract import \
-    AbstractNode, FactorValue
+    AbstractNode, FactorValue, JacobianValue
 from autofit.mapper.variable import Variable
 
 
@@ -400,4 +400,206 @@ class Factor(AbstractFactor):
     def __repr__(self) -> str:
         args = ", ".join(chain(
             map("{0[0]}={0[1]}".format, self._kwargs.items())))
-        return f"Factor({self.name})({args})"
+        return f"Factor({self.name}, {args})"
+
+
+
+class FactorJacobian(Factor):
+    def __init__(
+            self,
+            factor_jacobian: Callable,
+            name=None,
+            vectorised=False,
+            variable_order=None, 
+            **kwargs: Variable
+    ):
+        """
+        A node in a graph representing a factor
+
+        Parameters
+        ----------
+        factor
+            A wrapper around some callable
+        args
+            Variables representing positional arguments for the function
+        kwargs
+            Variables representing keyword arguments for the function
+        """
+        self.vectorised = vectorised
+        self._factor_jacobian = factor_jacobian
+        AbstractFactor.__init__(
+            self, 
+            **kwargs,
+            name=name or factor_jacobian.__name__
+        )
+        self._variables = tuple(
+            self._kwargs.values() if variable_order is None else variable_order)
+
+
+    def __hash__(self) -> int:
+        # TODO: might this break factor repetition somewhere?
+        return hash(self._factor)
+
+    def _call_factor(
+            self,
+            values: Dict[str, np.ndarray],
+            variables: Optional[Tuple[str, ...]] = None, 
+    ) -> Union[np.ndarray, Tuple[np.ndarray, Tuple[np.ndarray, ...]]]:
+        """
+        Call the underlying function
+
+        Parameters
+        ----------
+        args
+            Positional arguments for the function
+        kwargs
+            Keyword arguments for the function
+
+        Returns
+        -------
+        Value returned by the factor
+        """
+        if self.vectorised:
+            return self._factor_jacobian(**values, _variables=variables)
+            
+        """Some factors may not be vectorised to broadcast over
+        multiple inputs
+
+        this method checks whether multiple input values have been
+        passed, and if so automatically loops over the inputs.
+        If any of the inputs have initial dimension one, it repeats
+        that value to match the length of the other inputs
+
+        If the other inputs do not match then it raises ValueError
+        """
+        kwargs_dims = {k: np.ndim(a) for k, a in values.items()}
+        # Check dimensions of inputs directly match plates
+        direct_call = (
+            all(dim == kwargs_dims[k] for k, dim in self._kwargs_dims.items()))
+        if direct_call:
+            return self._factor_jacobian(**values, _variables=variables)
+
+        # Check dimensions of inputs match plates + 1
+        vectorised = (
+            all(dim + 1 == kwargs_dims[k]
+                for k, dim in self._kwargs_dims.items()))
+
+        if not vectorised:
+            raise ValueError(
+                "input dimensions do not match required dims"
+                f"input: **kwargs={kwargs_dims}"
+                f"required: "
+                f"**kwargs={self._kwargs_dims}")
+
+        kw_lens = {k: len(a) for k, a in values.items()}
+
+        # checking 1st dimensions match
+        sizes = set(kw_lens.values())
+        dim0 = max(sizes)
+        if sizes.difference({1, dim0}):
+            raise ValueError(
+                f"size mismatch first dimensions passed: {sizes}")
+
+        iter_kws = {
+            k: iter(a) if kw_lens[k] == dim0 else iter(repeat(a[0]))
+            for k, a in values.items()}
+
[…]
