+++ autofit/graphical/factor_graphs/__init__.py
@@ -1,4 +1,6 @@
 from autofit.graphical.factor_graphs.abstract import AbstractNode
-from autofit.graphical.factor_graphs.factor import FactorValue, Factor, AbstractFactor
-from autofit.graphical.factor_graphs.graph import DeterministicFactorNode, FactorGraph
+from autofit.graphical.factor_graphs.factor import \
+    FactorValue, Factor, FactorJacobian, AbstractFactor
+from autofit.graphical.factor_graphs.graph import \
+    DeterministicFactorNode, FactorGraph
 from autofit.mapper.variable import Variable, Plate
