### 2020-11-10 6f3dcc9
adding test_factor_jacobian
