### 2020-11-11 67ebfe3
some bugfixes for DeterministicFactor
