### 2020-11-11 6014f25
setting up search tree
