### 2020-11-11 df0ca19
results and consistency
