+++ autofit/sensitivity.py
@@ -1,6 +1,9 @@
-from typing import List, Generator, Callable
+from abc import ABC
+from typing import List, Generator, Callable, Type
 
-from autofit import AbstractPriorModel, ModelInstance, Paths, CollectionPriorModel, Result
+import numpy as np
+
+from autofit import AbstractPriorModel, ModelInstance, Paths, CollectionPriorModel, Result, Analysis, NonLinearSearch
 from autofit.non_linear.parallel import AbstractJob, Process
 from .non_linear.grid_search import make_lists
 
@@ -11,6 +14,14 @@ class JobResult:
             result: Result,
             perturbed_result: Result
     ):
+        """
+        The result of a single sensitivity comparison
+
+        Parameters
+        ----------
+        result
+        perturbed_result
+        """
         self.result = result
         self.perturbed_result = perturbed_result
 
@@ -18,17 +29,30 @@ class JobResult:
 class Job(AbstractJob):
     def __init__(
             self,
-            image,
-            model,
-            perturbation_model,
-            analysis_class,
-            search
+            analysis: Analysis,
+            model: AbstractPriorModel,
+            perturbation_model: AbstractPriorModel,
+            search: NonLinearSearch
     ):
-        self.image = image
+        """
+        Job to run non-linear searches comparing how well a model and a model with a perturbation
+        fit the image.
+
+        Parameters
+        ----------
+        model
+            A base model that fits the image without a perturbation
+        perturbation_model
+            A model of the perturbation which has been added to the underlying image
+        analysis
+            A class definition which can compares instances of a model to a perturbed image
+        search
+            A non-linear search
+        """
+        self.analysis = analysis
         self.model = model
 
         self.perturbation_model = perturbation_model
-        self.analysis_class = analysis_class
 
         paths = search.paths
 
@@ -42,13 +66,15 @@ class Job(AbstractJob):
             )
         )
 
-    @property
-    def analysis(self):
-        return self.analysis_class(
-            self.image
-        )
+    def perform(self) -> JobResult:
+        """
+        - Create one model with a perturbation and another without
+        - Fit each model against the perturbed image
 
-    def perform(self):
+        Returns
+        -------
+        An object comprising the results of the two fits
+        """
         model = CollectionPriorModel()
         model.model = self.model
         result = self.search.fit(
@@ -70,18 +96,58 @@ class Job(AbstractJob):
         )
 
 
+class ImageAnalysis(Analysis, ABC):
+    def __init__(self, image: np.array):
+        self.image = image
+
+
 class Sensitivity:
     def __init__(
             self,
             instance,
-            model,
-            search,
-            analysis_class,
+            model: AbstractPriorModel,
+            search: NonLinearSearch,
+            analysis_class: Type[ImageAnalysis],
             perturbation_model: AbstractPriorModel,
             image_function: Callable,
-            step_size=0.1,
-            number_of_cores=2
+            step_size: float = 0.1,
+            number_of_cores: int = 2
     ):
+        """
+        Perform sensitivity mapping to evaluate whether a perturbation
+        can be detected if it occurs in different parts of an image.
+
+        For a range from 0 to 1 with step_size, for each dimension of the
+        perturbation_model, a perturbation is created and used in conjunction
+        with the instance to create an image.
+
+        For each of these images, a fit is run with just the model and with both
+        the model and perturbation_model to compare how much better the image
+        can be fit if the perturbation is included.
+
+        Parameters
+        ----------
+        instance
+            An instance of a model to which perturbations are applied prior to
+            images being generated
+        model
+            A model that fits the instance well
+        search
+            A NonLinear search class which is copied and used to evaluate fitness
+        analysis_class
+            A class which can compare an image to an instance and evaluate fitness
+        perturbation_model
[…]
