### 2020-11-11 b02c318
extra assertion on global run
