### 2020-11-11 5c2138d
refactored FactorValue and JacobianValue
