### 2020-11-12 77d14e2
refactoring factor_graphs
