+++ autofit/graphical/factor_graphs/factor.py
@@ -4,12 +4,13 @@ from itertools import chain, repeat
 from typing import \
     Tuple, Dict, Union, Set, NamedTuple, Callable, Optional
 from functools import lru_cache
+
 import numpy as np
 
-
-from autofit.graphical.utils import aggregate, Axis
+from autofit.graphical.utils import \
+    aggregate, Axis, FactorValue, JacobianValue
 from autofit.graphical.factor_graphs.abstract import \
-    AbstractNode, FactorValue, JacobianValue
+    AbstractNode
 from autofit.mapper.variable import Variable
 
 
@@ -25,9 +26,6 @@ class AbstractFactor(AbstractNode, ABC):
 
     @property
     def deterministic_variables(self) -> Set[Variable]:
-        """
-        Dictionary mapping the names of deterministic variables to those variables
-        """
         return self._deterministic_variables
 
     @property
@@ -104,6 +102,32 @@ class AbstractFactor(AbstractNode, ABC):
     
 
 class Factor(AbstractFactor):
+    """
+    A node in a graph representing a factor with analytic evaluation 
+    of its Jacobian
+
+    Parameters
+    ----------
+    factor
+        the function being wrapped, must accept calls through keyword 
+        argument
+
+    name: optional, str
+        the name of the factor, if not passed then uses the name
+        of the function passed
+
+    vectorised: optional, bool
+        if true the factor will call it the function directly over multiple 
+        inputs. If false the factor will call the function iteratively over
+        each argument. 
+
+    is_scalar: optional, bool
+        if true the factor returns a scalar value. Note if multiple arguments
+        are passed then a vector will still be returned
+        
+    kwargs: Variables
+        Variables for each keyword argument for the function
+    """
     def __init__(
             self,
             factor: Callable,
@@ -174,7 +198,7 @@ class Factor(AbstractFactor):
             k: np.shape(x) for k, x in kwargs.items()}
         return self._var_shape(**var_shapes)
     
-    @lru_cache(maxsize=8)
+    @lru_cache
     def _var_shape(self, **kwargs: Tuple[int, ...]) -> Tuple[int, ...]:
         """This is called by _function_shape
         
@@ -256,12 +280,7 @@ class Factor(AbstractFactor):
 
         if self.vectorised:
             return self._factor(**kwargs)
-        return self._py_vec_call(**kwargs)
-
-    def _py_vec_call(
-            self,
-            **kwargs: np.ndarray
-    ) -> np.ndarray:
+            
         """Some factors may not be vectorised to broadcast over
         multiple inputs
 
@@ -368,8 +387,8 @@ class Factor(AbstractFactor):
             agg_func=np.sum
     ) -> np.ndarray:
         """
-        broadcasts the value of a variable to match the specific shape
-        of the factor
+        broadcasts `value` to match the specific shape of the factor,
+        where `value` has the shape of the factor
 
         if the number of dimensions passed of the variable is 1
         greater than the dimensions of the variable then it's assumed
@@ -404,8 +423,7 @@ class Factor(AbstractFactor):
             else:
                 return False
 
-        from autofit.graphical.factor_graphs import DeterministicFactorNode
-        return DeterministicFactorNode(
+        return DeterministicFactor(
             self._factor,
             other,
             **self._kwargs
@@ -417,214 +435,112 @@ class Factor(AbstractFactor):
         return f"Factor({self.name}, {args})"
 
 
+class DeterministicFactor(Factor):
+    """
+    A deterministic factor is used to convert a function f(g(x)) to f(y)g(x) (integrating over y wit
+    a delta function) so that it can be represented in a factor graph.
 
-class FactorJacobian(Factor):
+    Parameters
+    ----------
+    factor
[…]

[…]
