### 2020-11-12 c18bc98
removing accept_variable_dict
