### 2020-11-12 6f84993
adding some documentation and refactoring
