+++ autofit/graphical/mean_field.py
@@ -39,7 +39,7 @@ def project_on_to_factor_approx(
     factor_projection = {}
 #     log_norm = 0.
     for v, q_fit in model_dist.items():
-        q_cavity = factor_approx.all_cavity_dist.get(v)
+        q_cavity = factor_approx.cavity_dist.get(v)
         if isinstance(q_fit, FixedMessage):
             factor_projection[v] = q_fit
         elif q_fit.is_valid:
@@ -96,7 +96,30 @@ def project_on_to_factor_approx(
 
     
 class MeanField(Dict[Variable, AbstractMessage], Factor):
-    """
+    """For a factor with multiple variables, this class represents the 
+    the mean field approximation to that factor, 
+
+    f(x₁, x₂, x₃) = q(x₁, x₂, x₃) = q₁(x₁) q₂(x₂) q₃(x₃)
+
+    Internally these variables approximations are stored in a 
+    dictionary with the variables as keys and the message or 
+    variable distribution as values
+
+
+    Methods
+    -------
+    keys()
+        returns the variables of the mean_field
+
+    logpdf({x₁: x1, x₂: x2, x₃: x3})
+        returns the q(x₁, x₂, x₃), axis defines the axes over which
+        to reduce the return, if the meanfield is duplicated over multiple
+        plates. 
+
+    logpdf_gradient({x₁: x1, x₂: x2, x₃: x3})
+        returns the q(x₁, x₂, x₃) and the gradients for each input. 
+        to save memory the gradients are always the shape of the input
+        values (i.e. this does not calculate the Jacobian)
     """
     def __init__(
             self, 
@@ -230,7 +253,7 @@ class MeanField(Dict[Variable, AbstractMessage], Factor):
             covar: Dict[Variable, np.ndarray], 
             fun: Optional[float] = None):
         """
-        Projects mode and covar
+        Projects the mode and covariance 
         """
         projection = MeanField({
             v: dist.from_mode(mode[v], covar.get(v))
@@ -260,6 +283,54 @@ class _FactorApproximation(NamedTuple):
     model_dist: MeanField
 
 class FactorApproximation(_FactorApproximation):
+    """
+    This class represents the 'tilted distribution' in EP,
+
+    When approximating a model distribution of the form,
+
+    m(x) = ∏ₐ fₐ(xₐ)
+
+    we can define an approximating distribution as the product of 
+    factor distributions,
+
+    q(x) = ∏ₐ qₐ(xₐ)
+
+    the 'cavity distribution' q⁻ᵃ for a factor can be viewed as a 
+    prior distribution for the factor,
+
+    q⁻ᵃ(xₐ) = ∏_{ᵦ ≠ a} qᵦ(xᵦ)
+
+    so the model can be approximated by the 'tilted distribution'
+
+    q⁺ᵃ(xₐ) = fₐ(xₐ) q⁻ᵃ(xₐ)
+
+    Parameters
+    ----------
+    is_valid
+    factor: Factor
+        fₐ(xₐ)
+    cavity_dist: MeanField
+        q⁻ᵃ(xₐ)
+    factor_dist: MeanField
+        qₐ(xₐ)
+    model_dist: MeanField
+        q(xₐ)
+
+    Methods
+    -------
+    __call__(values={xₐ: x₀}, axis=axis)
+        returns q⁺ᵃ(x₀)
+
+    func_jacobian(values={xₐ: x₀}, variables=(xₐ,), axis=axis)
+        returns q⁺ᵃ(x₀), {xₐ: dq⁺ᵃ(x₀)/dxₐ}
+
+    project_mean_field(mean_field, delta=1., status=Status())
[…]
