### 2020-11-12 77764ef
refactoring location of FactorValue, JAcobianValue and HessianValue
