+++ autofit/graphical/factor_graphs/graph.py
@@ -5,11 +5,11 @@ from functools import reduce
 
 import numpy as np
 
-from autofit.graphical.factor_graphs.abstract import AbstractNode
+from autofit.graphical.factor_graphs.abstract import FactorValue, AbstractNode
 from autofit.graphical.factor_graphs.factor import Factor
 from autofit.mapper.variable import Variable, Plate
 from autofit.graphical.utils import \
-    add_arrays, aggregate, Axis, cached_property, FactorValue
+    add_arrays, aggregate, Axis, cached_property
 
 
 class FactorGraph(AbstractNode):
