+++ autofit/graphical/factor_graphs/transform.py
@@ -1,0 +1,278 @@
+
+
+from abc import ABC, abstractmethod
+from functools import wraps
+from typing import Dict, Tuple, Optional, List
+
+import numpy as np
+from scipy.linalg import cho_factor, solve_triangular, get_blas_funcs
+
+from autofit.graphical.factor_graphs import \
+    AbstractNode, Variable, FactorValue, JacobianValue, HessianValue
+from autofit.graphical.utils import cached_property, Axis
+
+class AbstractTransform(ABC):
+
+    @property 
+    @abstractmethod
+    def variables(self):
+        pass
+
+    @abstractmethod
+    def whiten(self, values: Dict[Variable, np.ndarray]):
+        pass
+    
+    @abstractmethod
+    def unwhiten(self, values: Dict[Variable, np.ndarray]):
+        pass
+
+    @abstractmethod
+    def whiten2d(self, values: Dict[Variable, np.ndarray]):
+        pass
+    
+    @abstractmethod
+    def unwhiten2d(self, values: Dict[Variable, np.ndarray]):
+        pass
+
+    @property
+    @abstractmethod
+    def log_det(self):
+        pass
+
+    @cached_property
+    def det(self):
+        return np.exp(self.log_det)
+
+
+class RescaleTransform(AbstractTransform):
+    def __init__(self, variables_scales: Dict[Variable, np.ndarray]):
+        self.scale = variables_scales
+        self.inv_scale = {v: scale**-1 for v, scale in self.scale.items()}
+
+    def whiten(
+        self, 
+        values: Dict[Variable, np.ndarray]
+    ) -> Dict[Variable, np.ndarray]:
+        return {
+            v: val * self.inv_scale[v] for v, val in values.items()}
+
+    def unwhiten(
+        self, 
+        values: Dict[Variable, np.ndarray]
+    ) -> Dict[Variable, np.ndarray]:
+        return {
+            v: val * self.scale[v] for v, val in values.items()}
+
+    def whiten2d(
+        self, 
+        values: Dict[Variable, np.ndarray]
+    ) -> Dict[Variable, np.ndarray]:
+        whitened = {}
+        for v, hess in values.items():
+            inv_scale = self.inv_scale[v]
+            shape = np.shape(inv_scale)
+            size = np.size(inv_scale)
+            hess2d = np.reshape(hess, (size, size))
+            inv_scale1d = inv_scale.ravel()
+            w_hess = hess2d * inv_scale1d[None, :] * inv_scale1d[:, None]
+            whitened[v] = w_hess.reshape(shape + shape)
+
+        return whitened
+
+    def unwhiten2d(
+        self, 
+        values: Dict[Variable, np.ndarray]
+    ) -> Dict[Variable, np.ndarray]:
+        unwhitened = {}
+        for v, hess in values.items():
+            scale = self.scale[v]
+            shape = np.shape(scale)
+            size = np.size(scale)
+            hess2d = np.reshape(hess, (size, size))
+            scale1d = scale.ravel()
+            w_hess = hess2d * scale1d[None, :] * scale1d[:, None]
+            unwhitened[v] = w_hess.reshape(shape + shape)
+
+        return unwhitened
+
+    @cached_property
+    def log_det(self):
+        return sum(np.log(scale).sum() for scale in self.scale.values())
+
+def _solve_triangular(c_and_lower, b, trans=False, overwrite_b=False):
+    c, lower = c_and_lower
+    return solve_triangular(
+        c, b, lower=lower, trans=trans, overwrite_b=overwrite_b)
+
+def _mul_triangular(c_and_lower, b, trans=False, overwrite_b=False):
[…]
