### 2020-11-12 f3a0c0f
refactor FlattenArrays to have flatten2d and unflatten2d and return views of arrays instead of copies
