+++ autofit/graphical/factor_graphs/transform.py
@@ -9,7 +9,7 @@ from scipy.linalg import cho_factor, solve_triangular, get_blas_funcs
 
 from autofit.graphical.factor_graphs import \
     AbstractNode, Variable, FactorValue, JacobianValue, HessianValue
-from autofit.graphical.utils import cached_property, Axis
+from autofit.graphical.utils import cached_property, Axis, FlattenArrays
 
 class AbstractTransform(ABC):
 
@@ -19,19 +19,19 @@ class AbstractTransform(ABC):
         pass
 
     @abstractmethod
-    def whiten(self, values: Dict[Variable, np.ndarray]):
+    def transform(self, values: Dict[Variable, np.ndarray]):
         pass
     
     @abstractmethod
-    def unwhiten(self, values: Dict[Variable, np.ndarray]):
+    def untransform(self, values: Dict[Variable, np.ndarray]):
         pass
 
     @abstractmethod
-    def whiten2d(self, values: Dict[Variable, np.ndarray]):
+    def transform2d(self, values: Dict[Variable, np.ndarray]):
         pass
     
     @abstractmethod
-    def unwhiten2d(self, values: Dict[Variable, np.ndarray]):
+    def untransform2d(self, values: Dict[Variable, np.ndarray]):
         pass
 
     @property
@@ -49,25 +49,29 @@ class RescaleTransform(AbstractTransform):
         self.scale = variables_scales
         self.inv_scale = {v: scale**-1 for v, scale in self.scale.items()}
 
-    def whiten(
+    @property
+    def variables(self):
+        return self.scale.keys()
+
+    def transform(
         self, 
         values: Dict[Variable, np.ndarray]
     ) -> Dict[Variable, np.ndarray]:
         return {
             v: val * self.inv_scale[v] for v, val in values.items()}
 
-    def unwhiten(
+    def untransform(
         self, 
         values: Dict[Variable, np.ndarray]
     ) -> Dict[Variable, np.ndarray]:
         return {
             v: val * self.scale[v] for v, val in values.items()}
 
-    def whiten2d(
+    def transform2d(
         self, 
         values: Dict[Variable, np.ndarray]
     ) -> Dict[Variable, np.ndarray]:
-        whitened = {}
+        transformed = {}
         for v, hess in values.items():
             inv_scale = self.inv_scale[v]
             shape = np.shape(inv_scale)
@@ -75,15 +79,15 @@ class RescaleTransform(AbstractTransform):
             hess2d = np.reshape(hess, (size, size))
             inv_scale1d = inv_scale.ravel()
             w_hess = hess2d * inv_scale1d[None, :] * inv_scale1d[:, None]
-            whitened[v] = w_hess.reshape(shape + shape)
+            transformed[v] = w_hess.reshape(shape + shape)
 
-        return whitened
+        return transformed
 
-    def unwhiten2d(
+    def untransform2d(
         self, 
         values: Dict[Variable, np.ndarray]
     ) -> Dict[Variable, np.ndarray]:
-        unwhitened = {}
+        actual_values = {}
         for v, hess in values.items():
             scale = self.scale[v]
             shape = np.shape(scale)
@@ -91,9 +95,9 @@ class RescaleTransform(AbstractTransform):
             hess2d = np.reshape(hess, (size, size))
             scale1d = scale.ravel()
             w_hess = hess2d * scale1d[None, :] * scale1d[:, None]
-            unwhitened[v] = w_hess.reshape(shape + shape)
+            actual_values[v] = w_hess.reshape(shape + shape)
 
-        return unwhitened
+        return actual_values
 
     @cached_property
     def log_det(self):
[…]
