+++ howtofit/chapter_1_introduction/scripts/tutorial_6_results_and_samples.py
@@ -30,9 +30,9 @@ Load the dataset from the `autofit_workspace/dataset` folder.
 """
 
 # %%
-dataset_path = "dataset/howtofit/chapter_1/gaussian_x1"
-data = af.util.numpy_array_from_json(file_path=f"{dataset_path}/data.json")
-noise_map = af.util.numpy_array_from_json(file_path=f"{dataset_path}/noise_map.json")
+dataset_path = path.join("dataset", "howtofit", "chapter_1","gaussian_x1")
+data = af.util.numpy_array_from_json(file_path=path.join(dataset_path, "data.json"))
+noise_map = af.util.numpy_array_from_json(file_path=path.join(dataset_path, "noise_map.json"))
 
 # %%
 """We'll reuse the `plot_line` and `Analysis` classes of the previous tutorial."""
