+++ autofit/graphical/factor_graphs/transform.py
@@ -8,122 +8,90 @@ import numpy as np
 from scipy.linalg import cho_factor, solve_triangular, get_blas_funcs
 
 from autofit.graphical.factor_graphs import \
-    AbstractNode, Variable, FactorValue, JacobianValue, HessianValue
+    AbstractNode, Variable, Value, FactorValue, JacobianValue, HessianValue
 from autofit.graphical.utils import cached_property, Axis, FlattenArrays
 
-class AbstractTransform(ABC):
-
-    @property 
+class AbstractLinearTransform(ABC):
     @abstractmethod
-    def variables(self):
-        pass
+    def __mul__(self, x:np.ndarray) -> np.ndarray:
+        pass 
 
     @abstractmethod
-    def transform(self, values: Dict[Variable, np.ndarray]):
-        pass
-    
-    @abstractmethod
-    def untransform(self, values: Dict[Variable, np.ndarray]):
+    def __rtruediv__(self, x:np.ndarray) -> np.ndarray:
         pass
 
     @abstractmethod
-    def transform2d(self, values: Dict[Variable, np.ndarray]):
-        pass
-    
-    @abstractmethod
-    def untransform2d(self, values: Dict[Variable, np.ndarray]):
-        pass
+    def __rmul__(self, x:np.ndarray) -> np.ndarray:
+        pass 
 
-    @property
     @abstractmethod
-    def log_det(self):
+    def ldiv(self, x: np.ndarray) -> np.ndarray:
         pass
 
     @cached_property
-    def det(self):
-        return np.exp(self.log_det)
-
-
-class RescaleTransform(AbstractTransform):
-    def __init__(self, variables_scales: Dict[Variable, np.ndarray]):
-        self.scale = variables_scales
-        self.inv_scale = {v: scale**-1 for v, scale in self.scale.items()}
-
-    @property
-    def variables(self):
-        return self.scale.keys()
-
-    def transform(
-        self, 
-        values: Dict[Variable, np.ndarray]
-    ) -> Dict[Variable, np.ndarray]:
-        return {
-            v: val * self.inv_scale[v] for v, val in values.items()}
-
-    def untransform(
-        self, 
-        values: Dict[Variable, np.ndarray]
-    ) -> Dict[Variable, np.ndarray]:
-        return {
-            v: val * self.scale[v] for v, val in values.items()}
-
-    def transform2d(
-        self, 
-        values: Dict[Variable, np.ndarray]
-    ) -> Dict[Variable, np.ndarray]:
-        transformed = {}
-        for v, hess in values.items():
-            inv_scale = self.inv_scale[v]
-            shape = np.shape(inv_scale)
-            size = np.size(inv_scale)
-            hess2d = np.reshape(hess, (size, size))
-            inv_scale1d = inv_scale.ravel()
-            w_hess = hess2d * inv_scale1d[None, :] * inv_scale1d[:, None]
-            transformed[v] = w_hess.reshape(shape + shape)
-
-        return transformed
-
-    def untransform2d(
-        self, 
-        values: Dict[Variable, np.ndarray]
-    ) -> Dict[Variable, np.ndarray]:
-        actual_values = {}
-        for v, hess in values.items():
-            scale = self.scale[v]
-            shape = np.shape(scale)
-            size = np.size(scale)
-            hess2d = np.reshape(hess, (size, size))
-            scale1d = scale.ravel()
-            w_hess = hess2d * scale1d[None, :] * scale1d[:, None]
-            actual_values[v] = w_hess.reshape(shape + shape)
-
-        return actual_values
-
-    @cached_property
+    @abstractmethod
     def log_det(self):
-        return sum(np.log(scale).sum() for scale in self.scale.values())
+        pass
+    
[…]
