+++ autofit/graphical/factor_graphs/__init__.py
@@ -1,7 +1,8 @@
 from typing import Union
 
 from .abstract import \
-    FactorValue, JacobianValue, HessianValue, AbstractNode, Variable, Plate
+    Value, FactorValue, JacobianValue, HessianValue, AbstractNode, Variable, \
+    Plate
 from .factor import \
     AbstractFactor, Factor, DeterministicFactor
 from .jacobians import \
