### 2020-11-14 6abf1d4
adding more complex tests and enabling scaling of multidimensional values
