+++ autofit/graphical/factor_graphs/__init__.py
@@ -1,5 +1,4 @@
 from typing import Union
-
 from .abstract import \
     Value, FactorValue, JacobianValue, HessianValue, AbstractNode, Variable, \
     Plate
@@ -8,6 +7,9 @@ from .factor import \
 from .jacobians import \
     FactorJacobian, DeterministicFactorJacobian
 from .graph import FactorGraph
+from .transform import \
+    DiagonalTransform, CholeskyTransform, VariableTransform, \
+    FullCholeskyTransform 
 
 FactorNode = Union[
     Factor, 
