+++ autofit/graphical/factor_graphs/transform.py
@@ -6,6 +6,7 @@ from typing import Dict, Tuple, Optional, List
 
 import numpy as np
 from scipy.linalg import cho_factor, solve_triangular, get_blas_funcs
+from scipy._lib._util import _asarray_validated
 
 from autofit.graphical.factor_graphs import \
     AbstractNode, Variable, Value, FactorValue, JacobianValue, HessianValue
@@ -55,16 +56,40 @@ class AbstractLinearTransform(ABC):
         else:    
             return NotImplemented
 
-def _solve_triangular(c_and_lower, b, trans=False, overwrite_b=False):
-    c, lower = c_and_lower
-    return solve_triangular(
-        c, b, lower=lower, trans=trans, overwrite_b=overwrite_b)
 
+def _mul_triangular(c, b, trans=False, lower=True, overwrite_b=False, 
+                    check_finite=True):
+    """wrapper for BLAS function trmv to perform triangular matrix
+    multiplications
 
-def _mul_triangular(c_and_lower, b, trans=False, overwrite_b=False):
-    c, lower = c_and_lower 
-    a1 = np.asarray(c)
-    b1 = np.asarray(b)
+    
+    Parameters
+    ----------
+    a : (M, M) array_like
+        A triangular matrix
+    b : (M,) or (M, N) array_like
+        vector/matrix being multiplied
+    lower : bool, optional
+        Use only data contained in the lower triangle of `a`.
+        Default is to use upper triangle.
+    trans : bool, optional
+        type of multiplication,
+        
+        ========  =========
+        trans     system
+        ========  =========
+        False     a b
+        True      a^T b
+    overwrite_b : bool, optional
+        Allow overwriting data in `b` (may enhance performance)
+        not fully tested
+    check_finite : bool, optional
+        Whether to check that the input matrices contain only finite numbers.
+        Disabling may give a performance gain, but may result in problems
+        (crashes, non-termination) if the inputs do contain infinities or NaNs.
+    """    
+    a1 = _asarray_validated(c, check_finite=check_finite)
+    b1 = _asarray_validated(b, check_finite=check_finite)
 
     n = c.shape[1]
     if c.shape[0] != n:
@@ -73,7 +98,6 @@ def _mul_triangular(c_and_lower, b, trans=False, overwrite_b=False):
         raise ValueError(
             f"shapes {c.shape} and {b.shape} not aligned: "
             f"{n} (dim 1) != {b.shape[0]} (dim 0)")
-
     
     trmv, = get_blas_funcs(('trmv',), (a1, b1))
     if a1.flags.f_contiguous:
@@ -91,8 +115,11 @@ def _mul_triangular(c_and_lower, b, trans=False, overwrite_b=False):
     if b1.ndim == 1:
         return _trmv(a1, b1, overwrite_b)
     elif b1.ndim == 2:
+        # trmv only works for vector multiplications
+        # set Fortran order so memory contiguous
         b2 = np.array(b1, order='F')
         for i in range(b2.shape[1]):
+            # overwrite results
             _trmv(a1, b2[:, i], True)
 
         if overwrite_b:
@@ -103,35 +130,6 @@ def _mul_triangular(c_and_lower, b, trans=False, overwrite_b=False):
     else:
         raise ValueError("b must have 1 or 2 dimensions, has {b.ndim}")
 
-def _unwhiten_cholesky(c_and_lower, b):
-    c, lower = c_and_lower
-    b = np.asarray(b)
-    n = c.shape[1]
-        
-    return _solve_triangular(
-        c_and_lower, b.reshape(n, -1), trans=lower, 
-    ).reshape(b.shape)
-
-def _whiten_cholesky(c_and_lower, b):
-    c, lower = c_and_lower
[…]
