### 2020-11-16 f3712fe
refactor OptFactor
