+++ autofit/graphical/optimise.py
@@ -6,9 +6,11 @@ import numpy as np
 from scipy.optimize import \
     minimize, OptimizeResult, least_squares, approx_fprime
 
-from autofit.graphical import FixedMessage
-from autofit.mapper.variable import Variable
-from autofit.graphical.factor_graphs import Factor, JacobianValue
+from autofit.graphical.factor_graphs import \
+    Variable, Factor, JacobianValue
+from autofit.graphical.factor_graphs.transform import \
+    AbstractLinearTransform, identity_transform
+from autofit.graphical.messages import FixedMessage
 from autofit.graphical.mean_field import \
     MeanField, FactorApproximation, MeanFieldApproximation, Status
 from autofit.graphical.utils import \
@@ -24,20 +26,29 @@ class OptFactor:
             param_shapes: FlattenArrays,
             fixed_kws: Optional[Dict[str, np.ndarray]] = None,
             model_dist: Optional[MeanField] = None, 
+            transform: Optional[AbstractLinearTransform] = None, 
             bounds: Optional[Dict[str, Tuple[float, float]]] = None,
             method: str = 'L-BFGS-B',
     ):
         self.factor = factor
         self.param_shapes = param_shapes
         self._model_dist = model_dist
-
+        
+        self.transform = identity_transform if transform is None else transform
         self.param_bounds = bounds
         self.free_vars = tuple(self.param_shapes.keys())
         self.deterministic_variables = self.factor.deterministic_variables
 
         self.sign = 1
         self.fixed_kws = fixed_kws
-        self.method = method
+
+        meth = method.lower()
+        # method needs to return Hessian information.
+        if meth not in ('bfgs', 'l-bfgs-b'):
+            raise ValueError('Unknown solver %s' % method)
+        
+        self.method = meth 
+
 
         if bounds:
             # TODO check that this is correct for composite
@@ -60,6 +71,7 @@ class OptFactor:
     def from_approx(
             cls,
             factor_approx: FactorApproximation,
+            transform: Optional[AbstractLinearTransform] = None, 
             **kwargs
     ) -> 'OptFactor':
         fixed_kws = {}
@@ -77,27 +89,29 @@ class OptFactor:
             FlattenArrays(kwargs),
             fixed_kws=fixed_kws,
             model_dist=factor_approx.model_dist, 
+            transform=transform, 
             bounds=bounds,
         )
 
     def flatten(self, values: Dict[Variable, np.ndarray]) -> np.ndarray:
-        return self.param_shapes.flatten(values)
+        x0 = self.param_shapes.flatten(values)
+        return x0
 
     def unflatten(self, x0: np.ndarray) -> Dict[Variable, np.ndarray]:
-        values = self.param_shapes.unflatten(x0)
-        values.update(self.fixed_kws)
+        values = {**self.param_shapes.unflatten(x0), **self.fixed_kws}
         return values
 
     def __call__(self, x0):
-        values = self.unflatten(x0)
+        values = self.unflatten(self.transform.ldiv(x0))
         return self.sign * np.sum(self.factor(values, axis=None))
 
-    def func_jacobian(self, args):
-        values = self.unflatten(args)
+    def func_jacobian(self, x0):
+        values = self.unflatten(self.transform.ldiv(x0))
         fval, jval = self.factor.func_jacobian(
             values, self.free_vars, 
             axis=None, _calc_deterministic=True)
-        grad = self.param_shapes.flatten(jval)
+        
+        grad = self.flatten(jval) / self.transform
         return self.sign * fval.log_value, self.sign * grad
 
     def jacobian(self, args):
@@ -123,7 +137,20 @@ class OptFactor:
             for x0 in x0s
         )
 
-    def _parse_result(
[…]
