### 2020-11-17 68dbcc1
removed pyc + docs
