+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-autoconf==0.7.0
+autoconf==0.7.2
 corner>=2.0.1
 decorator>=4.2.1
 dill>=0.3.1.1
