### 2020-11-17 bc8f8c7
Set up LaplaceFactorOptimser
