### 2020-11-17 807ff2f
fixing test_linear_regression to work with EPOptimise
