### 2020-11-17 2b8b3dc
fixing test_logistic_regression to work with EPOptimise
