### 2020-11-17 4b655d7
fixing GammaMessage kl and adding gradients and hessians
