### 2020-11-17 6e7f5c7
bugfixing deterministic Factor Node
