### 2020-11-17 fd49be1
caching plates property for performance and minor refactoring
