### 2020-11-18 86dac84
sort results in gridsearch, resolves #136
