+++ autofit/database/model.py
@@ -1,0 +1,2 @@
+class PriorModel:
+    pass
