+++ autofit/database/serialize.py
@@ -1,0 +1,5 @@
+from . import model as m
+
+
+def serialize_model(model):
+    return m.PriorModel()
