### 2020-11-18 47df434
priormodel to db and back again
