+++ autofit/database/__init__.py
@@ -1,2 +1,1 @@
 from .model import *
-from .serialize import *
