+++ autofit/database/model.py
@@ -11,14 +11,6 @@ Base = declarative_base()
 _schema_version = 1
 
 
-def class_for_name(module_name, class_name):
-    # load the module, will raise ImportError if module cannot be loaded
-    m = importlib.import_module(module_name)
-    # get the class, will raise AttributeError if class cannot be found
-    c = getattr(m, class_name)
-    return c
-
-
 class PriorModel(Base):
     __tablename__ = "prior_model"
 
