### 2020-11-18 b8b5aca
bugfixing calculation of deterministic Jacobians
