### 2020-11-18 faefd1a
minor bugfixes
