### 2020-11-18 75c2bb4
Bugfixing FactorJacobian and AbstractMessage.as_factor for vectorised inputs
