### 2020-11-18 c00fd1e
priors to database and back
