### 2020-11-18 28f2f20
instances and simplification
