### 2020-11-18 66b1e6a
more bugfixes and fixing evidence calculation for LaplaceOptimiser
