+++ autofit/graphical/sampling.py
@@ -7,7 +7,10 @@ import numpy as np
 
 from autofit.graphical.factor_graphs import Factor
 from autofit.graphical.messages.abstract import AbstractMessage
-from autofit.graphical.mean_field import MeanField, MeanFieldApproximation, FactorApproximation, Status
+from autofit.graphical.mean_field import \
+    MeanField, MeanFieldApproximation, FactorApproximation, Status
+from autofit.graphical.expectation_propagation import \
+    EPMeanField, AbstractFactorOptimiser
 from autofit.graphical.messages import map_dists
 from autofit.graphical.utils import add_arrays
 
@@ -73,12 +76,35 @@ class SamplingHistory(NamedTuple):
             for f in self._fields))
 
 
-class AbstractSampler(ABC):
+class AbstractSampler(AbstractFactorOptimiser):
+    def __init__(self, delta=1., deltas=None, sample_kws=None):
+        self.deltas = defaultdict(lambda: delta)
+        if deltas:
+            self.deltas.update(deltas)
+
+        self.sample_kws = defaultdict(dict)
+        if sample_kws:
+            self.sample_kws.update(sample_kws)
+
     @abstractmethod
     def __call__(self, factor_approx: "FactorApproximation",
                  last_samples: Optional[SamplingResult] = None) -> SamplingResult:
         pass
 
+    def optimise(
+            self, 
+            factor: Factor, 
+            model_approx: EPMeanField, 
+            status: Status = Status(),
+    ) -> Tuple[EPMeanField, Status]:
+        delta = self.deltas[factor]
+        sample_kws = self.sample_kws[factor]
+
+        factor_approx = model_approx.factor_approximation(factor)
+        sample = self(factor_approx, **sample_kws)
+        model_dist = project_factor_approx_sample(factor_approx, sample)
+        projection, status = factor_approx.project(model_dist, delta=delta)
+        return model_approx.project(projection, status=status)
 
 class ImportanceSampler(AbstractSampler):
     def __init__(
@@ -86,17 +112,26 @@ class ImportanceSampler(AbstractSampler):
             n_samples: int = 200,
             n_resample: int = 100,
             min_n_eff: int = 100,
-            max_samples: int = 1000
+            max_samples: int = 1000,
+            force_sample: bool = True,
+            delta: float = 1.,
+            deltas = None, 
+            sample_kws = None, 
     ):
 
         self.params = dict(
             n_samples=n_samples, n_resample=n_resample,
-            min_n_eff=min_n_eff, max_samples=max_samples)
+            min_n_eff=min_n_eff, max_samples=max_samples,
+            force_sample=force_sample)
         self._history = defaultdict(SamplingHistory)
 
-    def sample(self, factor_approx: "FactorApproximation") -> SamplingResult:
+        super().__init__(
+            delta=delta, deltas=deltas, sample_kws=sample_kws)
+
+    def sample(self, factor_approx: "FactorApproximation", **kwargs) -> SamplingResult:
         # Update default params 
-        n_samples = self.params['n_samples']
+        params = {**self.params, **kwargs}
+        n_samples = params['n_samples']
         messages = ()
 
         factor = factor_approx.factor
@@ -183,21 +218,23 @@ class ImportanceSampler(AbstractSampler):
             proposal_dist=factor_approx.model_dist,
             n_samples=sampling_result.n_samples)
 
-    def stop_criterion(self, sample: SamplingResult) -> bool:
+    def stop_criterion(self, sample: SamplingResult, **kwargs) -> bool:
+        params = {**self.params, **kwargs}
         ess = effective_sample_size(sample.weights, 0).mean()
         n = len(sample.weights)
 
[…]
