+++ autofit/graphical/__init__.py
@@ -9,3 +9,5 @@ from .messages import FixedMessage, NormalMessage, GammaMessage, AbstractMessage
 from .optimise import OptFactor, LaplaceFactorOptimiser, lstsq_laplace_factor_approx
 from .sampling import ImportanceSampler, project_factor_approx_sample
 from ..mapper.variable import Variable, Plate
+
+from . import optimise as optimize
