### 2020-11-18 d5b4743
adding ImportanceSampler to linear_regression EPOptimiser tests
