+++ test_autofit/unit/text/test_samples_text.py
@@ -41,5 +41,5 @@ def test__latex(samples):
 
     latex_results_at_sigma = samples_text.latex(samples=samples, sigma=3.0)
 
-    assert r"one_label_{\mathrm{a}} = 1.00^{+0.20}_{-0.00} & " in latex_results_at_sigma
-    assert r"two_label_{\mathrm{a}} = 2.00^{+0.20}_{-0.00}" in latex_results_at_sigma
+    assert r"one_label_{\mathrm{o}} = 1.00^{+0.20}_{-0.00} & " in latex_results_at_sigma
+    assert r"two_label_{\mathrm{t}} = 2.00^{+0.20}_{-0.00}" in latex_results_at_sigma
