### 2020-11-25 3373b4f
adding docstrings and changing variable_names to name_variable_dict
