### 2020-11-25 47c3277
splitting _call_factor into two functions
