### 2020-11-25 c1dcd46
as_meanfield -> from_dist
