### 2020-11-25 0f6ade3
changed typing as relying on index method
