+++ autofit/graphical/factor_graphs/abstract.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from functools import wraps
 from typing import \
-    List, Tuple, Dict, cast, Collection, Set, NamedTuple, Optional, Union
+    List, Tuple, Dict, cast, Set, NamedTuple, Optional, Union
 from itertools import count
 
 import numpy as np
@@ -244,7 +244,7 @@ class AbstractNode(ABC):
 
     def _match_plates(
             self,
-            plates: Collection[Plate]
+            plates: List[Plate]
     ) -> np.ndarray:
         """
         Find indices plates from some factor in the collection of
