+++ test_autofit/unit/mapper/test_take_attributes.py
@@ -1,20 +1,60 @@
+import pytest
+
 import autofit as af
 from autofit.mock import mock as m
 
 
-def test_simple():
-    target = af.PriorModel(
+@pytest.fixture(
+    name="target_gaussian"
+)
+def make_target_gaussian():
+    return af.PriorModel(
         m.Gaussian
     )
 
-    prior = af.UniformPrior()
-    source = af.PriorModel(
+
+@pytest.fixture(
+    name="prior"
+)
+def make_prior():
+    return af.UniformPrior()
+
+
+@pytest.fixture(
+    name="source_gaussian"
+)
+def make_source_gaussian(prior):
+    return af.PriorModel(
         m.Gaussian,
         centre=prior
     )
 
+
+def test_simple(
+        source_gaussian,
+        target_gaussian,
+        prior
+):
+    target_gaussian.take_attributes(
+        source_gaussian
+    )
+
+    assert target_gaussian.centre is prior
+
+
+def test_in_collection(
+        source_gaussian,
+        target_gaussian,
+        prior
+):
+    target = af.CollectionPriorModel(
+        gaussian=target_gaussian
+    )
+    source = af.CollectionPriorModel(
+        gaussian=source_gaussian
+    )
     target.take_attributes(
         source
     )
 
-    assert target.centre is prior
+    assert target.gaussian.centre is prior
