### 2020-11-25 34f252a
some minor ep refactoring
