### 2020-11-25 90abdae
removing MeanFieldApproximation
