### 2020-11-25 287d076
attempt to pass aggregators from model to db
