### 2020-11-26 6940aee
removing MeanFieldApproximation
