+++ autofit/graphical/mean_field.py
@@ -492,155 +492,3 @@ class FactorApproximation(AbstractNode):
     def __repr__(self):
         # TODO make this nicer
         return f"{type(self).__name__}({self.factor}, ...)"
-
-    
-class _MeanFieldApproximation:
-    '''
-    TODO: delete this class
-    '''
-    def __init__(
-            self,
-            factor_graph: FactorGraph,
-            variable_factor_dist: VariableFactorDist,
-            factor_evidence: Optional[Dict[Factor, float]] = None
-    ):
-        self._factor_graph = factor_graph
-        self._variable_factor_dist = variable_factor_dist
-        if factor_evidence is None:
-            factor_evidence = {f: 0. for f in self.factor_graph.factors}
-
-        self._factor_evidence = factor_evidence
-
-    def __getitem__(self, item):
-        if isinstance(item, Variable):
-            return self.approx[item]
-        elif isinstance(item, Factor):
-            return self.factor_approximation(item)
-        else:
-            raise TypeError(
-                f"type passed {(type(item))} is not `Variable` or `Factor`")
-
-    @classmethod
-    def from_approx_dists(
-            cls,
-            factor_graph: FactorGraph,
-            approx_dists: Dict[Variable, AbstractMessage],
-            factor_evidence: Optional[Dict[Factor, float]] = None,
-    ) -> "MeanFieldApproximation":
-        variable_factor_dist = {}
-        for factor, variables in factor_graph.factor_all_variables.items():
-            for variable in variables:
-                variable_factor_dist.setdefault(
-                    variable, {}
-                ).setdefault(
-                    factor, approx_dists[variable]
-                )
-
-        return cls(
-            factor_graph,
-            variable_factor_dist,
-            factor_evidence=factor_evidence
-        )
-
-    @classmethod
-    def from_kws(
-            cls,
-            factor_graph: FactorGraph,
-            approx_dists: Dict[Variable, AbstractMessage],
-            factor_evidence: Optional[Dict[Factor, float]] = None,
-    ) -> "MeanFieldApproximation":
-        return cls.from_approx_dists(
-            factor_graph=factor_graph,
-            approx_dists=approx_dists,
-            factor_evidence=factor_evidence
-        )
-
-    def project(self, projection: FactorApproximation,
-                status: Optional[Status] = None
-                ) -> Tuple["MeanFieldApproximation", Status]:
-        """
-        """
-        success, messages = Status() if status is None else status
-
-        factor = projection.factor
-        factor_projection = projection.factor_dist
-        factor_evidence = self._factor_evidence.copy()
-        variable_factor_dist = {
-            variable: factor_dist.copy()
-            for variable, factor_dist in self._variable_factor_dist.items()}
-
-        if success and projection.is_valid:
-            for v, dist in factor_projection.items():
-                variable_factor_dist[v][factor] = dist
-
-            # factor_evidence[factor] = projection.log_norm
-        else:
-            messages += f"projection for {factor} is invalid",
-            success = False
-
-        status = Status(success, messages)
-        approx = type(self)(factor_graph=self._factor_graph,
-                            variable_factor_dist=variable_factor_dist,
-                            factor_evidence=factor_evidence)
-        return approx, status
-
-    @property
-    def factor_graph(self) -> FactorGraph:
-        return self._factor_graph
-
-    def _variable_cavity_dist(self, variable: str,
-                              cavity_factor: Factor
-                              ) -> Optional[AbstractMessage]:
-        dists = [dist for factor, dist in
-                 self._variable_factor_dist[variable].items()
-                 if factor != cavity_factor]
-        if dists:
-            return prod(dists)
-        return None
-
[…]
