### 2020-11-26 7658320
adding some type annotation to AbstractMessage
