+++ autofit/graphical/messages/abstract.py
@@ -2,10 +2,12 @@ from abc import ABC, abstractmethod
 from functools import reduce
 from itertools import chain
 from operator import and_
-from typing import Optional, Tuple, Union
+from typing import Optional, Tuple, Union, Iterator
 from inspect import getfullargspec
 
 import numpy as np
+
+from autofit.mapper.variable import Variable
 
 
 class AbstractMessage(ABC):
@@ -20,39 +22,40 @@ class AbstractMessage(ABC):
         pass
 
     @abstractmethod
-    def sample(self, n_samples=None):
+    def sample(self, n_samples: Optional[int] = None):
         pass
 
     @staticmethod
     @abstractmethod
-    def invert_natural_parameters(natural_parameters):
+    def invert_natural_parameters(natural_parameters: np.ndarray
+    ) -> Tuple[np.ndarray,  ...]:
         pass
 
     @staticmethod
     @abstractmethod
-    def to_canonical_form(x):
+    def to_canonical_form(x: np.ndarray) -> np.ndarray:
         pass
 
     @property
     @abstractmethod
-    def log_partition(self):
+    def log_partition(self) -> np.ndarray:
         pass
 
     @property
     @abstractmethod
-    def mean(self):
+    def mean(self) -> np.ndarray:
         pass
 
     @property
     @abstractmethod
-    def variance(self):
+    def variance(self) -> np.ndarray:
         pass
 
     @property
-    def scale(self):
+    def scale(self) -> np.ndarray:
         return self.variance ** 0.5
 
-    def __init__(self, parameters, log_norm=0.):
+    def __init__(self, parameters: Tuple[np.ndarray, ...], log_norm=0.):
         self.log_norm = log_norm
         self._broadcast = np.broadcast(*parameters)
         if self.shape:
@@ -62,33 +65,40 @@ class AbstractMessage(ABC):
             self.parameters = tuple(parameters)
 
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator[np.ndarray]:
         return iter(self.parameters)
 
     @property
-    def shape(self):
+    def shape(self) -> Tuple[int, ...]:
         return self._broadcast.shape
 
     @property
-    def ndim(self):
+    def ndim(self) -> int: 
         return self._broadcast.ndim
 
     @classmethod
-    def from_natural_parameters(cls, parameters, **kwargs):
+    def from_natural_parameters(
+            cls, 
+            parameters: Tuple[np.ndarray, ...], 
+            **kwargs
+    ) -> "AbstractMessage":
         args = cls.invert_natural_parameters(parameters)
         return cls(*args, **kwargs)
 
     @classmethod
     @abstractmethod
-    def invert_sufficient_statistics(cls, sufficient_statistics):
+    def invert_sufficient_statistics(cls, sufficient_statistics: np.ndarray
+    ) -> Tuple[np.ndarray, ...]:
         pass
 
     @classmethod
-    def from_sufficient_statistics(cls, suff_stats, **kwargs):
+    def from_sufficient_statistics(cls, suff_stats: np.ndarray, **kwargs
+    ) -> "AbstractMessage":
         natural_params = cls.invert_sufficient_statistics(suff_stats)
         return cls.from_natural_parameters(natural_params, **kwargs)
 
-    def sum_natural_parameters(self, *dists):
+    def sum_natural_parameters(self, *dists: "AbstractMessage"
+    ) -> "AbstractMessage":
         """return the unnormalised result of multiplying the pdf
         of this distribution with another distribution of the same
         type
@@ -100,7 +110,8 @@ class AbstractMessage(ABC):
         mul_dist = self.from_natural_parameters(new_params)
         return mul_dist
 
-    def sub_natural_parameters(self, other):
+    def sub_natural_parameters(self, other: "AbstractMessage"
[…]
