### 2020-12-01 b6bc9c6
aggregator can now output its own results
