### 2020-12-02 13199bb
handle dicts in take_attributes
