### 2020-12-02 967ec99
fixed lru_cache decorator
