+++ autofit/graphical/optimise.py
@@ -1,41 +1,62 @@
-from itertools import repeat
 from collections import defaultdict
+from itertools import repeat
 from typing import (
-    Optional, Dict, Tuple, NamedTuple, Any, List, Iterator)
+    Optional, Dict, Tuple, Any, List, Iterator
+)
 
 import numpy as np
-from scipy.optimize import \
-    minimize, OptimizeResult, least_squares, approx_fprime
+from scipy.optimize import (
+    minimize,
+    OptimizeResult,
+    least_squares,
+    approx_fprime
+)
 
-from autofit.graphical.utils import \
-    propagate_uncertainty, FlattenArrays, OptResult
-from autofit.graphical.factor_graphs import \
-    Variable, Factor, JacobianValue
-from autofit.graphical.factor_graphs.transform import \
-    AbstractLinearTransform, identity_transform, CovarianceTransform
+from autofit.graphical.expectation_propagation import (
+    EPMeanField,
+    AbstractFactorOptimiser
+)
+from autofit.graphical.factor_graphs import (
+    Variable,
+    Factor,
+    JacobianValue
+)
+from autofit.graphical.factor_graphs.transform import (
+    AbstractLinearTransform,
+    identity_transform,
+    CovarianceTransform
+)
+from autofit.graphical.mean_field import (
+    MeanField,
+    FactorApproximation,
+    Status
+)
 from autofit.graphical.messages import FixedMessage
-from autofit.graphical.mean_field import \
-    MeanField, FactorApproximation, Status
-from autofit.graphical.expectation_propagation import \
-    EPMeanField, AbstractFactorOptimiser
+from autofit.graphical.utils import (
+    propagate_uncertainty,
+    FlattenArrays,
+    OptResult
+)
+
 
 class OptFactor:
     """
     """
+
     def __init__(
             self,
             factor: Factor,
             param_shapes: FlattenArrays,
             fixed_kws: Optional[Dict[str, np.ndarray]] = None,
-            model_dist: Optional[MeanField] = None, 
-            transform: Optional[AbstractLinearTransform] = None, 
+            model_dist: Optional[MeanField] = None,
+            transform: Optional[AbstractLinearTransform] = None,
             bounds: Optional[Dict[str, Tuple[float, float]]] = None,
             method: str = 'L-BFGS-B',
     ):
         self.factor = factor
         self.param_shapes = param_shapes
         self._model_dist = model_dist
-        
+
         self.transform = identity_transform if transform is None else transform
         self.param_bounds = bounds
         self.free_vars = tuple(self.param_shapes.keys())
@@ -48,9 +69,8 @@ class OptFactor:
         # method needs to return Hessian information.
         if meth not in ('bfgs', 'l-bfgs-b'):
             raise ValueError('Unknown solver %s' % method)
-        
-        self.method = meth 
 
+        self.method = meth
 
         if bounds:
             # TODO check that this is correct for composite
@@ -62,7 +82,7 @@ class OptFactor:
         else:
             self.bounds = bounds
 
-    @property 
+    @property
     def model_dist(self):
         if self._model_dist is None:
             raise ValueError("Model dist not defined")
@@ -73,7 +93,7 @@ class OptFactor:
     def from_approx(
             cls,
             factor_approx: FactorApproximation,
-            transform: Optional[AbstractLinearTransform] = None, 
+            transform: Optional[AbstractLinearTransform] = None,
     ) -> 'OptFactor':
         value_shapes = {}
         fixed_kws = {}
@@ -90,8 +110,8 @@ class OptFactor:
             factor_approx,
             FlattenArrays(value_shapes),
             fixed_kws=fixed_kws,
-            model_dist=factor_approx.model_dist, 
-            transform=transform, 
[…]
