### 2020-12-02 31409b5
changed declarative implementation to use conditionally terminating optimiser
