### 2020-12-02 cdf5fbf
removing dictionary default arguments
