### 2020-12-02 079182f
optionally compute jacobian
