### 2020-12-02 3fed437
model factor takes analysis instead of function
