### 2020-12-02 3c4deee
customise optimiser by passing it to a model factor
