### 2020-12-02 7170ed8
Revert "Merge branch 'feature/graphical_analysis' into develop"
This reverts commit b6a43d526f2cd5668cfa2a8e4ef1cdfbd4b406a8, reversing
changes made to e402deed8e99fc397724a8007e548e23ebedfc2a.
