+++ autofit/graphical/expectation_propagation.py
@@ -1,20 +1,21 @@
+
 from abc import ABC, abstractmethod
-from collections import defaultdict
+from collections import defaultdict 
 from itertools import count
 from typing import (
-    Dict, Tuple, Optional, List,
+    Dict, Tuple, Optional, NamedTuple, Iterator, List,
     Callable
 )
 
 import numpy as np
 
-from autofit.graphical.factor_graphs import (
-    Factor, AbstractNode, FactorGraph
-)
-from autofit.graphical.mean_field import MeanField, FactorApproximation
 from autofit.graphical.messages.abstract import AbstractMessage
-from autofit.graphical.utils import Status
+
+from autofit.graphical.factor_graphs import (
+    Factor, AbstractNode, FactorGraph)
 from autofit.mapper.variable import Variable
+from autofit.graphical.utils import Status, cached_property
+from autofit.graphical.mean_field import MeanField, FactorApproximation 
 
 
 class EPMeanField(AbstractNode):
@@ -59,7 +60,6 @@ class EPMeanField(AbstractNode):
         given the passed FactorApproximation, return a new `EPMeanField`
         object encoding the updated mean-field approximation
     '''
-
     def __init__(
             self,
             factor_graph: FactorGraph,
@@ -72,25 +72,25 @@ class EPMeanField(AbstractNode):
             for v in vs:
                 variable_factor.setdefault(v, set()).add(factor)
         self._variable_factor = variable_factor
-
+        
         super().__init__(**self.factor_graph._kwargs)
 
-    @property
+    @property 
     def name(self):
         return f"EP_{self.factor_graph.name}"
 
     @property
-    def variables(self):
+    def variables(self): 
         return self.factor_graph.variables
 
     @property
-    def deterministic_variables(self):
+    def deterministic_variables(self): 
         return self.factor_graph.deterministic_variables
 
     # @property
     # def variable_names(self) -> Dict[str, Variable]: 
     #     return self.factor_graph.variable_names
-
+ 
     @property
     def factor_mean_field(self) -> Dict[Factor, MeanField]:
         return self._factor_mean_field.copy()
@@ -115,7 +115,7 @@ class EPMeanField(AbstractNode):
             factor_mean_field)
 
     from_kws = from_approx_dists
-
+    
     def factor_approximation(self, factor: Factor) -> FactorApproximation:
         factor_mean_field = self._factor_mean_field.copy()
         factor_dist = factor_mean_field.pop(factor)
@@ -129,7 +129,7 @@ class EPMeanField(AbstractNode):
             factor, cavity_dist, factor_dist, model_dist)
 
     def project_factor_approx(
-            self, projection: FactorApproximation, status: Optional[Status] = None,
+        self, projection: FactorApproximation, status: Optional[Status] = None,
     ) -> "EPMeanField":
         """
         """
@@ -151,7 +151,7 @@ class EPMeanField(AbstractNode):
 
     @property
     def variable_factor_message(self
-                                ) -> Dict[Variable, Dict[Factor, AbstractMessage]]:
+    ) -> Dict[Variable, Dict[Factor, AbstractMessage]]:
         variable_factor_message = {
             v: {} for v in self.all_variables}
         for factor, meanfield in self.factor_mean_field.items():
@@ -167,7 +167,7 @@ class EPMeanField(AbstractNode):
         for meanfield in self.factor_mean_field.values():
             for v, message in meanfield.items():
                 variable_messages[v].append(message)
-
+        
         return variable_messages
 
     @property
@@ -177,13 +177,13 @@ class EPMeanField(AbstractNode):
             for v, ms in self.variable_messages.items()
         }
 
-    @property
+    @property 
[…]
