### 2020-12-02 f4039b4
set a seed for graphical
