### 2020-12-02 844f7e8
more take attribute tests
