+++ test_autofit/unit/mapper/test_take_attributes.py
@@ -60,6 +60,72 @@ def test_in_collection(
     assert target.gaussian.centre is prior
 
 
+def test_tuple(
+        source_gaussian,
+        target_gaussian,
+        prior
+):
+    source_gaussian.centre = (prior, 1.0)
+    target_gaussian.take_attributes(
+        source_gaussian
+    )
+
+    assert target_gaussian.centre == (prior, 1.0)
+
+
+def test_tuple_in_instance(
+        target_gaussian,
+        prior
+):
+    # noinspection PyTypeChecker
+    source_gaussian = m.Gaussian(
+        centre=(prior, 1.0)
+    )
+    target_gaussian.take_attributes(
+        source_gaussian
+    )
+
+    assert target_gaussian.centre == (prior, 1.0)
+
+
+def test_tuple_in_collection(
+        source_gaussian,
+        target_gaussian,
+        prior
+):
+    source_gaussian.centre = (prior, 1.0)
+
+    source = af.CollectionPriorModel(
+        gaussian=source_gaussian
+    )
+    target = af.CollectionPriorModel(
+        gaussian=target_gaussian
+    )
+
+    target.take_attributes(source)
+    assert target.gaussian.centre == (prior, 1.0)
+
+
+def test_tuple_in_instance_in_collection(
+        target_gaussian,
+        prior
+):
+    # noinspection PyTypeChecker
+    source_gaussian = m.Gaussian(
+        centre=(prior, 1.0)
+    )
+
+    source = af.CollectionPriorModel(
+        gaussian=source_gaussian
+    )
+    target = af.CollectionPriorModel(
+        gaussian=target_gaussian
+    )
+
+    target.take_attributes(source)
+    assert target.gaussian.centre == (prior, 1.0)
+
+
 def test_source_is_dict(
         source_gaussian,
         target_gaussian,
