### 2020-12-02 99253ac
account for tuple priors
