### 2020-12-02 ad738a1
Revert "Revert "Merge branch 'feature/graphical_analysis' into develop""
This reverts commit 7170ed81405490265d403124382e04d060ec6f6d.
