### 2020-12-02 9b4a4db
docs and typing
