### 2020-12-02 01c1384
db handling None
