### 2020-12-02 a6f8d5e
use class method rather than new for factory (avoid clash with sqlalchemy)
