### 2020-12-02 8915043
fixed none type - instantiation of database model works
