+++ autofit/database/__init__.py
@@ -1,2 +1,3 @@
+from .instance import *
 from .model import *
 from .prior import *
