### 2020-12-09 rhayes777/PyAutoFit#142 @karllark
Error when running welcome.py
There was an error when running the install instructions.  Below is a portion of the output showing the error that occurred when running `welcome.py`.

    backend=TKAgg
    backend=Qt5Agg
    backeknd=Qt4Agg

    [Press Enter to continue]

    Traceback (most recent call last):
     File "welcome.py", line 63, in <module>
       dataset_path = path.join("dataset", "example", "gaussian_x1")
     NameError: name 'path' is not defined
