### 2020-12-09 rhayes777/PyAutoFit#142 @Jammy2211
Ah, there is a `from os import path` missing at the top of the welcome.py script. We did a big refactor to support Windows recently and I inserted that everywhere except there.

Could you add it manually and run the script? In the middle of a release so wary of pushing a fix myself currently.
