### 2020-12-09 rhayes777/PyAutoFit#142 @karllark
That fixes the issue.  Moving on with `welcome.py`, there is a type.

"The reaming folders..." should be "The remaining folders...".
