### 2020-12-09 rhayes777/PyAutoFit#142 @Jammy2211
Cheers, workspace was also spelled workpace :P.
