### 2020-12-09 rhayes777/PyAutoFit#143 @karllark
Readme and readthedocs
The readthedocs link is at the very end of the README.  Better to have it earlier.   In fact, the entire "Getting Started" section would be better much closer to the beginning - even before installation.
