### 2020-12-09 #2550 @karllark
Once the (minor) issues https://github.com/rhayes777/PyAutoFit/issues/142 and https://github.com/rhayes777/PyAutoFit/issues/143 are closed, then I am happy and would recommend acceptance.
