+++ autofit/non_linear/samples.py
@@ -1,6 +1,7 @@
 import csv
 import json
 import math
+from collections import defaultdict
 from typing import List
 
 import numpy as np
@@ -9,38 +10,51 @@ from autofit.mapper.model import ModelInstance
 from autofit.mapper.model_mapper import ModelMapper
 
 
-def load_from_table(filename, model):
+# parameters, log_likelihoods, log_priors, log_posteriors, weights
 
-    parameters = []
-    log_likelihoods = []
-    log_priors = []
-    log_posteriors = []
-    weights = []
+
+def load_from_table(filename, model: ModelMapper):
+    output = defaultdict(list)
 
     with open(filename, "r+", newline="") as f:
         reader = csv.reader(f)
-        for index, row in enumerate(reader):
-            if index > 0:
-                parameters.append(
-                    [float(param) for param in row[0: model.prior_count]]
+        headers = next(reader)
+        for row in reader:
+            for header, value in zip(
+                    headers,
+                    row
+            ):
+                output[header].append(
+                    float(value)
                 )
-                log_likelihoods.append(float(row[model.prior_count]))
-                log_priors.append(float(row[model.prior_count + 1]))
-                log_posteriors.append(float(row[model.prior_count + 2]))
-                weights.append(float(row[model.prior_count + 3]))
 
-    return parameters, log_likelihoods, log_priors, log_posteriors, weights
+    path_prior_tuples = model.path_priors_tuples
+    parameters = [
+        output["_".join(path)]
+        for path, _
+        in path_prior_tuples
+    ]
+
+    parameters = [
+        [
+            parameters[i][j]
+            for i in range(len(parameters))
+        ]
+        for j in range(len(parameters[0]))
+    ]
+
+    return parameters, output["log_likelihood"], output["log_prior"], output["log_posterior"], output["weights"]
 
 
 class OptimizerSamples:
     def __init__(
-        self,
-        model: ModelMapper,
-        parameters: List[List[float]],
-        log_likelihoods: List[float],
-        log_priors: List[float],
-        weights: List[float],
-        time: float = None,
+            self,
+            model: ModelMapper,
+            parameters: List[List[float]],
+            log_likelihoods: List[float],
+            log_priors: List[float],
+            weights: List[float],
+            time: float = None,
     ):
         """The `Samples` of a non-linear search, specifically the samples of an search which only provides
         information on the global maximum likelihood solutions, but does not map-out the posterior and thus does
@@ -175,14 +189,14 @@ class OptimizerSamples:
 
 class PDFSamples(OptimizerSamples):
     def __init__(
-        self,
-        model: ModelMapper,
-        parameters: List[List[float]],
-        log_likelihoods: List[float],
-        log_priors: List[float],
-        weights: List[float],
-        unconverged_sample_size: int = 100,
-        time: float = None,
+            self,
+            model: ModelMapper,
+            parameters: List[List[float]],
+            log_likelihoods: List[float],
+            log_priors: List[float],
+            weights: List[float],
+            unconverged_sample_size: int = 100,
+            time: float = None,
     ):
         """The `Samples` of a non-linear search, specifically the samples of a `NonLinearSearch` which maps out the
         posterior of parameter space and thus does provide information on parameter errors.
@@ -215,7 +229,8 @@ class PDFSamples(OptimizerSamples):
             Where the table is to be written
         """
 
-        parameters, log_likelihoods, log_priors, log_posteriors, weights = load_from_table(filename=filename, model=model)
+        parameters, log_likelihoods, log_priors, log_posteriors, weights = load_from_table(filename=filename,
+                                                                                           model=model)
 
[…]
