### 2020-12-10 a7992aa
refactoring sample
