### 2020-12-10 31f9fc7
convenience method to convert tests
