### 2020-12-10 837a886
illustrating advantages of sample class
