### 2020-12-10 c006927
model factors implement analysis interface
