### 2020-12-10 8f53c93
renamed collection to GraphicalModel
