+++ test_autofit/unit/graphical/gaussian/test_declarative.py
@@ -60,7 +60,7 @@ def test_shared_intensity():
     The factor model can compute all the variables and messages required as well as construct
     a factor graph representing a fit on the ensemble.
     """
-    factor_model = ep.ModelFactorCollection(
+    factor_model = ep.GraphicalModel(
         make_factor_model(centre=40, sigma=10), make_factor_model(centre=60, sigma=15)
     )
 
