+++ autofit/graphical/__init__.py
@@ -1,4 +1,4 @@
-from .declarative import ModelFactor, ModelFactorCollection
+from .declarative import ModelFactor, GraphicalModel
 from .factor_graphs import \
     Factor, FactorJacobian, FactorGraph, AbstractFactor, FactorValue, \
     DiagonalTransform, CholeskyTransform, VariableTransform, \
