### 2020-12-10 bc0bf42
copy in take attributes
