### 2020-12-11 28a16ee
Bug fixes for OptFactor to improve early stopping
