+++ test_autofit/unit/tools/test_pipeline.py
@@ -50,11 +50,12 @@ class TestPipeline:
         phase1 = MockPhase(search=af.MockSearch("one"))
         phase2 = MockPhase(search=af.MockSearch("two"))
 
-        af.Pipeline("name", "", phase1, phase2)
+        af.Pipeline("name", "", None, phase1, phase2)
         with pytest.raises(af.exc.PipelineException):
             af.Pipeline(
                 "name",
                 "",
+                None,
                 MockPhase(search=af.MockSearch("one")),
                 MockPhase(search=af.MockSearch("one")),
             )
@@ -73,8 +74,8 @@ class TestPipeline:
         phase.model.profile.centre_0 = af.UniformPrior()
 
     def test_name_composition(self):
-        first = af.Pipeline("first", "")
-        second = af.Pipeline("second", "")
+        first = af.Pipeline("first", "", None)
+        second = af.Pipeline("second", "", None)
 
         assert (first + second).pipeline_name == "first + second"
 
@@ -84,7 +85,7 @@ class TestPhasePipelineName:
     def test_name_stamping(self):
         one = MockPhase(search=af.MockSearch("one"))
         two = MockPhase(search=af.MockSearch("two"))
-        af.Pipeline("name", "", one, two)
+        af.Pipeline("name", "", None, one, two)
 
         assert one.pipeline_name == "name"
         assert two.pipeline_name == "name"
@@ -92,8 +93,8 @@ class TestPhasePipelineName:
     def test_no_restamping(self):
         one = MockPhase(search=af.MockSearch("one"))
         two = MockPhase(search=af.MockSearch("two"))
-        pipeline_one = af.Pipeline("one", "", one)
-        pipeline_two = af.Pipeline("two", "", two)
+        pipeline_one = af.Pipeline("one", "", None, one)
+        pipeline_two = af.Pipeline("two", "", None, two)
 
         composed_pipeline = pipeline_one + pipeline_two
 
