### 2020-12-14 ba7cae9
added function which creates new Samples within a specified rane
