### 2020-12-15 039325c
changing defaults to result in less computation
