### 2020-12-15 d5e1174
refactoring test_logistic_regression to have analytic gradients for speed
