### 2020-12-16 191709f
GraphicalModel -> FactorGraphModel
