+++ autofit/graphical/declarative.py
@@ -314,7 +314,7 @@ class ModelFactor(Factor, AbstractModelFactor):
         )[0]
 
 
-class GraphicalModel(AbstractModelFactor):
+class FactorGraphModel(AbstractModelFactor):
     def __init__(self, *model_factors: ModelFactor):
         """
         A collection of factors that describe models, which can be
