### 2020-12-16 2a58fa1
global switch to ignore prior limits
