+++ test_autofit/unit/config/general.ini
@@ -1,11 +1,14 @@
 [output]
-log_to_file=False
-log_file=output.log
-log_level=INFO
-model_results_decimal_places=3
-remove_files=True
-force_pickle_overwrite=False
+log_to_file = False
+log_file = output.log
+log_level = INFO
+model_results_decimal_places = 3
+remove_files = True
+force_pickle_overwrite = False
 
 [hpc]
-hpc_mode=False
-iterations_per_update=5000
+hpc_mode = False
+iterations_per_update = 5000
+
+[model]
+ignore_prior_limits = False
