### 2020-12-16 789d444
vaguely making progress with querying
