### 2020-12-17 c692381
fixing gradient bug in GammaMessage
