### 2020-12-17 7e4097a
adding networkx visualisations to FactorGraph
