+++ README.rst
@@ -18,8 +18,8 @@ Getting Started
 
 To get started checkout our `readthedocs <https://pyautofit.readthedocs.io/>`_,
 where you'll find our installation guide, a complete overview of **PyAutoFit**'s features, examples scripts and
-tutorials, detailed API documentation and the `HowToFit lecture series <https://pyautofit.readthedocs.io/en/latest/howtofit/howtofit.html>`_
-on how to integrate **PyAutoFit** into your modeling software.
+tutorials, and the `HowToFit Jupyter notebook tutorials <https://pyautofit.readthedocs.io/en/latest/howtofit/howtofit.html>`_
+which introduces new users to **PyAutoFit**.
 
 Installation
 ------------
