### 2020-12-18 f342f79
Refactoring `AbstractFactorOptimiser` for more general interface
