### 2020-12-18 be06cd6
adding exact updates to AbstractMessage and refactor AbstractMessage.__init__
