+++ autofit/config/general.ini
@@ -11,4 +11,7 @@ hpc_mode = False
 iterations_per_update = 5000
 
 [model]
-ignore_prior_limits = False
+ignore_prior_limits=False
+
+[test]
+test_mode=False
