### 2020-12-21 de90303
embedded equality query
