+++ test_autofit/unit/database/test_query.py
@@ -1,5 +1,6 @@
 import pytest
 
+import autofit as af
 from autofit import database as db
 from autofit.mock import mock as m
 
@@ -58,6 +59,49 @@ def test_embedded_equality_query(
     assert (aggregator.lens.centre == 1).string == string
 
 
+def test_embedded_query(
+        session,
+        aggregator
+):
+    model_1 = db.Object.from_object(
+        af.Collection(
+            gaussian=m.Gaussian(
+                centre=1
+            )
+        )
+    )
+    model_2 = db.Object.from_object(
+        af.Collection(
+            gaussian=m.Gaussian(
+                centre=2
+            )
+        )
+    )
+
+    session.add_all([
+        model_1,
+        model_2
+    ])
+
+    result = aggregator.filter(
+        aggregator.centre == 0
+    )
+
+    assert result == []
+
+    result = aggregator.filter(
+        aggregator.gaussian.centre == 1
+    )
+
+    assert result == [model_1]
+
+    result = aggregator.filter(
+        aggregator.gaussian.centre == 2
+    )
+
+    assert result == [model_2]
+
+
 def test_query(
         session,
         aggregator
@@ -77,7 +121,6 @@ def test_query(
         gaussian_1,
         gaussian_2
     ])
-    session.commit()
 
     result = aggregator.filter(
         aggregator.centre == 0
