### 2020-12-21 6499c33
string based queries
