+++ howtofit/chapter_phase_api/src/plot/fit_plots.py
@@ -111,7 +111,7 @@ def data(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.data,
         title="Data",
@@ -145,7 +145,7 @@ def noise_map(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.noise_map,
         title="Noise-Map",
@@ -179,7 +179,7 @@ def model_data(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.model_data,
         title="Model Data",
@@ -213,7 +213,7 @@ def residual_map(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.residual_map,
         title="Residual-Map",
@@ -247,7 +247,7 @@ def normalized_residual_map(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.normalized_residual_map,
         title="Normalized Residual-Map",
@@ -281,7 +281,7 @@ def chi_squared_map(
     output_format : str
         Whether the data is output as a `.png` file ("png") or displayed on the screen ("show").
     """
-    line_plots.line(
+    line_plots.line_plot(
         xvalues=fit.xvalues,
         line=fit.chi_squared_map,
         title="Chi-Squared Map",
