### 2021-01-06 52970b8
trivial query combination
