+++ test_autofit/unit/database/test_query/test_query.py
@@ -1,31 +1,6 @@
-import pytest
-
 import autofit as af
 from autofit import database as db
 from autofit.mock import mock as m
-
-
-@pytest.fixture(
-    name="aggregator"
-)
-def make_aggregator(
-        session
-):
-    return db.Aggregator(session)
-
-
-@pytest.fixture(
-    name="centre_query"
-)
-def make_centre_query():
-    return "SELECT parent_id FROM object WHERE name = 'centre'"
-
-
-@pytest.fixture(
-    name="equality_query"
-)
-def make_equality_query():
-    return "SELECT parent_id FROM object, value WHERE name = 'centre' AND value = 1 AND value.id = object.id"
 
 
 def test_attribute_query(
@@ -70,16 +45,6 @@ def test_string_equality_query(
     assert (aggregator.centre == "one").string == string
 
 
-@pytest.fixture(
-    name="type_equality_query"
-)
-def make_type_equality_query():
-    return (
-        "SELECT parent_id FROM object WHERE class_path = 'autofit.mock.mock.Gaussian' "
-        "AND name = 'centre'"
-    )
-
-
 def test_type_equality_query(
         aggregator,
         type_equality_query
@@ -96,44 +61,6 @@ def test_embedded_equality_query(
 
     string = f"SELECT parent_id FROM object WHERE name = 'galaxies' AND id IN ({string})"
     assert (aggregator.galaxies.lens.centre == 1).string == string
-
-
-def test_trivial_combined_query(
-        aggregator,
-        type_equality_query
-):
-    query = (aggregator.centre == m.Gaussian) & (aggregator.centre == m.Gaussian)
-    assert query.string == type_equality_query
-
-
-def test_top_level_combined_query(
-        aggregator,
-        type_equality_query
-):
-    other_query = (
-        "SELECT parent_id FROM object WHERE class_path = 'autofit.mock.mock.ComplexClass' "
-        "AND name = 'other'"
-    )
-    string = (
-        f"SELECT t0.parent_id FROM "
-        f"({type_equality_query}) as t0, "
-        f"({other_query}) as t1 "
-        f"WHERE t0.parent_id = t1.parent_id"
-    )
-    query = (aggregator.centre == m.Gaussian) & (aggregator.other == m.ComplexClass)
-    assert query.string == string
-
-
-def test_combined_query(
-        aggregator,
-        type_equality_query,
-        equality_query
-):
-    string = ((aggregator.lens == m.Gaussian) & (aggregator.lens.centre == 1)).string
-
-    assert string == (
-        f"SELECT parent_id FROM object WHERE name = 'lens' AND id IN ({type_equality_query} AND id in {equality_query})"
-    )
 
 
 def test_embedded_query(
