+++ test_autofit/unit/database/test_query/test_combined.py
@@ -1,7 +1,7 @@
 from autofit.mock import mock as m
 
 
-def test_trivial_combined_query(
+def test_trivial(
         aggregator,
         type_equality_query
 ):
@@ -9,12 +9,14 @@ def test_trivial_combined_query(
     assert query.string == type_equality_query
 
 
-def test_top_level_combined_query(
+def test_top_level(
         aggregator,
         type_equality_query
 ):
     other_query = (
-        "SELECT parent_id FROM object WHERE class_path = 'autofit.mock.mock.ComplexClass' "
+        "SELECT parent_id "
+        "FROM object "
+        "WHERE class_path = 'autofit.mock.mock.ComplexClass' "
         "AND name = 'other'"
     )
     string = (
@@ -27,13 +29,20 @@ def test_top_level_combined_query(
     assert query.string == string
 
 
-def test_combined_query(
+def test_second_level(
         aggregator,
-        type_equality_query,
         equality_query
 ):
-    string = ((aggregator.lens == m.Gaussian) & (aggregator.lens.centre == 1)).string
+    equality_query_2 = (
+        "SELECT parent_id "
+        "FROM object, value "
+        "WHERE name = 'intensity' "
+        "AND value = 0 "
+        "AND value.id = object.id"
+    )
+
+    string = ((aggregator.lens.intensity == 0) & (aggregator.lens.centre == 1)).string
 
     assert string == (
-        f"SELECT parent_id FROM object WHERE name = 'lens' AND id IN ({type_equality_query} AND id in {equality_query})"
+        f"SELECT parent_id FROM object WHERE name = 'lens' AND id IN ({equality_query_2} AND id in {equality_query})"
     )
