### 2021-01-09 rhayes777/PyAutoFit#143 @Jammy2211
Agreed and changed.
