### 2021-01-19 #2550 @arfon
@Jammy2211 @dfm - not sure what was going on here but there was something in the BibTeX entries causing compilation issues. I swapped them out here and the paper now seems to compile for me locally: https://github.com/rhayes777/PyAutoFit/pull/149
