### 2021-01-23 #2550 @Jammy2211
Hopefully this is ok - I've added a third author to the paper who made a lot of contributions to the software in the past few months.
