### 2021-01-24 #2550 @dfm
@Jammy2211: I've made a few more edits to the manuscript and opened a PR. Once you've taken a look at those and merged, here are the steps that we need to do before final processing:

1. Comment `@whedon generate pdf` on this issue to get the proofs of the manuscript. Read these over and double check that you're happy with it (especially things like author names and affiliations). It is difficult to make changes later.
2. Bump the version number of PyAutoFit and report the new version number back here.
3. Create an archived release (using Zenodo or similar) and report the DOI of that release back here. **Please note that the title and author list of this archive must EXACTLY match the paper.** (On Zenodo, you can edit the metadata after the release has been made if you're using an automated method.)

Thanks for your patience and let me know if you run into any issues!
