### 2021-01-29 #2550 @Jammy2211
We will go with autofit v0.72.2.

I made a couple of changes to the manuscript, primarily typos and adding Binder links (and removing a couple of older links).
