### 2021-02-04 #2550 @kyleniemeyer
Hi @Jammy2211, I am the EIC on duty this week doing some final checks before publishing.

I noticed that your paper does not have an explicit Statement of Need section, which is one of our [required paper elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). This was caught by whedon at the very beginning of the submission process, but looks like it got missed more recently. (We are working on making this requirement more explicit in a few places, and ensuring the automated check happens at other steps in the process.)

Could you add this section? You can likely just repurpose some existing text in your paper. Thanks!

The only other editorial comments I have are minor: in the summary paragraph, could you add a comma after "e.g" in the parenthetical, and remove the redundant "etc."? There is another instance of an "e.g." missing its comma later in the Model Abstraction and Composition section.

The paper looks great otherwise!
