### 2021-02-05 #2550 @Jammy2211
Hi @kyleniemeyer,

I've added a statement of need, which invovled me rejigging a few sentenses and adding a couple :).

I've also sorted out the e.g. / etc issues.

Thanks!
