### 2021-02-11 #2550 @JohannesBuchner
Can the typo PyMulitNest->PyMultiNest still be fixed?
