### 2021-02-11 #2550 @kyleniemeyer
@openjournals/dev The typo is in the Software Citation section of the paper, can that be fixed at this point?
