### 2021-02-11 #2550 @arfon
Yep, we can regenerate the paper if the fix has been made.
