### 2021-02-11 #2550 @kyleniemeyer
Hi @Jammy2211, could you fix the typo pointed out in the paper source?
